Improved frontend player.
