--- MoQPlugin/src/main/resources/moq-ams-player-build/assets/play-C3Y9epmR.js
@@ -1,476 +1,0 @@
-import{S as l,E as S,e as In,M as h,a as je,b as Me,d as Ln,_ as at,R as Bn,f as ct}from"./time-Cy1-YUvq.js";function g(t,e,n){function s(o,c){if(o._zod||Object.defineProperty(o,"_zod",{value:{def:c,constr:a,traits:new Set},enumerable:!1}),o._zod.traits.has(t))return;o._zod.traits.add(t),e(o,c);const u=a.prototype,d=Object.keys(u);for(let f=0;f<d.length;f++){const m=d[f];m in o||(o[m]=u[m].bind(o))}}const i=n?.Parent??Object;class r extends i{}Object.defineProperty(r,"name",{value:t});function a(o){var c;const u=n?.Parent?new r:this;s(u,o),(c=u._zod).deferred??(c.deferred=[]);for(const d of u._zod.deferred)d();return u}return Object.defineProperty(a,"init",{value:s}),Object.defineProperty(a,Symbol.hasInstance,{value:o=>n?.Parent&&o instanceof n.Parent?!0:o?._zod?.traits?.has(t)}),Object.defineProperty(a,"name",{value:t}),a}class Q extends Error{constructor(){super("Encountered Promise during synchronous parse. Use .parseAsync() instead.")}}class Rn extends Error{constructor(e){super(`Encountered unidirectional transform during encode: ${e}`),this.name="ZodEncodeError"}}const On={};function ie(t){return On}function Dn(t){const e=Object.values(t).filter(n=>typeof n=="number");return Object.entries(t).filter(([n,s])=>e.indexOf(+n)===-1).map(([n,s])=>s)}function Fn(t,e){return typeof e=="bigint"?e.toString():e}function jt(t){return{get value(){{const e=t();return Object.defineProperty(this,"value",{value:e}),e}}}}function Pt(t){const e=t.startsWith("^")?1:0,n=t.endsWith("$")?t.length-1:t.length;return t.slice(e,n)}const ut=Symbol("evaluating");function z(t,e,n){let s;Object.defineProperty(t,e,{get(){if(s!==ut)return s===void 0&&(s=ut,s=n()),s},set(i){Object.defineProperty(t,e,{value:i})},configurable:!0})}const Vt="captureStackTrace"in Error?Error.captureStackTrace:(...t)=>{};function Ae(t){return typeof t=="object"&&t!==null&&!Array.isArray(t)}function Wt(t){if(Ae(t)===!1)return!1;const e=t.constructor;if(e===void 0||typeof e!="function")return!0;const n=e.prototype;return!(Ae(n)===!1||Object.prototype.hasOwnProperty.call(n,"isPrototypeOf")===!1)}function $n(t){return Wt(t)?{...t}:Array.isArray(t)?[...t]:t}const Nn=new Set(["string","number","symbol"]);function Pe(t){return t.replace(/[.*+?^${}()|[\]\\]/g,"\\$&")}function jn(t,e,n){const s=new t._zod.constr(e??t._zod.def);return(!e||n?.parent)&&(s._zod.parent=t),s}function I(t){return{}}function Pn(t){return Object.keys(t).filter(e=>t[e]._zod.optin==="optional"&&t[e]._zod.optout==="optional")}const Vn={safeint:[Number.MIN_SAFE_INTEGER,Number.MAX_SAFE_INTEGER],int32:[-2147483648,2147483647],uint32:[0,4294967295],float32:[-34028234663852886e22,34028234663852886e22],float64:[-Number.MAX_VALUE,Number.MAX_VALUE]};function ue(t,e=0){if(t.aborted===!0)return!0;for(let n=e;n<t.issues.length;n++)if(t.issues[n]?.continue!==!0)return!0;return!1}function J(t,e){return e.map(n=>{var s;return(s=n).path??(s.path=[]),n.path.unshift(t),n})}function we(t){return typeof t=="string"?t:t?.message}function re(t,e,n){const s={...t,path:t.path??[]};if(!t.message){const i=we(t.inst?._zod.def?.error?.(t))??we(e?.error?.(t))??we(n.customError?.(t))??we(n.localeError?.(t))??"Invalid input";s.message=i}return delete s.inst,delete s.continue,e?.reportInput||delete s.input,s}const Zt=(t,e)=>{t.name="$ZodError",Object.defineProperty(t,"_zod",{value:t._zod,enumerable:!1}),Object.defineProperty(t,"issues",{value:e,enumerable:!1}),t.message=JSON.stringify(e,Fn,2),Object.defineProperty(t,"toString",{value:()=>t.message,enumerable:!1})},Wn=g("$ZodError",Zt),ze=g("$ZodError",Zt,{Parent:Error}),Zn=t=>(e,n,s,i)=>{const r=s?Object.assign(s,{async:!1}):{async:!1},a=e._zod.run({value:n,issues:[]},r);if(a instanceof Promise)throw new Q;if(a.issues.length){const o=new(i?.Err??t)(a.issues.map(c=>re(c,r,ie())));throw Vt(o,i?.callee),o}return a.value},Hn=Zn(ze),qn=t=>async(e,n,s,i)=>{const r=s?Object.assign(s,{async:!0}):{async:!0};let a=e._zod.run({value:n,issues:[]},r);if(a instanceof Promise&&(a=await a),a.issues.length){const o=new(i?.Err??t)(a.issues.map(c=>re(c,r,ie())));throw Vt(o,i?.callee),o}return a.value},Gn=qn(ze),Xn=t=>(e,n,s)=>{const i=s?{...s,async:!1}:{async:!1},r=e._zod.run({value:n,issues:[]},i);if(r instanceof Promise)throw new Q;return r.issues.length?{success:!1,error:new(t??Wn)(r.issues.map(a=>re(a,i,ie())))}:{success:!0,data:r.value}},Ht=Xn(ze),Yn=t=>async(e,n,s)=>{const i=s?Object.assign(s,{async:!0}):{async:!0};let r=e._zod.run({value:n,issues:[]},i);return r instanceof Promise&&(r=await r),r.issues.length?{success:!1,error:new t(r.issues.map(a=>re(a,i,ie())))}:{success:!0,data:r.value}},qt=Yn(ze),Kn=/^$|^(?:[0-9a-zA-Z+/]{4})*(?:(?:[0-9a-zA-Z+/]{2}==)|(?:[0-9a-zA-Z+/]{3}=))?$/,Jn=t=>{const e=t?`[\\s\\S]{${t?.minimum??0},${t?.maximum??""}}`:"[\\s\\S]*";return new RegExp(`^${e}$`)},Qn=/^-?\d+$/,Gt=/^-?\d+(?:\.\d+)?$/,es=/^(?:true|false)$/i,Te=g("$ZodCheck",(t,e)=>{var n;t._zod??(t._zod={}),t._zod.def=e,(n=t._zod).onattach??(n.onattach=[])}),Xt={number:"number",bigint:"bigint",object:"date"},ts=g("$ZodCheckLessThan",(t,e)=>{Te.init(t,e);const n=Xt[typeof e.value];t._zod.onattach.push(s=>{const i=s._zod.bag,r=(e.inclusive?i.maximum:i.exclusiveMaximum)??Number.POSITIVE_INFINITY;e.value<r&&(e.inclusive?i.maximum=e.value:i.exclusiveMaximum=e.value)}),t._zod.check=s=>{(e.inclusive?s.value<=e.value:s.value<e.value)||s.issues.push({origin:n,code:"too_big",maximum:typeof e.value=="object"?e.value.getTime():e.value,input:s.value,inclusive:e.inclusive,inst:t,continue:!e.abort})}}),ns=g("$ZodCheckGreaterThan",(t,e)=>{Te.init(t,e);const n=Xt[typeof e.value];t._zod.onattach.push(s=>{const i=s._zod.bag,r=(e.inclusive?i.minimum:i.exclusiveMinimum)??Number.NEGATIVE_INFINITY;e.value>r&&(e.inclusive?i.minimum=e.value:i.exclusiveMinimum=e.value)}),t._zod.check=s=>{(e.inclusive?s.value>=e.value:s.value>e.value)||s.issues.push({origin:n,code:"too_small",minimum:typeof e.value=="object"?e.value.getTime():e.value,input:s.value,inclusive:e.inclusive,inst:t,continue:!e.abort})}}),ss=g("$ZodCheckNumberFormat",(t,e)=>{Te.init(t,e),e.format=e.format||"float64";const n=e.format?.includes("int"),s=n?"int":"number",[i,r]=Vn[e.format];t._zod.onattach.push(a=>{const o=a._zod.bag;o.format=e.format,o.minimum=i,o.maximum=r,n&&(o.pattern=Qn)}),t._zod.check=a=>{const o=a.value;if(n){if(!Number.isInteger(o)){a.issues.push({expected:s,format:e.format,code:"invalid_type",continue:!1,input:o,inst:t});return}if(!Number.isSafeInteger(o)){o>0?a.issues.push({input:o,code:"too_big",maximum:Number.MAX_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:t,origin:s,inclusive:!0,continue:!e.abort}):a.issues.push({input:o,code:"too_small",minimum:Number.MIN_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:t,origin:s,inclusive:!0,continue:!e.abort});return}}o<i&&a.issues.push({origin:"number",input:o,code:"too_small",minimum:i,inclusive:!0,inst:t,continue:!e.abort}),o>r&&a.issues.push({origin:"number",input:o,code:"too_big",maximum:r,inclusive:!0,inst:t,continue:!e.abort})}}),is=g("$ZodCheckStringFormat",(t,e)=>{var n,s;Te.init(t,e),t._zod.onattach.push(i=>{const r=i._zod.bag;r.format=e.format,e.pattern&&(r.patterns??(r.patterns=new Set),r.patterns.add(e.pattern))}),e.pattern?(n=t._zod).check??(n.check=i=>{e.pattern.lastIndex=0,!e.pattern.test(i.value)&&i.issues.push({origin:"string",code:"invalid_format",format:e.format,input:i.value,...e.pattern?{pattern:e.pattern.toString()}:{},inst:t,continue:!e.abort})}):(s=t._zod).check??(s.check=()=>{})}),rs={major:4,minor:3,patch:6},L=g("$ZodType",(t,e)=>{var n;t??(t={}),t._zod.def=e,t._zod.bag=t._zod.bag||{},t._zod.version=rs;const s=[...t._zod.def.checks??[]];t._zod.traits.has("$ZodCheck")&&s.unshift(t);for(const i of s)for(const r of i._zod.onattach)r(t);if(s.length===0)(n=t._zod).deferred??(n.deferred=[]),t._zod.deferred?.push(()=>{t._zod.run=t._zod.parse});else{const i=(a,o,c)=>{let u=ue(a),d;for(const f of o){if(f._zod.def.when){if(!f._zod.def.when(a))continue}else if(u)continue;const m=a.issues.length,b=f._zod.check(a);if(b instanceof Promise&&c?.async===!1)throw new Q;if(d||b instanceof Promise)d=(d??Promise.resolve()).then(async()=>{await b,a.issues.length!==m&&(u||(u=ue(a,m)))});else{if(a.issues.length===m)continue;u||(u=ue(a,m))}}return d?d.then(()=>a):a},r=(a,o,c)=>{if(ue(a))return a.aborted=!0,a;const u=i(o,s,c);if(u instanceof Promise){if(c.async===!1)throw new Q;return u.then(d=>t._zod.parse(d,c))}return t._zod.parse(u,c)};t._zod.run=(a,o)=>{if(o.skipChecks)return t._zod.parse(a,o);if(o.direction==="backward"){const u=t._zod.parse({value:a.value,issues:[]},{...o,skipChecks:!0});return u instanceof Promise?u.then(d=>r(d,a,o)):r(u,a,o)}const c=t._zod.parse(a,o);if(c instanceof Promise){if(o.async===!1)throw new Q;return c.then(u=>i(u,s,o))}return i(c,s,o)}}z(t,"~standard",()=>({validate:i=>{try{const r=Ht(t,i);return r.success?{value:r.data}:{issues:r.error?.issues}}catch{return qt(t,i).then(r=>r.success?{value:r.data}:{issues:r.error?.issues})}},vendor:"zod",version:1}))}),Yt=g("$ZodString",(t,e)=>{L.init(t,e),t._zod.pattern=[...t?._zod.bag?.patterns??[]].pop()??Jn(t._zod.bag),t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=String(n.value)}catch{}return typeof n.value=="string"||n.issues.push({expected:"string",code:"invalid_type",input:n.value,inst:t}),n}}),Kt=g("$ZodStringFormat",(t,e)=>{is.init(t,e),Yt.init(t,e)});function os(t){if(t==="")return!0;if(t.length%4!==0)return!1;try{return atob(t),!0}catch{return!1}}const as=g("$ZodBase64",(t,e)=>{e.pattern??(e.pattern=Kn),Kt.init(t,e),t._zod.bag.contentEncoding="base64",t._zod.check=n=>{os(n.value)||n.issues.push({code:"invalid_format",format:"base64",input:n.value,inst:t,continue:!e.abort})}}),Jt=g("$ZodNumber",(t,e)=>{L.init(t,e),t._zod.pattern=t._zod.bag.pattern??Gt,t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=Number(n.value)}catch{}const i=n.value;if(typeof i=="number"&&!Number.isNaN(i)&&Number.isFinite(i))return n;const r=typeof i=="number"?Number.isNaN(i)?"NaN":Number.isFinite(i)?void 0:"Infinity":void 0;return n.issues.push({expected:"number",code:"invalid_type",input:i,inst:t,...r?{received:r}:{}}),n}}),cs=g("$ZodNumberFormat",(t,e)=>{ss.init(t,e),Jt.init(t,e)}),us=g("$ZodBoolean",(t,e)=>{L.init(t,e),t._zod.pattern=es,t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=!!n.value}catch{}const i=n.value;return typeof i=="boolean"||n.issues.push({expected:"boolean",code:"invalid_type",input:i,inst:t}),n}});function dt(t,e,n){t.issues.length&&e.issues.push(...J(n,t.issues)),e.value[n]=t.value}const ds=g("$ZodArray",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{const i=n.value;if(!Array.isArray(i))return n.issues.push({expected:"array",code:"invalid_type",input:i,inst:t}),n;n.value=Array(i.length);const r=[];for(let a=0;a<i.length;a++){const o=i[a],c=e.element._zod.run({value:o,issues:[]},s);c instanceof Promise?r.push(c.then(u=>dt(u,n,a))):dt(c,n,a)}return r.length?Promise.all(r).then(()=>n):n}});function xe(t,e,n,s,i){if(t.issues.length){if(i&&!(n in s))return;e.issues.push(...J(n,t.issues))}t.value===void 0?n in s&&(e.value[n]=void 0):e.value[n]=t.value}function ls(t){const e=Object.keys(t.shape);for(const s of e)if(!t.shape?.[s]?._zod?.traits?.has("$ZodType"))throw new Error(`Invalid element at key "${s}": expected a Zod schema`);const n=Pn(t.shape);return{...t,keys:e,keySet:new Set(e),numKeys:e.length,optionalKeys:new Set(n)}}function fs(t,e,n,s,i,r){const a=[],o=i.keySet,c=i.catchall._zod,u=c.def.type,d=c.optout==="optional";for(const f in e){if(o.has(f))continue;if(u==="never"){a.push(f);continue}const m=c.run({value:e[f],issues:[]},s);m instanceof Promise?t.push(m.then(b=>xe(b,n,f,e,d))):xe(m,n,f,e,d)}return a.length&&n.issues.push({code:"unrecognized_keys",keys:a,input:e,inst:r}),t.length?Promise.all(t).then(()=>n):n}const hs=g("$ZodObject",(t,e)=>{if(L.init(t,e),!Object.getOwnPropertyDescriptor(e,"shape")?.get){const a=e.shape;Object.defineProperty(e,"shape",{get:()=>{const o={...a};return Object.defineProperty(e,"shape",{value:o}),o}})}const n=jt(()=>ls(e));z(t._zod,"propValues",()=>{const a=e.shape,o={};for(const c in a){const u=a[c]._zod;if(u.values){o[c]??(o[c]=new Set);for(const d of u.values)o[c].add(d)}}return o});const s=Ae,i=e.catchall;let r;t._zod.parse=(a,o)=>{r??(r=n.value);const c=a.value;if(!s(c))return a.issues.push({expected:"object",code:"invalid_type",input:c,inst:t}),a;a.value={};const u=[],d=r.shape;for(const f of r.keys){const m=d[f],b=m._zod.optout==="optional",w=m._zod.run({value:c[f],issues:[]},o);w instanceof Promise?u.push(w.then(k=>xe(k,a,f,c,b))):xe(w,a,f,c,b)}return i?fs(u,c,a,o,n.value,t):u.length?Promise.all(u).then(()=>a):a}});function lt(t,e,n,s){for(const r of t)if(r.issues.length===0)return e.value=r.value,e;const i=t.filter(r=>!ue(r));return i.length===1?(e.value=i[0].value,i[0]):(e.issues.push({code:"invalid_union",input:e.value,inst:n,errors:t.map(r=>r.issues.map(a=>re(a,s,ie())))}),e)}const Qt=g("$ZodUnion",(t,e)=>{L.init(t,e),z(t._zod,"optin",()=>e.options.some(i=>i._zod.optin==="optional")?"optional":void 0),z(t._zod,"optout",()=>e.options.some(i=>i._zod.optout==="optional")?"optional":void 0),z(t._zod,"values",()=>{if(e.options.every(i=>i._zod.values))return new Set(e.options.flatMap(i=>Array.from(i._zod.values)))}),z(t._zod,"pattern",()=>{if(e.options.every(i=>i._zod.pattern)){const i=e.options.map(r=>r._zod.pattern);return new RegExp(`^(${i.map(r=>Pt(r.source)).join("|")})$`)}});const n=e.options.length===1,s=e.options[0]._zod.run;t._zod.parse=(i,r)=>{if(n)return s(i,r);let a=!1;const o=[];for(const c of e.options){const u=c._zod.run({value:i.value,issues:[]},r);if(u instanceof Promise)o.push(u),a=!0;else{if(u.issues.length===0)return u;o.push(u)}}return a?Promise.all(o).then(c=>lt(c,i,t,r)):lt(o,i,t,r)}}),ms=g("$ZodDiscriminatedUnion",(t,e)=>{e.inclusive=!1,Qt.init(t,e);const n=t._zod.parse;z(t._zod,"propValues",()=>{const i={};for(const r of e.options){const a=r._zod.propValues;if(!a||Object.keys(a).length===0)throw new Error(`Invalid discriminated union option at index "${e.options.indexOf(r)}"`);for(const[o,c]of Object.entries(a)){i[o]||(i[o]=new Set);for(const u of c)i[o].add(u)}}return i});const s=jt(()=>{const i=e.options,r=new Map;for(const a of i){const o=a._zod.propValues?.[e.discriminator];if(!o||o.size===0)throw new Error(`Invalid discriminated union option at index "${e.options.indexOf(a)}"`);for(const c of o){if(r.has(c))throw new Error(`Duplicate discriminator value "${String(c)}"`);r.set(c,a)}}return r});t._zod.parse=(i,r)=>{const a=i.value;if(!Ae(a))return i.issues.push({code:"invalid_type",expected:"object",input:a,inst:t}),i;const o=s.value.get(a?.[e.discriminator]);return o?o._zod.run(i,r):e.unionFallback?n(i,r):(i.issues.push({code:"invalid_union",errors:[],note:"No matching discriminator",discriminator:e.discriminator,input:a,path:[e.discriminator],inst:t}),i)}}),ps=g("$ZodRecord",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{const i=n.value;if(!Wt(i))return n.issues.push({expected:"record",code:"invalid_type",input:i,inst:t}),n;const r=[],a=e.keyType._zod.values;if(a){n.value={};const o=new Set;for(const u of a)if(typeof u=="string"||typeof u=="number"||typeof u=="symbol"){o.add(typeof u=="number"?u.toString():u);const d=e.valueType._zod.run({value:i[u],issues:[]},s);d instanceof Promise?r.push(d.then(f=>{f.issues.length&&n.issues.push(...J(u,f.issues)),n.value[u]=f.value})):(d.issues.length&&n.issues.push(...J(u,d.issues)),n.value[u]=d.value)}let c;for(const u in i)o.has(u)||(c=c??[],c.push(u));c&&c.length>0&&n.issues.push({code:"unrecognized_keys",input:i,inst:t,keys:c})}else{n.value={};for(const o of Reflect.ownKeys(i)){if(o==="__proto__")continue;let c=e.keyType._zod.run({value:o,issues:[]},s);if(c instanceof Promise)throw new Error("Async schemas not supported in object keys currently");if(typeof o=="string"&&Gt.test(o)&&c.issues.length){const d=e.keyType._zod.run({value:Number(o),issues:[]},s);if(d instanceof Promise)throw new Error("Async schemas not supported in object keys currently");d.issues.length===0&&(c=d)}if(c.issues.length){e.mode==="loose"?n.value[o]=i[o]:n.issues.push({code:"invalid_key",origin:"record",issues:c.issues.map(d=>re(d,s,ie())),input:o,path:[o],inst:t});continue}const u=e.valueType._zod.run({value:i[o],issues:[]},s);u instanceof Promise?r.push(u.then(d=>{d.issues.length&&n.issues.push(...J(o,d.issues)),n.value[c.value]=d.value})):(u.issues.length&&n.issues.push(...J(o,u.issues)),n.value[c.value]=u.value)}}return r.length?Promise.all(r).then(()=>n):n}}),gs=g("$ZodEnum",(t,e)=>{L.init(t,e);const n=Dn(e.entries),s=new Set(n);t._zod.values=s,t._zod.pattern=new RegExp(`^(${n.filter(i=>Nn.has(typeof i)).map(i=>typeof i=="string"?Pe(i):i.toString()).join("|")})$`),t._zod.parse=(i,r)=>{const a=i.value;return s.has(a)||i.issues.push({code:"invalid_value",values:n,input:a,inst:t}),i}}),bs=g("$ZodLiteral",(t,e)=>{if(L.init(t,e),e.values.length===0)throw new Error("Cannot create literal schema with no valid values");const n=new Set(e.values);t._zod.values=n,t._zod.pattern=new RegExp(`^(${e.values.map(s=>typeof s=="string"?Pe(s):s?Pe(s.toString()):String(s)).join("|")})$`),t._zod.parse=(s,i)=>{const r=s.value;return n.has(r)||s.issues.push({code:"invalid_value",values:e.values,input:r,inst:t}),s}}),ws=g("$ZodTransform",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{if(s.direction==="backward")throw new Rn(t.constructor.name);const i=e.transform(n.value,n);if(s.async)return(i instanceof Promise?i:Promise.resolve(i)).then(r=>(n.value=r,n));if(i instanceof Promise)throw new Q;return n.value=i,n}});function ft(t,e){return t.issues.length&&e===void 0?{issues:[],value:void 0}:t}const ys=g("$ZodOptional",(t,e)=>{L.init(t,e),t._zod.optin="optional",t._zod.optout="optional",z(t._zod,"values",()=>e.innerType._zod.values?new Set([...e.innerType._zod.values,void 0]):void 0),z(t._zod,"pattern",()=>{const n=e.innerType._zod.pattern;return n?new RegExp(`^(${Pt(n.source)})?$`):void 0}),t._zod.parse=(n,s)=>{if(e.innerType._zod.optin==="optional"){const i=e.innerType._zod.run(n,s);return i instanceof Promise?i.then(r=>ft(r,n.value)):ft(i,n.value)}return n.value===void 0?n:e.innerType._zod.run(n,s)}}),vs=g("$ZodDefault",(t,e)=>{L.init(t,e),t._zod.optin="optional",z(t._zod,"values",()=>e.innerType._zod.values),t._zod.parse=(n,s)=>{if(s.direction==="backward")return e.innerType._zod.run(n,s);if(n.value===void 0)return n.value=e.defaultValue,n;const i=e.innerType._zod.run(n,s);return i instanceof Promise?i.then(r=>ht(r,e)):ht(i,e)}});function ht(t,e){return t.value===void 0&&(t.value=e.defaultValue),t}const ks=g("$ZodPipe",(t,e)=>{L.init(t,e),z(t._zod,"values",()=>e.in._zod.values),z(t._zod,"optin",()=>e.in._zod.optin),z(t._zod,"optout",()=>e.out._zod.optout),z(t._zod,"propValues",()=>e.in._zod.propValues),t._zod.parse=(n,s)=>{if(s.direction==="backward"){const r=e.out._zod.run(n,s);return r instanceof Promise?r.then(a=>ye(a,e.in,s)):ye(r,e.in,s)}const i=e.in._zod.run(n,s);return i instanceof Promise?i.then(r=>ye(r,e.out,s)):ye(i,e.out,s)}});function ye(t,e,n){return t.issues.length?(t.aborted=!0,t):e._zod.run({value:t.value,issues:t.issues},n)}var mt;class As{constructor(){this._map=new WeakMap,this._idmap=new Map}add(e,...n){const s=n[0];return this._map.set(e,s),s&&typeof s=="object"&&"id"in s&&this._idmap.set(s.id,e),this}clear(){return this._map=new WeakMap,this._idmap=new Map,this}remove(e){const n=this._map.get(e);return n&&typeof n=="object"&&"id"in n&&this._idmap.delete(n.id),this._map.delete(e),this}get(e){const n=e._zod.parent;if(n){const s={...this.get(n)??{}};delete s.id;const i={...s,...this._map.get(e)};return Object.keys(i).length?i:void 0}return this._map.get(e)}has(e){return this._map.has(e)}}function xs(){return new As}(mt=globalThis).__zod_globalRegistry??(mt.__zod_globalRegistry=xs());function _s(t,e){return new t({type:"string",...I()})}function Ss(t,e){return new t({type:"string",format:"base64",check:"string_format",abort:!1,...I()})}function Es(t,e){return new t({type:"number",checks:[],...I()})}function Us(t,e){return new t({type:"number",check:"number_format",abort:!1,format:"safeint",...I()})}function Ms(t,e){return new t({type:"boolean",...I()})}function en(t,e){return new ts({check:"less_than",...I(),value:t,inclusive:!0})}function zs(t,e){return new ns({check:"greater_than",...I(),value:t,inclusive:!0})}function tn(t){return zs(0)}const B=g("ZodMiniType",(t,e)=>{if(!t._zod)throw new Error("Uninitialized schema in ZodMiniType.");L.init(t,e),t.def=e,t.type=e.type,t.parse=(n,s)=>Hn(t,n,s,{callee:t.parse}),t.safeParse=(n,s)=>Ht(t,n,s),t.parseAsync=async(n,s)=>Gn(t,n,s,{callee:t.parseAsync}),t.safeParseAsync=async(n,s)=>qt(t,n,s),t.check=(...n)=>t.clone({...e,checks:[...e.checks??[],...n.map(s=>typeof s=="function"?{_zod:{check:s,def:{check:"custom"},onattach:[]}}:s)]},{parent:!0}),t.with=t.check,t.clone=(n,s)=>jn(t,n,s),t.brand=()=>t,t.register=((n,s)=>(n.add(t,s),t)),t.apply=n=>n(t)}),nn=g("ZodMiniString",(t,e)=>{Yt.init(t,e),B.init(t,e)});function v(t){return _s(nn)}const Ts=g("ZodMiniStringFormat",(t,e)=>{Kt.init(t,e),nn.init(t,e)}),Cs=g("ZodMiniBase64",(t,e)=>{as.init(t,e),Ts.init(t,e)});function Is(t){return Ss(Cs)}const sn=g("ZodMiniNumber",(t,e)=>{Jt.init(t,e),B.init(t,e)});function M(t){return Es(sn)}const Ls=g("ZodMiniNumberFormat",(t,e)=>{cs.init(t,e),sn.init(t,e)});function rn(t){return Us(Ls)}const Bs=g("ZodMiniBoolean",(t,e)=>{us.init(t,e),B.init(t,e)});function Ye(t){return Ms(Bs)}const Rs=g("ZodMiniArray",(t,e)=>{ds.init(t,e),B.init(t,e)});function q(t,e){return new Rs({type:"array",element:t,...I()})}const Os=g("ZodMiniObject",(t,e)=>{hs.init(t,e),B.init(t,e),z(t,"shape",()=>e.shape)});function A(t,e){const n={type:"object",shape:t??{},...I()};return new Os(n)}const Ds=g("ZodMiniUnion",(t,e)=>{Qt.init(t,e),B.init(t,e)});function Ce(t,e){return new Ds({type:"union",options:t,...I()})}const Fs=g("ZodMiniDiscriminatedUnion",(t,e)=>{ms.init(t,e),B.init(t,e)});function $s(t,e,n){return new Fs({type:"union",options:e,discriminator:t,...I()})}const Ns=g("ZodMiniRecord",(t,e)=>{ps.init(t,e),B.init(t,e)});function on(t,e,n){return new Ns({type:"record",keyType:t,valueType:e,...I()})}const js=g("ZodMiniEnum",(t,e)=>{gs.init(t,e),B.init(t,e),t.options=Object.values(e.entries)});function an(t,e){const n=Array.isArray(t)?Object.fromEntries(t.map(s=>[s,s])):t;return new js({type:"enum",entries:n,...I()})}const Ps=g("ZodMiniLiteral",(t,e)=>{bs.init(t,e),B.init(t,e)});function Ve(t,e){return new Ps({type:"literal",values:Array.isArray(t)?t:[t],...I()})}const Vs=g("ZodMiniTransform",(t,e)=>{ws.init(t,e),B.init(t,e)});function cn(t){return new Vs({type:"transform",transform:t})}const Ws=g("ZodMiniOptional",(t,e)=>{ys.init(t,e),B.init(t,e)});function p(t){return new Ws({type:"optional",innerType:t})}const Zs=g("ZodMiniDefault",(t,e)=>{vs.init(t,e),B.init(t,e)});function Hs(t,e){return new Zs({type:"default",innerType:t,get defaultValue(){return typeof e=="function"?e():$n(e)}})}const qs=g("ZodMiniPipe",(t,e)=>{ks.init(t,e),B.init(t,e)});function un(t,e){return new qs({type:"pipe",in:t,out:e})}const dn=Hs($s("kind",[A({kind:Ve("legacy")}),A({kind:Ve("cmaf"),init:Is()})]),{kind:"legacy"});M().check(rn(),tn(),en(255)).brand("u8");const R=M().check(rn(),tn(),en(Number.MAX_SAFE_INTEGER)).brand("u53");function Z(t){return R.parse(t)}const Gs=A({name:v()}),pt=A({codec:v(),container:dn,description:p(v()),sampleRate:R,numberOfChannels:R,bitrate:p(R),jitter:p(R)}),Xs=Ce([A({renditions:on(v(),pt)}),un(A({track:Gs,config:pt}),cn(t=>({renditions:{[t.track.name]:t.config}})))]),Ys=A({hardware:p(q(v())),software:p(q(v())),unsupported:p(q(v()))}),Ks=A({hardware:p(q(v())),software:p(q(v())),unsupported:p(q(v()))}),Js=A({video:p(Ys),audio:p(Ks)}),pe=A({name:v()}),Qs=A({message:p(pe),typing:p(pe)}),ei=A({x:p(M()),y:p(M()),z:p(M()),s:p(M())}),ti=A({initial:p(ei),track:p(pe),handle:p(v()),peers:p(pe)}),te={catalog:100,audio:80,video:60},ni=A({id:p(v()),name:p(v()),avatar:p(v()),color:p(v())}),si=A({name:v()}),gt=A({codec:v(),container:dn,description:p(v()),codedWidth:p(R),codedHeight:p(R),displayAspectWidth:p(R),displayAspectHeight:p(R),framerate:p(M()),bitrate:p(R),optimizeForLatency:p(Ye()),jitter:p(R)}),ii=Ce([A({renditions:on(v(),gt),display:p(A({width:R,height:R})),rotation:p(M()),flip:p(Ye())}),un(q(A({track:si,config:gt})),cn(t=>{const e=t[0]?.config;return{renditions:Object.fromEntries(t.map(n=>[n.track.name,n.config])),display:e?.displayAspectWidth!==void 0&&e?.displayAspectHeight!==void 0?{width:e.displayAspectWidth,height:e.displayAspectHeight}:void 0,rotation:void 0,flip:void 0}}))]),ri=A({video:p(ii),audio:p(Xs),location:p(ti),user:p(ni),chat:p(Qs),capabilities:p(Js),preview:p(pe)});function oi(t){const e=new TextDecoder().decode(t);try{const n=JSON.parse(e);return ri.parse(n)}catch(n){throw console.warn("invalid catalog",e),n}}async function ai(t){const e=await t.readFrame();if(e)return oi(e)}function ln(t){return t instanceof ArrayBuffer||typeof SharedArrayBuffer<"u"&&t instanceof SharedArrayBuffer}const ci="utf-16",Be="utf-16be",bt="utf-16le",fe="utf-8";function fn(t,e={}){let n;ln(t)?n=new DataView(t):n=new DataView(t.buffer,t.byteOffset,t.byteLength);let s=0,{encoding:i}=e;if(!i){const u=n.getUint8(0),d=n.getUint8(1);u==239&&d==187&&n.getUint8(2)==191?(i=fe,s=3):u==254&&d==255?(i=Be,s=2):u==255&&d==254?(i=bt,s=2):i=fe}if(typeof TextDecoder<"u")return new TextDecoder(i).decode(n);const{byteLength:r}=n,a=i!==Be;let o="",c;for(;s<r;){switch(i){case fe:if(c=n.getUint8(s),c<128)s++;else if(c>=194&&c<=223)if(s+1<r){const u=n.getUint8(s+1);u>=128&&u<=191?(c=(c&31)<<6|u&63,s+=2):s++}else s++;else if(c>=224&&c<=239)if(s+2<=r-1){const u=n.getUint8(s+1),d=n.getUint8(s+2);u>=128&&u<=191&&d>=128&&d<=191?(c=(c&15)<<12|(u&63)<<6|d&63,s+=3):s++}else s++;else if(c>=240&&c<=244)if(s+3<=r-1){const u=n.getUint8(s+1),d=n.getUint8(s+2),f=n.getUint8(s+3);u>=128&&u<=191&&d>=128&&d<=191&&f>=128&&f<=191?(c=(c&7)<<18|(u&63)<<12|(d&63)<<6|f&63,s+=4):s++}else s++;else s++;break;case Be:case ci:case bt:c=n.getUint16(s,a),s+=2;break}o+=String.fromCodePoint(c)}return o}function ui(t){return new TextEncoder().encode(t)}function di(t){return{writers:t?.writers??{}}}const li=["dinf","edts","grpl","mdia","meco","mfra","minf","moof","moov","mvex","schi","sinf","stbl","strk","traf","trak","tref","udta","vttc"];function hn(t){return"boxes"in t||li.includes(t.type)}const wt="utf8",C="uint",ne="template",yt="string",vt="int",kt="data";var E=class{constructor(e,n){this.writeUint=(s,i)=>{const{dataView:r,cursor:a}=this;switch(i){case 1:r.setUint8(a,s);break;case 2:r.setUint16(a,s);break;case 3:{const o=(s&16776960)>>8,c=s&255;r.setUint16(a,o),r.setUint8(a+2,c);break}case 4:r.setUint32(a,s);break;case 8:{const o=Math.floor(s/Math.pow(2,32)),c=s-o*Math.pow(2,32);r.setUint32(a,o),r.setUint32(a+4,c);break}}this.cursor+=i},this.writeInt=(s,i)=>{const{dataView:r,cursor:a}=this;switch(i){case 1:r.setInt8(a,s);break;case 2:r.setInt16(a,s);break;case 4:r.setInt32(a,s);break;case 8:const o=Math.floor(s/Math.pow(2,32)),c=s-o*Math.pow(2,32);r.setUint32(a,o),r.setUint32(a+4,c);break}this.cursor+=i},this.writeString=s=>{for(let i=0,r=s.length;i<r;i++)this.writeUint(s.charCodeAt(i),1)},this.writeTerminatedString=s=>{if(s.length!==0){for(let i=0,r=s.length;i<r;i++)this.writeUint(s.charCodeAt(i),1);this.writeUint(0,1)}},this.writeUtf8TerminatedString=s=>{const i=ui(s);new Uint8Array(this.dataView.buffer).set(i,this.cursor),this.cursor+=i.length,this.writeUint(0,1)},this.writeBytes=s=>{Array.isArray(s)||(s=[s]);for(const i of s)new Uint8Array(this.dataView.buffer).set(i,this.cursor),this.cursor+=i.length},this.writeArray=(s,i,r,a)=>{const o=i===C?this.writeUint:i===ne?this.writeTemplate:this.writeInt;for(let c=0;c<a;c++)o(s[c]??0,r)},this.writeTemplate=(s,i)=>{const r=Math.round(s*Math.pow(2,i===4?16:8));this.writeUint(r,i)},this.writeBoxHeader=(s,i)=>{i>4294967295?(this.writeUint(1,4),this.writeString(s),this.writeUint(i,8)):(this.writeUint(i,4),this.writeString(s))},this.dataView=new DataView(new ArrayBuffer(n)),this.cursor=0,this.writeBoxHeader(e,n)}get buffer(){return this.dataView.buffer}get byteLength(){return this.dataView.byteLength}get byteOffset(){return this.dataView.byteOffset}writeFullBox(e,n){this.writeUint(e,1),this.writeUint(n,3)}};function mn(t,e){return Array.from(t,n=>hi(n,e))}function Ke(t,e){const n=mn(t,e);return{bytes:n,size:n.reduce((s,i)=>s+i.byteLength,0)}}function fi(t,e){const{bytes:n,size:s}=Ke(t.boxes,e),i=8+s,r=new E(t.type,i);return r.writeBytes(n),r}function hi(t,e){let n=null;if("type"in t){const{type:s}=t,i=e.writers?.[s];if(i?n=i(t,e):hn(t)?n=fi(t,e):"view"in t&&(n=t.view),!n)throw new Error(`No writer found for box type: ${s}`)}if("buffer"in t&&(n=t),!n)throw new Error("Invalid box");return new Uint8Array(n.buffer,n.byteOffset,n.byteLength)}function mi(t,e,n){const s=n>0?n:t.byteLength-(e-t.byteOffset);return new Uint8Array(t.buffer,e,Math.max(s,0))}function pi(t,e,n){let s=NaN;const i=e-t.byteOffset;switch(n){case 1:s=t.getInt8(i);break;case 2:s=t.getInt16(i);break;case 4:s=t.getInt32(i);break;case 8:const r=t.getInt32(i),a=t.getInt32(i+4);s=r*Math.pow(2,32)+a;break}return s}function X(t,e,n){const s=e-t.byteOffset;let i=NaN,r,a;switch(n){case 1:i=t.getUint8(s);break;case 2:i=t.getUint16(s);break;case 3:r=t.getUint16(s),a=t.getUint8(s+2),i=(r<<8)+a;break;case 4:i=t.getUint32(s);break;case 8:r=t.getUint32(s),a=t.getUint32(s+4),i=r*Math.pow(2,32)+a;break}return i}function At(t,e,n){let s="";for(let i=0;i<n;i++){const r=X(t,e+i,1);s+=String.fromCharCode(r)}return s}function gi(t,e,n){const s=n/2;return X(t,e,s)+X(t,e+s,s)/Math.pow(2,s)}function bi(t,e){let n="",s=e;for(;s-t.byteOffset<t.byteLength;){const i=X(t,s,1);if(i===0)break;n+=String.fromCharCode(i),s++}return n}function wi(t,e){const n=t.byteLength-(e-t.byteOffset);return n>0?fn(new DataView(t.buffer,e,n),{encoding:fe}):""}function yi(t,e){const n=t.byteLength-(e-t.byteOffset);let s="";if(n>0){const i=new DataView(t.buffer,e,n);let r=0;for(;r<n&&i.getUint8(r)!==0;r++);s=fn(new DataView(t.buffer,e,r),{encoding:fe})}return s}var vi=class pn{constructor(e,n){this.truncated=!1,this.slice=(s,i)=>{const r=new pn(new DataView(this.dataView.buffer,s,i),this.config),a=this.offset-s,o=i-a;return this.offset+=o,r.jump(a),r},this.read=(s,i=0)=>{const{dataView:r,offset:a}=this;let o,c=i;switch(s){case C:o=X(r,a,i);break;case vt:o=pi(r,a,i);break;case ne:o=gi(r,a,i);break;case yt:i===-1?(o=bi(r,a),c=o.length+1):o=At(r,a,i);break;case kt:o=mi(r,a,i),c=o.length;break;case wt:i===-1?(o=yi(r,a),c=o.length+1):o=wi(r,a);break;default:o=-1}return this.offset+=c,o},this.readUint=s=>this.read(C,s),this.readInt=s=>this.read(vt,s),this.readString=s=>this.read(yt,s),this.readTemplate=s=>this.read(ne,s),this.readData=s=>this.read(kt,s),this.readUtf8=s=>this.read(wt,s),this.readFullBox=()=>({version:this.readUint(1),flags:this.readUint(3)}),this.readArray=(s,i,r)=>{const a=[];for(let o=0;o<r;o++)a.push(this.read(s,i));return a},this.jump=s=>{this.offset+=s},this.readBox=()=>{const{dataView:s,offset:i}=this;let r=0;const a=X(s,i,4),o=At(s,i+4,4),c={size:a,type:o};r+=8,c.size===1&&(c.largesize=X(s,i+r,8),r+=8);const u=c.size===0?this.bytesRemaining:c.largesize??c.size;if(this.cursor+u>s.byteLength)throw this.truncated=!0,new Error("Truncated box");return this.jump(r),o==="uuid"&&(c.usertype=this.readArray("uint",1,16)),c.view=this.slice(i,u),c},this.readBoxes=(s=-1)=>{const i=[];for(const r of this)if(i.push(r),s>0&&i.length>=s)break;return i},this.readEntries=(s,i)=>{const r=[];for(let a=0;a<s;a++)r.push(i());return r},this.dataView=ln(e)?new DataView(e):e instanceof DataView?e:new DataView(e.buffer,e.byteOffset,e.byteLength),this.offset=this.dataView.byteOffset,this.config=n||{}}get buffer(){return this.dataView.buffer}get byteOffset(){return this.dataView.byteOffset}get byteLength(){return this.dataView.byteLength}get cursor(){return this.offset-this.dataView.byteOffset}get done(){return this.cursor>=this.dataView.byteLength||this.truncated}get bytesRemaining(){return this.dataView.byteLength-this.cursor}*[Symbol.iterator](){const{readers:e={}}=this.config;for(;!this.done;)try{const n=this.readBox(),{type:s,view:i}=n,r=e[s]||e[s.trim()];if(r&&Object.assign(n,r(i,s)),hn(n)&&!n.boxes){const a=[];for(const o of i)a.push(o);n.boxes=a}yield n}catch(n){if(n instanceof Error&&n.message==="Truncated box")break;throw n}}};function Je(t,e){const n=[];for(const s of new vi(t,e))n.push(s);return n}function ki(t,e){return mn(t,di(e))}function Ai(t,e){const{readArray:n,readUint:s,readTemplate:i,readBoxes:r}=e;return{type:t,reserved1:n(C,1,6),dataReferenceIndex:s(2),reserved2:n(C,4,2),channelcount:s(2),samplesize:s(2),preDefined:s(2),reserved3:s(2),samplerate:i(4),boxes:r()}}function Qe(t,e){const{readArray:n,readUint:s,readInt:i,readTemplate:r,readBoxes:a}=e;return{type:t,reserved1:n(C,1,6),dataReferenceIndex:s(2),preDefined1:s(2),reserved2:s(2),preDefined2:n(C,4,3),width:s(2),height:s(2),horizresolution:r(4),vertresolution:r(4),reserved3:s(4),frameCount:s(2),compressorName:n(C,1,32),depth:s(2),preDefined3:i(2),boxes:a()}}function xt(t){return Qe("avc1",t)}function xi(t){return Qe("hev1",t)}function _i(t){return Qe("hvc1",t)}function Si(t){return{type:"mdat",data:t.readData(-1)}}function Ei(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(e==1?8:4),i=t.readUint(e==1?8:4),r=t.readUint(4),a=t.readUint(e==1?8:4),o=t.readUint(2);return{type:"mdhd",version:e,flags:n,creationTime:s,modificationTime:i,timescale:r,duration:a,language:String.fromCharCode((o>>10&31)+96,(o>>5&31)+96,(o&31)+96),preDefined:t.readUint(2)}}function Ui(t){return{type:"mfhd",...t.readFullBox(),sequenceNumber:t.readUint(4)}}function Mi(t){return Ai("mp4a",t)}function zi(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(4);return{type:"stsd",version:e,flags:n,entryCount:s,entries:t.readBoxes(s)}}function Ti(t){const{version:e,flags:n}=t.readFullBox();return{type:"tfdt",version:e,flags:n,baseMediaDecodeTime:t.readUint(e==1?8:4)}}function Ci(t){const{version:e,flags:n}=t.readFullBox();return{type:"tfhd",version:e,flags:n,trackId:t.readUint(4),baseDataOffset:n&1?t.readUint(8):void 0,sampleDescriptionIndex:n&2?t.readUint(4):void 0,defaultSampleDuration:n&8?t.readUint(4):void 0,defaultSampleSize:n&16?t.readUint(4):void 0,defaultSampleFlags:n&32?t.readUint(4):void 0}}function Ii(t){const{version:e,flags:n}=t.readFullBox(),s=e===1?8:4;return{type:"tkhd",version:e,flags:n,creationTime:t.readUint(s),modificationTime:t.readUint(s),trackId:t.readUint(4),reserved1:t.readUint(4),duration:t.readUint(s),reserved2:t.readArray(C,4,2),layer:t.readUint(2),alternateGroup:t.readUint(2),volume:t.readTemplate(2),reserved3:t.readUint(2),matrix:t.readArray(ne,4,9),width:t.readTemplate(4),height:t.readTemplate(4)}}function Li(t){return{type:"trex",...t.readFullBox(),trackId:t.readUint(4),defaultSampleDescriptionIndex:t.readUint(4),defaultSampleDuration:t.readUint(4),defaultSampleSize:t.readUint(4),defaultSampleFlags:t.readUint(4)}}function Bi(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(4);let i,r;n&1&&(i=t.readInt(4)),n&4&&(r=t.readUint(4));const a=t.readEntries(s,()=>{const o={};return n&256&&(o.sampleDuration=t.readUint(4)),n&512&&(o.sampleSize=t.readUint(4)),n&1024&&(o.sampleFlags=t.readUint(4)),n&2048&&(o.sampleCompositionTimeOffset=e===1?t.readInt(4):t.readUint(4)),o});return{type:"trun",version:e,flags:n,sampleCount:s,dataOffset:i,firstSampleFlags:r,samples:a}}function Ri(t,e){const n=t.entries.length,{bytes:s,size:i}=Ke(t.entries,e),r=new E("dref",16+i);return r.writeFullBox(t.version,t.flags),r.writeUint(n,4),r.writeBytes(s),r}function Oi(t){const e=t.compatibleBrands.length*4,n=new E("ftyp",16+e);n.writeString(t.majorBrand),n.writeUint(t.minorVersion,4);for(const s of t.compatibleBrands)n.writeString(s);return n}function Di(t){const e=t.name.length+1,n=new E("hdlr",32+e);return n.writeFullBox(t.version,t.flags),n.writeUint(t.preDefined,4),n.writeString(t.handlerType),n.writeArray(t.reserved,C,4,3),n.writeTerminatedString(t.name),n}function Fi(t){const e=new E("mdat",8+t.data.length);return e.writeBytes(t.data),e}function $i(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("mdhd",n+s+i+4+2+2);r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.timescale,4),r.writeUint(t.duration,e);const a=t.language.length>=3?(t.language.charCodeAt(0)-96&31)<<10|(t.language.charCodeAt(1)-96&31)<<5|t.language.charCodeAt(2)-96&31:0;return r.writeUint(a,2),r.writeUint(t.preDefined,2),r}function Ni(t){const e=new E("mfhd",16);return e.writeFullBox(t.version,t.flags),e.writeUint(t.sequenceNumber,4),e}function ji(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("mvhd",n+s+i+4+4+2+2+8+36+24+4);return r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.timescale,4),r.writeUint(t.duration,e),r.writeTemplate(t.rate,4),r.writeTemplate(t.volume,2),r.writeUint(t.reserved1,2),r.writeArray(t.reserved2,C,4,2),r.writeArray(t.matrix,ne,4,9),r.writeArray(t.preDefined,C,4,6),r.writeUint(t.nextTrackId,4),r}function Pi(t){const e=new E("smhd",16);return e.writeFullBox(t.version,t.flags),e.writeUint(t.balance,2),e.writeUint(t.reserved,2),e}function Vi(t,e){const n=t.entries.length,{bytes:s,size:i}=Ke(t.entries,e),r=new E("stsd",16+i);return r.writeFullBox(t.version,t.flags),r.writeUint(n,4),r.writeBytes(s),r}function Wi(t){const e=t.entryCount*8,n=new E("stts",16+e);n.writeFullBox(t.version,t.flags),n.writeUint(t.entryCount,4);for(const s of t.entries)n.writeUint(s.sampleCount,4),n.writeUint(s.sampleDelta,4);return n}function Zi(t){const e=t.version===1?8:4,n=8,s=4,i=e,r=new E("tfdt",n+s+i);return r.writeFullBox(t.version,t.flags),r.writeUint(t.baseMediaDecodeTime,e),r}function Hi(t){const e=t.flags&1?8:0,n=t.flags&2?4:0,s=t.flags&8?4:0,i=t.flags&16?4:0,r=t.flags&32?4:0,a=new E("tfhd",16+e+n+s+i+r);return a.writeFullBox(t.version,t.flags),a.writeUint(t.trackId,4),t.flags&1&&a.writeUint(t.baseDataOffset??0,8),t.flags&2&&a.writeUint(t.sampleDescriptionIndex??0,4),t.flags&8&&a.writeUint(t.defaultSampleDuration??0,4),t.flags&16&&a.writeUint(t.defaultSampleSize??0,4),t.flags&32&&a.writeUint(t.defaultSampleFlags??0,4),a}function qi(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("tkhd",n+s+i+4+4+8+2+2+2+2+36+4+4);return r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.trackId,4),r.writeUint(t.reserved1,4),r.writeUint(t.duration,e),r.writeArray(t.reserved2,C,4,2),r.writeUint(t.layer,2),r.writeUint(t.alternateGroup,2),r.writeTemplate(t.volume,2),r.writeUint(t.reserved3,2),r.writeArray(t.matrix,ne,4,9),r.writeTemplate(t.width,4),r.writeTemplate(t.height,4),r}function Gi(t){const e=new E("trex",32);return e.writeFullBox(t.version,t.flags),e.writeUint(t.trackId,4),e.writeUint(t.defaultSampleDescriptionIndex,4),e.writeUint(t.defaultSampleDuration,4),e.writeUint(t.defaultSampleSize,4),e.writeUint(t.defaultSampleFlags,4),e}function Xi(t){const e=t.flags&1?4:0,n=t.flags&4?4:0;let s=0;t.flags&256&&(s+=4),t.flags&512&&(s+=4),t.flags&1024&&(s+=4),t.flags&2048&&(s+=4);const i=s*t.sampleCount,r=new E("trun",16+e+n+i);r.writeFullBox(t.version,t.flags),r.writeUint(t.sampleCount,4),t.flags&1&&r.writeUint(t.dataOffset??0,4),t.flags&4&&r.writeUint(t.firstSampleFlags??0,4);for(const a of t.samples)t.flags&256&&r.writeUint(a.sampleDuration??0,4),t.flags&512&&r.writeUint(a.sampleSize??0,4),t.flags&1024&&r.writeUint(a.sampleFlags??0,4),t.flags&2048&&r.writeUint(a.sampleCompositionTimeOffset??0,4);return r}function Yi(t){const e=t.location.length+1,n=new E("url ",12+e);return n.writeFullBox(t.version,t.flags),n.writeTerminatedString(t.location),n}function Ki(t){const e=new E("vmhd",20);return e.writeFullBox(t.version,t.flags),e.writeUint(t.graphicsmode,2),e.writeArray(t.opcolor,C,2,3),e}const Ji={avc1:xt,avc3:xt,hvc1:_i,hev1:xi,mp4a:Mi,stsd:zi,mdhd:Ei,tkhd:Ii,trex:Li},gn={mfhd:Ui,tfhd:Ci,tfdt:Ti,trun:Bi,mdat:Si};function j(t,e){for(const n of t){if(e(n))return n;const s=n.boxes;if(s&&Array.isArray(s)){const i=j(s,e);if(i)return i}}}function et(t){const e=new ArrayBuffer(t.byteLength);return new Uint8Array(e).set(t),e}function H(t){return e=>e.type===t}function Ie(t){const e=Je(et(t),{readers:Ji}),n=j(e,H("mdhd"));if(!n)throw new Error("No mdhd box found in init segment");const s=j(e,H("tkhd"))?.trackId??1,i=j(e,H("stsd"));if(!i?.entries||i.entries.length===0)throw new Error("No stsd box found in init segment");const r=i.entries[0],a=Qi(r),o=j(e,c=>c.type==="trex"&&c.trackId===s);return{description:a,timescale:n.timescale,trackId:s,defaultSampleDuration:o?.defaultSampleDuration??0,defaultSampleSize:o?.defaultSampleSize??0,defaultSampleFlags:o?.defaultSampleFlags??0}}function Qi(t){if(!(!t.boxes||!Array.isArray(t.boxes)))for(const e of t.boxes){if(e instanceof Uint8Array){if(e.length>8){const s=String.fromCharCode(e[4],e[5],e[6],e[7]);if(s==="avcC"||s==="hvcC"||s==="dOps")return new Uint8Array(e.slice(8));if(s==="esds")return _t(new Uint8Array(e.slice(8)))}continue}const n=e.type;if(n==="avcC"||n==="hvcC"||n==="dOps"){if(e.view){const s=e.view,i=8,r=s.byteOffset+i,a=e.size-i;return new Uint8Array(s.buffer,r,a)}if(e.data instanceof Uint8Array)return new Uint8Array(e.data);if(e.raw instanceof Uint8Array)return new Uint8Array(e.raw.slice(8))}if(n==="esds"){let s;if(e.view){const i=e.view,r=8;s=new Uint8Array(i.buffer,i.byteOffset+r,e.size-r)}else e.data instanceof Uint8Array?s=new Uint8Array(e.data):e.raw instanceof Uint8Array&&(s=new Uint8Array(e.raw.slice(8)));if(s)return _t(s)}}}function _t(t){let e=4;for(;e<t.length;){const n=t[e++];let s=0;for(let i=0;i<4&&e<t.length;i++){const r=t[e++];if(s=s<<7|r&127,(r&128)===0)break}if(n===5)return e+s<=t.length?new Uint8Array(t.buffer,t.byteOffset+e,s):void 0;n===3?e+=3:n===4?e+=13:e+=s}}function bn(t,e){const n=Je(et(t),{readers:gn});return(j(n,H("tfdt"))?.baseMediaDecodeTime??0)*1e6/e.timescale}function er(t,e){const n=Je(et(t),{readers:gn}),s=j(n,H("tfdt"))?.baseMediaDecodeTime??0,i=j(n,H("tfhd")),r=i?.defaultSampleDuration??e.defaultSampleDuration,a=i?.defaultSampleSize??e.defaultSampleSize,o=i?.defaultSampleFlags??e.defaultSampleFlags,c=j(n,H("trun"));if(!c)throw new Error("No trun box found in data segment");const u=j(n,H("mdat"));if(!u)throw new Error("No mdat box found in data segment");const d=u.data;if(!d)throw new Error("No data in mdat box");const f=[];let m=0,b=s;for(let w=0;w<c.sampleCount;w++){const k=c.samples[w]??{},_=k.sampleSize??a,T=k.sampleDuration??r;if(_<=0)throw new Error(`Invalid sample size ${_} for sample ${w} in trun`);if(T<0)throw new Error(`Invalid sample duration ${T} for sample ${w} in trun`);if(m+_>d.length)throw new Error(`Sample ${w} would overflow mdat: offset=${m}, size=${_}, mdatLength=${d.length}`);const x=w===0&&c.firstSampleFlags!==void 0?c.firstSampleFlags:k.sampleFlags??o,P=k.sampleCompositionTimeOffset??0,F=new Uint8Array(d.slice(m,m+_));m+=_;const O=b+P,y=Math.round(O*1e6/e.timescale),$=x===0||(x&65536)===0;f.push({data:F,timestamp:y,keyframe:$}),b+=T}return f}function V(t){if(t=t.startsWith("0x")?t.slice(2):t,t.length%2)throw new Error("invalid hex string length");const e=t.match(/.{2}/g);if(!e)throw new Error("invalid hex string format");return new Uint8Array(e.map(n=>parseInt(n,16)))}const _e=[65536,0,0,0,65536,0,0,0,1073741824],tr={ftyp:Oi,mvhd:ji,tkhd:qi,mdhd:$i,hdlr:Di,vmhd:Ki,smhd:Pi,"url ":Yi,dref:Ri,stsd:Vi,stts:Wi,trex:Gi,mfhd:Ni,tfhd:Hi,tfdt:Zi,trun:Xi,mdat:Fi};function he(t){return ki(t,{writers:tr})}function tt(t,e,n,s){const i=12+s.length,r=new Uint8Array(i),a=new DataView(r.buffer);return a.setUint32(0,i,!1),r[4]=t.charCodeAt(0),r[5]=t.charCodeAt(1),r[6]=t.charCodeAt(2),r[7]=t.charCodeAt(3),a.setUint32(8,e<<24|n,!1),r.set(s,12),r}function wn(){const t=new Uint8Array(4);return tt("stsc",0,0,t)}function yn(){const t=new Uint8Array(8);return tt("stsz",0,0,t)}function vn(){const t=new Uint8Array(4);return tt("stco",0,0,t)}function nr(t,e,n){const s=8+n.length,i=8+(78+s),r=new Uint8Array(i),a=new DataView(r.buffer);let o=0;return a.setUint32(o,i,!1),o+=4,r[o++]=97,r[o++]=118,r[o++]=99,r[o++]=49,o+=6,a.setUint16(o,1,!1),o+=2,a.setUint16(o,0,!1),o+=2,a.setUint16(o,0,!1),o+=2,o+=12,a.setUint16(o,t,!1),o+=2,a.setUint16(o,e,!1),o+=2,a.setUint32(o,4718592,!1),o+=4,a.setUint32(o,4718592,!1),o+=4,a.setUint32(o,0,!1),o+=4,a.setUint16(o,1,!1),o+=2,o+=32,a.setUint16(o,24,!1),o+=2,a.setUint16(o,65535,!1),o+=2,a.setUint32(o,s,!1),o+=4,r[o++]=97,r[o++]=118,r[o++]=99,r[o++]=67,r.set(n,o),r}function sr(t){const{codedWidth:e,codedHeight:n,description:s}=t;if(!e||!n||!s)throw new Error("Missing required fields to create video init segment");const i=1e6,r=1,a={type:"ftyp",majorBrand:"isom",minorVersion:512,compatibleBrands:["isom","iso6","mp41"]},o={type:"mvhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:i,duration:0,rate:65536,volume:256,reserved1:0,reserved2:[0,0],matrix:_e,preDefined:[0,0,0,0,0,0],nextTrackId:r+1},c={type:"tkhd",version:0,flags:3,creationTime:0,modificationTime:0,trackId:r,reserved1:0,duration:0,reserved2:[0,0],layer:0,alternateGroup:0,volume:0,reserved3:0,matrix:_e,width:e*65536,height:n*65536},u={type:"mdhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:i,duration:0,language:"und",preDefined:0},d={type:"hdlr",version:0,flags:0,preDefined:0,handlerType:"vide",reserved:[0,0,0],name:"VideoHandler"},f={type:"vmhd",version:0,flags:1,graphicsmode:0,opcolor:[0,0,0]},m={type:"dinf",boxes:[{type:"dref",version:0,flags:0,entryCount:1,entries:[{type:"url ",version:0,flags:1,location:""}]}]},b={type:"stsd",version:0,flags:0,entryCount:1,entries:[nr(e,n,V(s))]},w={type:"stts",version:0,flags:0,entryCount:0,entries:[]},k=wn(),_=yn(),T=vn(),x=he([a,{type:"moov",boxes:[o,{type:"trak",boxes:[c,{type:"mdia",boxes:[u,d,{type:"minf",boxes:[f,m,{type:"stbl",boxes:[b,w,k,_,T]}]}]}]},{type:"mvex",boxes:[{type:"trex",version:0,flags:0,trackId:r,defaultSampleDescriptionIndex:1,defaultSampleDuration:0,defaultSampleSize:0,defaultSampleFlags:0}]}]}]),P=x.reduce((y,$)=>y+$.byteLength,0),F=new Uint8Array(P);let O=0;for(const y of x)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;return F}function ir(t){const{sampleRate:e,numberOfChannels:n,description:s,codec:i}=t,r=1e6,a=1,o={type:"ftyp",majorBrand:"isom",minorVersion:512,compatibleBrands:["isom","iso6","mp41"]},c={type:"mvhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:r,duration:0,rate:65536,volume:256,reserved1:0,reserved2:[0,0],matrix:_e,preDefined:[0,0,0,0,0,0],nextTrackId:a+1},u={type:"tkhd",version:0,flags:3,creationTime:0,modificationTime:0,trackId:a,reserved1:0,duration:0,reserved2:[0,0],layer:0,alternateGroup:0,volume:256,reserved3:0,matrix:_e,width:0,height:0},d={type:"mdhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:r,duration:0,language:"und",preDefined:0},f={type:"hdlr",version:0,flags:0,preDefined:0,handlerType:"soun",reserved:[0,0,0],name:"SoundHandler"},m={type:"smhd",version:0,flags:0,balance:0,reserved:0},b={type:"dinf",boxes:[{type:"dref",version:0,flags:0,entryCount:1,entries:[{type:"url ",version:0,flags:1,location:""}]}]},w={type:"stsd",version:0,flags:0,entryCount:1,entries:[rr(i,e,n,s)]},k={type:"stts",version:0,flags:0,entryCount:0,entries:[]},_=wn(),T=yn(),x=vn(),P=he([o,{type:"moov",boxes:[c,{type:"trak",boxes:[u,{type:"mdia",boxes:[d,f,{type:"minf",boxes:[m,b,{type:"stbl",boxes:[w,k,_,T,x]}]}]}]},{type:"mvex",boxes:[{type:"trex",version:0,flags:0,trackId:a,defaultSampleDescriptionIndex:1,defaultSampleDuration:0,defaultSampleSize:0,defaultSampleFlags:0}]}]}]),F=P.reduce(($,Cn)=>$+Cn.byteLength,0),O=new Uint8Array(F);let y=0;for(const $ of P)O.set(new Uint8Array($.buffer,$.byteOffset,$.byteLength),y),y+=$.byteLength;return O}function rr(t,e,n,s){if(t.startsWith("mp4a"))return or(e,n,s);if(t==="opus")return ar(e,n,s);throw new Error(`Unsupported audio codec: ${t}`)}function or(t,e,n){const s=ur(t,e,n),i=8+(28+s.length),r=new Uint8Array(i),a=new DataView(r.buffer);let o=0;return a.setUint32(o,i,!1),o+=4,r[o++]=109,r[o++]=112,r[o++]=52,r[o++]=97,o+=6,a.setUint16(o,1,!1),o+=2,o+=8,a.setUint16(o,e,!1),o+=2,a.setUint16(o,16,!1),o+=2,a.setUint16(o,0,!1),o+=2,a.setUint16(o,0,!1),o+=2,a.setUint32(o,t*65536,!1),o+=4,r.set(s,o),r}function ar(t,e,n){const s=dr(e,t,n),i=8+(28+s.length),r=new Uint8Array(i),a=new DataView(r.buffer);let o=0;return a.setUint32(o,i,!1),o+=4,r[o++]=79,r[o++]=112,r[o++]=117,r[o++]=115,o+=6,a.setUint16(o,1,!1),o+=2,o+=8,a.setUint16(o,e,!1),o+=2,a.setUint16(o,16,!1),o+=2,a.setUint16(o,0,!1),o+=2,a.setUint16(o,0,!1),o+=2,a.setUint32(o,t*65536,!1),o+=4,r.set(s,o),r}function cr(t,e){const n={96e3:0,88200:1,64e3:2,48e3:3,44100:4,32e3:5,24e3:6,22050:7,16e3:8,12e3:9,11025:10,8e3:11,7350:12}[t]??4,s=16|n>>1,i=(n&1)<<7|e<<3;return new Uint8Array([s,i])}function ur(t,e,n){const s=n?V(n):cr(t,e),i=s.length,r=15+i,a=5+r+3,o=14+a,c=new Uint8Array(o),u=new DataView(c.buffer);let d=0;return u.setUint32(d,o,!1),d+=4,c[d++]=101,c[d++]=115,c[d++]=100,c[d++]=115,u.setUint32(d,0,!1),d+=4,c[d++]=3,c[d++]=a,u.setUint16(d,0,!1),d+=2,c[d++]=0,c[d++]=4,c[d++]=r,c[d++]=64,c[d++]=21,c[d++]=0,c[d++]=0,c[d++]=0,u.setUint32(d,0,!1),d+=4,u.setUint32(d,0,!1),d+=4,c[d++]=5,c[d++]=i,c.set(s,d),d+=i,c[d++]=6,c[d++]=1,c[d++]=2,c}function dr(t,e,n){if(n){const o=V(n),c=8+o.length,u=new Uint8Array(c);return new DataView(u.buffer).setUint32(0,c,!1),u[4]=100,u[5]=79,u[6]=112,u[7]=115,u.set(o,8),u}const s=19,i=new Uint8Array(s),r=new DataView(i.buffer);let a=0;return r.setUint32(a,s,!1),a+=4,i[a++]=100,i[a++]=79,i[a++]=112,i[a++]=115,i[a++]=0,i[a++]=t,r.setUint16(a,312,!1),a+=2,r.setUint32(a,e,!1),a+=4,r.setInt16(a,0,!1),a+=2,i[a++]=0,i}function kn(t){const{data:e,timestamp:n,duration:s,keyframe:i,sequence:r,trackId:a=1}=t,o=i?33554432:16842752,c={type:"mfhd",version:0,flags:0,sequenceNumber:r},u={type:"tfhd",version:0,flags:131072,trackId:a},d={type:"tfdt",version:1,flags:0,baseMediaDecodeTime:n},f={type:"trun",version:0,flags:1793,sampleCount:1,dataOffset:0,samples:[{sampleDuration:s,sampleSize:e.byteLength,sampleFlags:o}]},m={type:"moof",boxes:[c,{type:"traf",boxes:[u,d,f]}]},b=he([m]);let w=0;for(const y of b)w+=y.byteLength;f.dataOffset=w+8;const k=he([m]);w=0;for(const y of k)w+=y.byteLength;const _=new ArrayBuffer(e.byteLength),T=new Uint8Array(_);T.set(e);const x=he([{type:"mdat",data:T}]);let P=0;for(const y of x)P+=y.byteLength;const F=new Uint8Array(w+P);let O=0;for(const y of k)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;for(const y of x)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;return F}let An=class{#e;constructor(t){this.#e=t}decode(t){return er(t,this.#e).map(e=>({data:e.data,timestamp:e.timestamp,keyframe:e.keyframe}))}};class se{#e;#t;#n;#i=[];#s;#r;#a=new l([]);buffered=this.#a;#o=new S;constructor(e,n){this.#e=e,this.#t=n.format,this.#n=l.from(n.latency??h.zero),this.#o.spawn(this.#c.bind(this)),this.#o.cleanup(()=>{this.#e.close();for(const s of this.#i)s.consumer.close();this.#i.length=0})}async#c(){for(;;){const e=await this.#e.recvGroup();if(!e)break;if(this.#s===void 0&&(this.#s=e.sequence),e.sequence<this.#s){console.warn(`skipping old group: ${e.sequence} < ${this.#s}`),e.close();continue}const n={consumer:e,frames:[]};this.#i.push(n),this.#i.sort((s,i)=>s.consumer.sequence-i.consumer.sequence),this.#o.spawn(this.#u.bind(this,n))}}async#u(e){try{let n=0;for(;;){const s=await e.consumer.readFrame();if(!s)break;const i=this.#t.decode(s);for(const r of i){const a={data:r.data,timestamp:r.timestamp,keyframe:n===0?!0:r.keyframe};n++,e.frames.push(a),(e.latest===void 0||a.timestamp>e.latest)&&(e.latest=a.timestamp),this.#d(),e.consumer.sequence===this.#s?(this.#r?.(),this.#r=void 0):this.#l()}}}catch{}finally{e.done=!0,e.consumer.sequence===this.#s&&(this.#s+=1),this.#d(),this.#r?.(),this.#r=void 0,e.consumer.close()}}#l(){if(this.#s===void 0)return;let e=!1;for(;this.#i.length>=2;){const n=Me.fromMilli(this.#n.peek());let s,i;for(const a of this.#i){if(a.latest===void 0)continue;const o=a.frames.at(0)?.timestamp??a.latest;(s===void 0||o<s)&&(s=o),(i===void 0||a.latest>i)&&(i=a.latest)}if(s===void 0||i===void 0||i-s<=n)break;const r=this.#i.shift();if(!r)break;this.#s=this.#i[0]?.consumer.sequence,console.warn(`skipping slow group: ${r.consumer.sequence} -> ${this.#s}`),r.consumer.close(),r.frames.length=0,e=!0}e&&(this.#d(),this.#r?.(),this.#r=void 0)}async next(){for(;;){if(this.#i.length>0&&this.#s!==void 0&&this.#i[0].consumer.sequence<=this.#s){const n=this.#i[0].frames.shift();if(n)return this.#d(),{frame:n,group:this.#i[0].consumer.sequence};if(this.#s>this.#i[0].consumer.sequence||this.#i[0].done){this.#i[0].consumer.sequence===this.#s&&(this.#s+=1);const s=this.#i.shift();if(s)return this.#d(),{frame:void 0,group:s.consumer.sequence}}}if(this.#r)throw new Error("multiple calls to next not supported");const e=new Promise(n=>{this.#r=n}).then(()=>!0);if(!await Promise.race([e,this.#o.closed])){this.#r=void 0;return}}}#d(){const e=[];let n;for(const s of this.#i){const i=s.frames.at(0);if(!i||s.latest===void 0)continue;const r=h.fromMicro(i.timestamp),a=h.fromMicro(s.latest),o=e.at(-1),c=n?.done&&n.consumer.sequence+1===s.consumer.sequence;o&&(o.end>=r||c)?o.end=h.max(o.end,a):e.push({start:r,end:a}),n=s}this.#a.set(e)}close(){this.#o.close()}}class Le{decode(e){const[n,s]=Ln(e);return[{data:s,timestamp:n,keyframe:!1}]}}function Se(t,e){if(t.length===0)return e;if(e.length===0)return t;const n=[],s=[...t,...e].sort((i,r)=>i.start-r.start);for(const i of s){const r=n.at(-1);r&&r.end>=i.start?r.end=h.max(r.end,i.end):n.push({...i})}return n}let Re;async function St(){return globalThis.AudioEncoder&&globalThis.AudioDecoder?!0:(Re||(console.warn("using Opus polyfill; performance may be degraded"),Re=Promise.all([at(()=>import("./libav-opus-af-BlMWboA7-CFTeN5TA.js"),[],import.meta.url),at(()=>import("./main-DGBFe0O7-CIZu5tmC.js"),[],import.meta.url)]).then(async([t,e])=>(await e.load({LibAV:t,polyfill:!0}),!0))),await Re)}function be(t){const e=atob(t),n=new Uint8Array(e.length);for(let s=0;s<e.length;s++)n[s]=e.charCodeAt(s);return n}const W=0,N=1,oe=2,Y=3,lr=4;function xn(t,e,n){if(t<=0)throw new Error("invalid channels");if(e<=0)throw new Error("invalid capacity");if(n<=0)throw new Error("invalid sample rate");const s=new SharedArrayBuffer(t*e*Float32Array.BYTES_PER_ELEMENT),i=new SharedArrayBuffer(lr*Int32Array.BYTES_PER_ELEMENT),r=new Int32Array(i);return Atomics.store(r,Y,1),{channels:t,capacity:e,rate:n,samples:s,control:i}}function fr(t,e){return(t-e|0)>0?t:e}function ae(t,e){return(t%e+e)%e}function Oe(t,e,n){for(;;){const s=Atomics.load(t,e);if((n-s|0)<=0)return s;if(Atomics.compareExchange(t,e,s,n)===s)return n}}class nt{channels;capacity;rate;init;#e;#t;constructor(e){this.channels=e.channels,this.capacity=e.capacity,this.rate=e.rate,this.init=e,this.#e=new Int32Array(e.control),this.#t=[];for(let n=0;n<this.channels;n++)this.#t.push(new Float32Array(e.samples,n*this.capacity*Float32Array.BYTES_PER_ELEMENT,this.capacity))}insert(e,n){if(n.length!==this.channels)throw new Error("wrong number of channels");let s=Math.round(je.fromMicro(e)*this.rate);const i=n[0].length;let r=0;const a=s+i|0,o=Atomics.load(this.#e,N),c=o-s|0;if(c>0){if(c>=i)return;r=c,s=s+c|0}const u=i-r;(a-o|0)>this.capacity&&Oe(this.#e,N,a-this.capacity|0);const d=Atomics.load(this.#e,W),f=s-d|0;if(f>0){const k=Math.min(f,this.capacity);for(let _=0;_<this.channels;_++){const T=this.#t[_];for(let x=0;x<k;x++)T[ae(d+x|0,this.capacity)]=0}}for(let k=0;k<this.channels;k++){const _=n[k],T=this.#t[k];for(let x=0;x<u;x++)T[ae(s+x|0,this.capacity)]=_[r+x]}Atomics.store(this.#e,W,fr(Atomics.load(this.#e,W),a));const m=Atomics.load(this.#e,N),b=Atomics.load(this.#e,W),w=Atomics.load(this.#e,oe);(b-m|0)>=w&&w>0&&Atomics.store(this.#e,Y,0)}read(e){if(Atomics.load(this.#e,Y)===1)return 0;let n=Atomics.load(this.#e,N);const s=Atomics.load(this.#e,W),i=Atomics.load(this.#e,oe),r=s-n|0;if(i>0&&r>i){const c=s-i|0;n=Oe(this.#e,N,c)}const a=s-n|0,o=Math.min(a,e[0].length);if(o<=0)return 0;for(let c=0;c<this.channels;c++){const u=this.#t[c],d=e[c];for(let f=0;f<o;f++)d[f]=u[ae(n+f|0,this.capacity)]}return Oe(this.#e,N,n+o|0),o}setLatency(e){Atomics.store(this.#e,oe,e)}resize(e){const n=xn(this.channels,e,this.rate),s=new nt(n),i=Atomics.load(this.#e,N),r=Atomics.load(this.#e,W),a=Atomics.load(this.#e,oe),o=Atomics.load(this.#e,Y),c=r-i|0,u=Math.max(0,Math.min(c,s.capacity)),d=r-u|0;for(let f=0;f<this.channels;f++){const m=this.#t[f],b=s.#t[f];for(let w=0;w<u;w++){const k=d+w|0;b[ae(k,s.capacity)]=m[ae(k,this.capacity)]}}return Atomics.store(s.#e,N,d),Atomics.store(s.#e,W,r),Atomics.store(s.#e,oe,a),Atomics.store(s.#e,Y,o),s}get timestamp(){const e=Atomics.load(this.#e,N);return Me.fromSecond(e/this.rate)}get stalled(){return Atomics.load(this.#e,Y)===1}get length(){return Atomics.load(this.#e,W)-Atomics.load(this.#e,N)|0}}function hr(){return!(typeof SharedArrayBuffer>"u"||typeof crossOriginIsolated<"u"&&!crossOriginIsolated)}function mr(t,e,n,s){return hr()?(console.log("[audio] using SharedArrayBuffer audio buffer"),new pr(t,e,n,s)):(console.log("[audio] using postMessage audio buffer (SharedArrayBuffer unavailable)"),new gr(t,e,n,s))}class pr{rate;channels;#e;#t;#n=new l(0);timestamp=this.#n;#i=new l(!0);stalled=this.#i;#s=new S;constructor(e,n,s,i){this.#e=e,this.channels=n,this.rate=s;const r=Math.max(s,i*2),a=xn(n,r,s);this.#t=new nt(a),this.#t.setLatency(i);const o={type:"init-shared",...a};e.port.postMessage(o),this.#s.interval(()=>{this.#n.set(this.#t.timestamp),this.#i.set(this.#t.stalled)},50)}insert(e,n){this.#t.insert(e,n)}setLatency(e){if(this.#t.capacity<e*1.5){const n=Math.max(this.rate,e*2);this.#t=this.#t.resize(n),this.#t.setLatency(e);const s={type:"init-shared",...this.#t.init};this.#e.port.postMessage(s)}else this.#t.setLatency(e)}close(){this.#s.close()}}class gr{rate;channels;#e;#t=new l(0);timestamp=this.#t;#n=new l(!0);stalled=this.#n;#i=new S;constructor(e,n,s,i){this.#e=e,this.channels=n,this.rate=s;const r=h.fromSecond(i/s),a={type:"init-post",channels:n,rate:s,latency:r};e.port.postMessage(a),this.#i.event(e.port,"message",o=>{const c=o.data;c?.type==="state"&&(this.#t.set(c.timestamp),this.#n.set(c.stalled))}),e.port.start()}insert(e,n){const s={type:"data",data:n,timestamp:e};this.#e.port.postMessage(s,n.map(i=>i.buffer))}setLatency(e){const n={type:"latency",latency:h.fromSecond(e/this.rate)};this.#e.port.postMessage(n)}close(){this.#i.close()}}const br=`var __defProp = Object.defineProperty;
-var __getOwnPropNames = Object.getOwnPropertyNames;
-var __esm = (fn, res) => function __init() {
-  return fn && (res = (0, fn[__getOwnPropNames(fn)[0]])(fn = 0)), res;
-};
-var __export = (target, all) => {
-  for (var name in all)
-    __defProp(target, name, { get: all[name], enumerable: true });
-};
-
-// ../lite/src/path.ts
-var init_path = __esm({
-  "../lite/src/path.ts"() {
-    "use strict";
-  }
-});
-
-// ../lite/src/varint.ts
-var MAX_U6, MAX_U14, MAX_U30, MAX_U53, MAX_U64, MAX_U62;
-var init_varint = __esm({
-  "../lite/src/varint.ts"() {
-    "use strict";
-    MAX_U6 = 2 ** 6 - 1;
-    MAX_U14 = 2 ** 14 - 1;
-    MAX_U30 = 2 ** 30 - 1;
-    MAX_U53 = Number.MAX_SAFE_INTEGER;
-    MAX_U64 = (1n << 64n) - 1n;
-    MAX_U62 = 2n ** 62n - 1n;
-  }
-});
-
-// ../lite/src/index.ts
-init_path();
-
-// ../lite/src/time.ts
-var time_exports = {};
-__export(time_exports, {
-  Micro: () => Micro,
-  Milli: () => Milli,
-  Nano: () => Nano,
-  Second: () => Second
-});
-var Nano = {
-  zero: 0,
-  fromMicro: (us) => us * 1e3,
-  fromMilli: (ms) => ms * 1e6,
-  fromSecond: (s) => s * 1e9,
-  toMicro: (ns) => ns / 1e3,
-  toMilli: (ns) => ns / 1e6,
-  toSecond: (ns) => ns / 1e9,
-  now: () => performance.now() * 1e6,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Micro = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e3,
-  fromMilli: (ms) => ms * 1e3,
-  fromSecond: (s) => s * 1e6,
-  toNano: (us) => us * 1e3,
-  toMilli: (us) => us / 1e3,
-  toSecond: (us) => us / 1e6,
-  now: () => performance.now() * 1e3,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Milli = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e6,
-  fromMicro: (us) => us / 1e3,
-  fromSecond: (s) => s * 1e3,
-  toNano: (ms) => ms * 1e6,
-  toMicro: (ms) => ms * 1e3,
-  toSecond: (ms) => ms / 1e3,
-  now: () => performance.now(),
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Second = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e9,
-  fromMicro: (us) => us / 1e6,
-  fromMilli: (ms) => ms / 1e3,
-  toNano: (s) => s * 1e9,
-  toMicro: (s) => s * 1e6,
-  toMilli: (s) => s * 1e3,
-  now: () => performance.now() / 1e3,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-
-// ../lite/src/index.ts
-init_varint();
-
-// src/audio/ring-buffer.ts
-var AudioRingBuffer = class {
-  #buffer;
-  #writeIndex = 0;
-  #readIndex = 0;
-  rate;
-  channels;
-  #stalled = true;
-  constructor(props) {
-    if (props.channels <= 0) throw new Error("invalid channels");
-    if (props.rate <= 0) throw new Error("invalid sample rate");
-    if (props.latency <= 0) throw new Error("invalid latency");
-    const samples = Math.ceil(props.rate * time_exports.Second.fromMilli(props.latency));
-    if (samples === 0) throw new Error("empty buffer");
-    this.rate = props.rate;
-    this.channels = props.channels;
-    this.#buffer = [];
-    for (let i = 0; i < this.channels; i++) {
-      this.#buffer[i] = new Float32Array(samples);
-    }
-  }
-  get stalled() {
-    return this.#stalled;
-  }
-  get timestamp() {
-    return time_exports.Micro.fromSecond(this.#readIndex / this.rate);
-  }
-  get length() {
-    return this.#writeIndex - this.#readIndex;
-  }
-  get capacity() {
-    return this.#buffer[0]?.length;
-  }
-  resize(latency) {
-    const newCapacity = Math.ceil(this.rate * time_exports.Second.fromMilli(latency));
-    if (newCapacity === this.capacity) return;
-    if (newCapacity === 0) throw new Error("empty buffer");
-    const newBuffer = [];
-    for (let i = 0; i < this.channels; i++) {
-      newBuffer[i] = new Float32Array(newCapacity);
-    }
-    const samplesToKeep = Math.min(this.length, newCapacity);
-    if (samplesToKeep > 0) {
-      const copyStart = this.#writeIndex - samplesToKeep;
-      for (let channel = 0; channel < this.channels; channel++) {
-        const src = this.#buffer[channel];
-        const dst = newBuffer[channel];
-        for (let i = 0; i < samplesToKeep; i++) {
-          const srcPos = (copyStart + i) % src.length;
-          const dstPos = (copyStart + i) % dst.length;
-          dst[dstPos] = src[srcPos];
-        }
-      }
-    }
-    this.#buffer = newBuffer;
-    this.#readIndex = this.#writeIndex - samplesToKeep;
-    if (samplesToKeep === 0) this.#stalled = true;
-  }
-  write(timestamp, data) {
-    if (data.length !== this.channels) throw new Error("wrong number of channels");
-    let start = Math.round(time_exports.Second.fromMicro(timestamp) * this.rate);
-    let samples = data[0].length;
-    let offset = this.#readIndex - start;
-    if (offset > samples) {
-      return;
-    } else if (offset > 0) {
-      samples -= offset;
-      start += offset;
-    } else {
-      offset = 0;
-    }
-    const end = start + samples;
-    const overflow = end - this.#readIndex - this.#buffer[0].length;
-    if (overflow >= 0) {
-      this.#stalled = false;
-      this.#readIndex += overflow;
-    }
-    if (start > this.#writeIndex) {
-      const gapSize = Math.min(start - this.#writeIndex, this.#buffer[0].length);
-      if (gapSize === 1) {
-        console.warn("floating point inaccuracy detected");
-      }
-      for (let channel = 0; channel < this.channels; channel++) {
-        const dst = this.#buffer[channel];
-        for (let i = 0; i < gapSize; i++) {
-          const writePos = (this.#writeIndex + i) % dst.length;
-          dst[writePos] = 0;
-        }
-      }
-    }
-    for (let channel = 0; channel < this.channels; channel++) {
-      let src = data[channel];
-      src = src.subarray(src.length - samples);
-      const dst = this.#buffer[channel];
-      if (src.length !== samples) throw new Error("mismatching number of samples");
-      for (let i = 0; i < samples; i++) {
-        const writePos = (start + i) % dst.length;
-        dst[writePos] = src[i];
-      }
-    }
-    if (end > this.#writeIndex) {
-      this.#writeIndex = end;
-    }
-  }
-  read(output) {
-    if (output.length !== this.channels) throw new Error("wrong number of channels");
-    if (this.#stalled) return 0;
-    const samples = Math.min(this.#writeIndex - this.#readIndex, output[0].length);
-    if (samples === 0) return 0;
-    for (let channel = 0; channel < this.channels; channel++) {
-      const dst = output[channel];
-      const src = this.#buffer[channel];
-      if (dst.length !== output[0].length) throw new Error("mismatching number of samples");
-      for (let i = 0; i < samples; i++) {
-        const readPos = (this.#readIndex + i) % src.length;
-        dst[i] = src[readPos];
-      }
-    }
-    this.#readIndex += samples;
-    return samples;
-  }
-};
-
-// src/audio/shared-ring-buffer.ts
-var WRITE = 0;
-var READ = 1;
-var LATENCY = 2;
-var STALLED = 3;
-var CONTROL_SLOTS = 4;
-function allocSharedRingBuffer(channels, capacity, rate) {
-  if (channels <= 0) throw new Error("invalid channels");
-  if (capacity <= 0) throw new Error("invalid capacity");
-  if (rate <= 0) throw new Error("invalid sample rate");
-  const samples = new SharedArrayBuffer(channels * capacity * Float32Array.BYTES_PER_ELEMENT);
-  const control = new SharedArrayBuffer(CONTROL_SLOTS * Int32Array.BYTES_PER_ELEMENT);
-  const ctrl = new Int32Array(control);
-  Atomics.store(ctrl, STALLED, 1);
-  return { channels, capacity, rate, samples, control };
-}
-function i32Max(a, b) {
-  return (a - b | 0) > 0 ? a : b;
-}
-function slot(idx, capacity) {
-  return (idx % capacity + capacity) % capacity;
-}
-function casAdvance(arr, idx, candidate) {
-  for (; ; ) {
-    const current = Atomics.load(arr, idx);
-    if ((candidate - current | 0) <= 0) return current;
-    const witnessed = Atomics.compareExchange(arr, idx, current, candidate);
-    if (witnessed === current) return candidate;
-  }
-}
-var SharedRingBuffer = class _SharedRingBuffer {
-  channels;
-  capacity;
-  rate;
-  init;
-  #control;
-  #samples;
-  constructor(init) {
-    this.channels = init.channels;
-    this.capacity = init.capacity;
-    this.rate = init.rate;
-    this.init = init;
-    this.#control = new Int32Array(init.control);
-    this.#samples = [];
-    for (let i = 0; i < this.channels; i++) {
-      this.#samples.push(
-        new Float32Array(init.samples, i * this.capacity * Float32Array.BYTES_PER_ELEMENT, this.capacity)
-      );
-    }
-  }
-  /**
-   * Insert audio samples at the given timestamp.
-   * Main thread only. Handles out-of-order writes, gap filling, and overflow.
-   */
-  insert(timestamp, data) {
-    if (data.length !== this.channels) throw new Error("wrong number of channels");
-    let start = Math.round(time_exports.Second.fromMicro(timestamp) * this.rate);
-    const originalLength = data[0].length;
-    let offset = 0;
-    const end = start + originalLength | 0;
-    const read = Atomics.load(this.#control, READ);
-    const behind = read - start | 0;
-    if (behind > 0) {
-      if (behind >= originalLength) {
-        return;
-      }
-      offset = behind;
-      start = start + behind | 0;
-    }
-    const samples = originalLength - offset;
-    if ((end - read | 0) > this.capacity) {
-      casAdvance(this.#control, READ, end - this.capacity | 0);
-    }
-    const write = Atomics.load(this.#control, WRITE);
-    const gap = start - write | 0;
-    if (gap > 0) {
-      const gapSize = Math.min(gap, this.capacity);
-      for (let channel = 0; channel < this.channels; channel++) {
-        const dst = this.#samples[channel];
-        for (let i = 0; i < gapSize; i++) {
-          dst[slot(write + i | 0, this.capacity)] = 0;
-        }
-      }
-    }
-    for (let channel = 0; channel < this.channels; channel++) {
-      const src = data[channel];
-      const dst = this.#samples[channel];
-      for (let i = 0; i < samples; i++) {
-        dst[slot(start + i | 0, this.capacity)] = src[offset + i];
-      }
-    }
-    Atomics.store(this.#control, WRITE, i32Max(Atomics.load(this.#control, WRITE), end));
-    const currentRead = Atomics.load(this.#control, READ);
-    const currentWrite = Atomics.load(this.#control, WRITE);
-    const latency = Atomics.load(this.#control, LATENCY);
-    if ((currentWrite - currentRead | 0) >= latency && latency > 0) {
-      Atomics.store(this.#control, STALLED, 0);
-    }
-  }
-  /**
-   * Read audio samples into the output buffers.
-   * AudioWorklet only. Returns the number of samples read.
-   */
-  read(output) {
-    if (Atomics.load(this.#control, STALLED) === 1) return 0;
-    let read = Atomics.load(this.#control, READ);
-    const write = Atomics.load(this.#control, WRITE);
-    const latency = Atomics.load(this.#control, LATENCY);
-    const buffered = write - read | 0;
-    if (latency > 0 && buffered > latency) {
-      const skipTo = write - latency | 0;
-      read = casAdvance(this.#control, READ, skipTo);
-    }
-    const available = write - read | 0;
-    const count = Math.min(available, output[0].length);
-    if (count <= 0) return 0;
-    for (let channel = 0; channel < this.channels; channel++) {
-      const src = this.#samples[channel];
-      const dst = output[channel];
-      for (let i = 0; i < count; i++) {
-        dst[i] = src[slot(read + i | 0, this.capacity)];
-      }
-    }
-    casAdvance(this.#control, READ, read + count | 0);
-    return count;
-  }
-  /** Update the target latency in samples. */
-  setLatency(samples) {
-    Atomics.store(this.#control, LATENCY, samples);
-  }
-  /**
-   * Allocate a new ring with \`newCapacity\` samples and copy the unread window
-   * [READ, WRITE) plus control state into it. Used when growing capacity so
-   * we don't drop buffered audio. If \`newCapacity\` is smaller than the unread
-   * span, the oldest samples are truncated.
-   *
-   * Main thread only. \`resize()\` reads from the source \`SharedRingBuffer\` and
-   * writes into a freshly allocated buffer from \`allocSharedRingBuffer\`, so it
-   * relies on the same invariant as \`insert()\`: no concurrent main-thread
-   * writers. The AudioWorklet reader is tolerated via the CAS discipline used
-   * by READ/WRITE elsewhere.
-   */
-  resize(newCapacity) {
-    const init = allocSharedRingBuffer(this.channels, newCapacity, this.rate);
-    const dst = new _SharedRingBuffer(init);
-    const read = Atomics.load(this.#control, READ);
-    const write = Atomics.load(this.#control, WRITE);
-    const latency = Atomics.load(this.#control, LATENCY);
-    const stalled = Atomics.load(this.#control, STALLED);
-    const available = write - read | 0;
-    const copyCount = Math.max(0, Math.min(available, dst.capacity));
-    const copyStart = write - copyCount | 0;
-    for (let channel = 0; channel < this.channels; channel++) {
-      const src = this.#samples[channel];
-      const out = dst.#samples[channel];
-      for (let i = 0; i < copyCount; i++) {
-        const idx = copyStart + i | 0;
-        out[slot(idx, dst.capacity)] = src[slot(idx, this.capacity)];
-      }
-    }
-    Atomics.store(dst.#control, READ, copyStart);
-    Atomics.store(dst.#control, WRITE, write);
-    Atomics.store(dst.#control, LATENCY, latency);
-    Atomics.store(dst.#control, STALLED, stalled);
-    return dst;
-  }
-  /** Current playback timestamp derived from READ position. */
-  get timestamp() {
-    const read = Atomics.load(this.#control, READ);
-    return time_exports.Micro.fromSecond(read / this.rate);
-  }
-  /** Whether the buffer is stalled (waiting to fill). */
-  get stalled() {
-    return Atomics.load(this.#control, STALLED) === 1;
-  }
-  /**
-   * Number of buffered samples (WRITE - READ).
-   *
-   * Non-atomic: WRITE and READ are loaded separately, so a concurrent
-   * writer/reader can make the two loads inconsistent. Intended for
-   * tests and diagnostics, not control-flow decisions.
-   */
-  get length() {
-    return Atomics.load(this.#control, WRITE) - Atomics.load(this.#control, READ) | 0;
-  }
-};
-
-// src/audio/render-worklet.ts
-var Render = class extends AudioWorkletProcessor {
-  // Set after init, depending on which path the main thread chose.
-  #backend;
-  #underflow = 0;
-  #stateCounter = 0;
-  constructor() {
-    super();
-    this.port.onmessage = (event) => {
-      const msg = event.data;
-      if (msg.type === "init-shared") {
-        console.log("[audio-worklet] init-shared: using SharedArrayBuffer path");
-        this.#backend = new SharedRingBuffer(msg);
-        this.#underflow = 0;
-      } else if (msg.type === "init-post") {
-        console.log("[audio-worklet] init-post: using postMessage path");
-        this.#backend = new AudioRingBuffer(msg);
-        this.#underflow = 0;
-      } else if (msg.type === "data") {
-        if (this.#backend instanceof AudioRingBuffer) this.#backend.write(msg.timestamp, msg.data);
-      } else if (msg.type === "latency") {
-        if (this.#backend instanceof AudioRingBuffer) this.#backend.resize(msg.latency);
-      }
-    };
-  }
-  process(_inputs, outputs, _parameters) {
-    const output = outputs[0];
-    const backend = this.#backend;
-    const samplesRead = backend?.read(output) ?? 0;
-    if (samplesRead < output[0].length) {
-      this.#underflow += output[0].length - samplesRead;
-    } else if (this.#underflow > 0 && backend) {
-      console.debug(\`audio underflow: \${Math.round(1e3 * this.#underflow / backend.rate)}ms\`);
-      this.#underflow = 0;
-    }
-    if (backend instanceof AudioRingBuffer) {
-      this.#stateCounter++;
-      if (this.#stateCounter >= 5) {
-        this.#stateCounter = 0;
-        const state = {
-          type: "state",
-          timestamp: backend.timestamp,
-          stalled: backend.stalled
-        };
-        this.port.postMessage(state);
-      }
-    }
-    return true;
-  }
-};
-registerProcessor("render", Render);
-`,wr=new Blob([br],{type:"application/javascript"}),yr=URL.createObjectURL(wr);let vr=class{source;enabled;#e=new l(void 0);context=this.#e;#t=new l(void 0);root=this.#t;#n=new l(void 0);sampleRate=this.#n;#i=new l(void 0);stats=this.#i;#s=new l(void 0);timestamp=this.#s;#r=new l(!0);stalled=this.#r;#a=new l([]);#o=new l([]);buffered=this.#o;#c;#u=new S;constructor(t,e){this.source=t,this.source.supported.set(kr),this.enabled=l.from(e?.enabled??!1),this.#u.run(this.#l.bind(this)),this.#u.run(this.#d.bind(this)),this.#u.run(this.#h.bind(this)),this.#u.run(this.#m.bind(this))}#l(t){const e=t.get(this.source.config);if(!e)return;const n=e.sampleRate,s=e.numberOfChannels,i=new AudioContext({latencyHint:"interactive",sampleRate:n});t.set(this.#e,i),t.cleanup(()=>i.close()),t.spawn(async()=>{if(await i.audioWorklet.addModule(yr),i.state==="closed")return;const r=new AudioWorkletNode(i,"render",{channelCount:s,channelCountMode:"explicit",outputChannelCount:[s]});t.cleanup(()=>r.disconnect());const a=this.source.sync.buffer.peek(),o=Math.ceil(n*je.fromMilli(a)),c=mr(r,s,n,o);this.#c=c,t.cleanup(()=>{c.close(),this.#c=void 0}),t.run(u=>{const d=h.fromMicro(u.get(c.timestamp));this.#s.set(d),this.#w(d)}),t.run(u=>{this.#r.set(u.get(c.stalled))}),t.set(this.#t,r)})}#d(t){const e=t.getAll([this.enabled,this.#e]);if(!e)return;const[n,s]=e;s.resume()}#h(t){if(!t.get(this.#t))return;const e=this.#c;if(!e)return;const n=t.get(this.source.sync.buffer),s=Math.ceil(e.rate*je.fromMilli(n));e.setLatency(s)}#m(t){if(!t.get(this.enabled))return;const e=t.get(this.source.broadcast);if(!e)return;const n=t.get(this.source.track);if(!n)return;const s=t.get(this.source.config);if(!s)return;const i=t.get(e.active);if(!i)return;const r=i.subscribe(n,te.audio);t.cleanup(()=>r.close()),s.container.kind==="cmaf"?this.#g(t,r,s):this.#p(t,r,s)}#p(t,e,n){const s=new se(e,{format:new Le,latency:this.source.sync.buffer});t.cleanup(()=>s.close()),t.run(i=>{const r=i.get(s.buffered),a=i.get(this.#a);this.#o.update(()=>Se(r,a))}),t.spawn(async()=>{if(!await St())return;let i=0;const r=new AudioDecoder({output:o=>{if(i++,i<=3){o.close();return}this.#f(o)},error:o=>console.error(o)});t.cleanup(()=>{r.state!=="closed"&&r.close()});const a=n.codec==="opus"?void 0:n.description?V(n.description):void 0;for(r.configure({...n,description:a});;){const o=await s.next();if(!o)break;const{frame:c}=o;if(!c)continue;const u=h.fromMicro(c.timestamp);this.source.sync.received(u,"audio"),this.#i.update(f=>({bytesReceived:(f?.bytesReceived??0)+c.data.byteLength}));const d=new EncodedAudioChunk({type:c.keyframe?"key":"delta",data:c.data,timestamp:c.timestamp});r.decode(d)}})}#g(t,e,n){if(n.container.kind!=="cmaf")return;const s=be(n.container.init),i=Ie(s),r=n.codec==="opus"?void 0:n.description?V(n.description):void 0,a=new se(e,{format:new An(i),latency:this.source.sync.buffer});t.cleanup(()=>a.close()),t.run(o=>{const c=o.get(a.buffered),u=o.get(this.#a);this.#o.update(()=>Se(c,u))}),t.spawn(async()=>{if(!await St())return;const o=new AudioDecoder({output:c=>this.#f(c),error:c=>console.error(c)});for(t.cleanup(()=>{o.state!=="closed"&&o.close()}),o.configure({codec:n.codec,sampleRate:n.sampleRate,numberOfChannels:n.numberOfChannels,description:r});;){const c=await Promise.race([a.next(),t.cancel]);if(!c)break;const{frame:u}=c;if(!u)continue;const d=h.fromMicro(u.timestamp);if(this.source.sync.received(d,"audio"),this.#i.update(f=>({bytesReceived:(f?.bytesReceived??0)+u.data.byteLength})),o.state==="closed")break;o.decode(new EncodedAudioChunk({type:u.keyframe?"key":"delta",data:u.data,timestamp:u.timestamp}))}})}#f(t){const e=t.timestamp,n=h.fromMicro(e),s=this.#c;if(!s){t.close();return}const i=t.numberOfFrames/t.sampleRate*1e6,r=h.fromMicro(i),a=h.add(n,r);this.#b(n,a);const o=Math.min(t.numberOfChannels,s.channels),c=[];for(let u=0;u<o;u++){const d=new Float32Array(t.numberOfFrames);t.copyTo(d,{format:"f32-planar",planeIndex:u}),c.push(d)}s.insert(e,c),t.close()}#b(t,e){t>e||this.#a.mutate(n=>{for(const s of n)if(t<=s.end+1&&e>=s.start){s.start=h.min(s.start,t),s.end=h.max(s.end,e);return}n.push({start:t,end:e}),n.sort((s,i)=>s.start-i.start)})}#w(t){this.#a.mutate(e=>{for(;e.length>0;){if(e[0].end>=t){e[0].start=h.max(e[0].start,t);break}e.shift()}})}close(){this.#u.close()}};async function kr(t){const e=t.codec==="opus"?void 0:t.description?V(t.description):void 0;return(await AudioDecoder.isConfigSupported({...t,description:e})).supported??!1}const Et=.001,De=.2;class Ar{source;volume;muted;paused;#e=new S;#t=.5;#n=new l(void 0);constructor(e,n){this.source=e,this.volume=l.from(n?.volume??.5),this.muted=l.from(n?.muted??!1),this.paused=l.from(n?.paused??n?.muted??!1),this.#e.run(s=>{s.get(this.muted)?(this.#t=this.volume.peek()||.5,this.volume.set(0)):this.volume.set(this.#t)}),this.#e.run(s=>{const i=!s.get(this.paused)&&!s.get(this.muted);this.source.enabled.set(i)}),this.#e.run(s=>{const i=s.get(this.volume);this.muted.set(i===0)}),this.#e.run(s=>{const i=s.get(this.source.root);if(!i)return;const r=new GainNode(i.context,{gain:s.get(this.volume)});i.connect(r),s.set(this.#n,r),s.run(a=>{a.get(this.source.enabled)&&(r.connect(i.context.destination),a.cleanup(()=>r.disconnect()))})}),this.#e.run(s=>{const i=s.get(this.#n);if(!i)return;s.cleanup(()=>i.gain.cancelScheduledValues(i.context.currentTime));const r=s.get(this.volume);r<Et?(i.gain.exponentialRampToValueAtTime(Et,i.context.currentTime+De),i.gain.setValueAtTime(0,i.context.currentTime+De+.01)):i.gain.exponentialRampToValueAtTime(r,i.context.currentTime+De)})}close(){this.#e.close()}}class xr{element;paused;#e;#t=new l(void 0);mediaSource=this.#t;#n=new S;constructor(e,n){this.element=l.from(n?.element),this.paused=l.from(n?.paused??!1),this.#e=e,this.#n.run(this.#i.bind(this)),this.#n.run(this.#s.bind(this)),this.#n.run(this.#r.bind(this)),this.#n.run(this.#a.bind(this)),this.#n.run(this.#o.bind(this))}#i(e){const n=e.get(this.element);if(!n)return;const s=new MediaSource;n.src=URL.createObjectURL(s),e.cleanup(()=>URL.revokeObjectURL(n.src)),e.event(s,"sourceopen",()=>{e.set(this.#t,s)},{once:!0}),e.event(s,"error",i=>{console.error("[MSE] MediaSource error event:",i)})}#s(e){const n=e.get(this.element);if(!n||e.get(this.paused))return;const s=h.toSecond(e.get(this.#e.buffer));e.interval(()=>{const i=n.buffered;if(i.length===0)return;const r=i.end(i.length-1)-s,a=r-n.currentTime;(a>.1||a<-.1)&&(console.warn("seeking",a>0?"forward":"backward",Math.abs(a).toFixed(3),"seconds"),n.currentTime=r)},100)}#r(e){const n=e.get(this.element);if(!n)return;const s=e.get(this.mediaSource);s&&e.interval(async()=>{for(const i of s.sourceBuffers){for(;i.updating;)await new Promise(r=>i.addEventListener("updateend",r,{once:!0}));n.currentTime>10&&i.remove(0,n.currentTime-10)}},1e3)}#a(e){const n=e.get(this.element);if(!n)return;const s=e.get(this.paused);s&&!n.paused?n.pause():!s&&n.paused&&n.play().catch(i=>{console.error("[MSE] MediaElement play error:",i)})}#o(e){const n=e.get(this.element);if(!n||e.get(this.paused))return;const s=e.get(this.#e.reference);if(s===void 0)return;const i=e.get(this.#e.buffer),r=h.sub(h.sub(h.now(),s),i);n.currentTime=h.toSecond(r)}close(){this.#n.close()}}const _r=20,Ut=100;class st{#e=new l(void 0);reference=this.#e;latency;jitter;audio;video;#t=new l(h.zero);buffer=this.#t;#n;timestamp=new l(void 0);#i=new Map;#s;#r;signals=new S;constructor(e){this.latency=l.from(e?.latency??"real-time"),this.jitter=new l(Ut),this.#s=e?.connection,this.audio=l.from(e?.audio),this.video=l.from(e?.video),this.#n=Promise.withResolvers(),this.signals.run(this.#a.bind(this)),this.signals.run(this.#o.bind(this))}#a(e){const n=e.get(this.latency);if(typeof n=="number"){this.#r=void 0,this.jitter.set(n);return}const s=(this.#s?e.get(this.#s):void 0)?.rtt,i=s?e.get(s):void 0;if(i!==void 0){this.#r=this.#r!==void 0?Math.min(this.#r,i):i;const r=Math.max(_r,this.#r*1.25);this.jitter.set(r);return}this.#r=void 0,this.jitter.set(Ut)}#o(e){const n=e.get(this.jitter),s=e.get(this.video)??h.zero,i=e.get(this.audio)??h.zero,r=h.add(h.max(s,i),n);this.#t.set(r),this.#n.resolve(),this.#n=Promise.withResolvers()}received(e,n=""){this.timestamp.update(a=>a===void 0||e>a?e:a);const s=h.now(),i=h.sub(s,e),r=this.#e.peek();if(r!==void 0){const a=h.add(h.sub(r,i),this.#t.peek());if(a<0){const o=this.#i.get(n);o?(o.count++,o.maxMs=Math.max(o.maxMs,-a)):this.#i.set(n,{count:1,maxMs:-a})}else{const o=this.#i.get(n);if(o){const c=n?`sync[${n}]`:"sync",u=st.#c(o.maxMs);console.debug(`${c}: ${o.count} late frame(s), max ${u} behind`),this.#i.delete(n)}}if(i>=r)return}this.#e.set(i),this.#n.resolve(),this.#n=Promise.withResolvers()}now(){const e=this.#e.peek();if(e!==void 0)return h.sub(h.sub(h.now(),e),this.#t.peek())}async wait(e){if(this.#e.peek()===void 0)throw new Error("reference not set; call update() first");for(;;){const n=h.now(),s=h.sub(n,e),i=this.#e.peek();if(i===void 0)return;const r=h.add(h.sub(i,s),this.#t.peek());if(r<=0||r<5)return;const a=new Promise(o=>setTimeout(o,r)).then(()=>!0);if(await Promise.race([this.#n.promise,a]))return}}static#c(e){if(e=Math.round(e),e<1e3)return`${e}ms`;const n=e/1e3;if(n<60)return`${Math.round(n*10)/10}s`;const s=n/60;return`${Math.round(s*10)/10}m`}close(){this.signals.close()}}const Sr=500,Er=100;class Ur{enabled;source;#e=new l(void 0);#t=new l(void 0);frame=this.#t;#n=new l(void 0);timestamp=this.#n;#i=new l(void 0);display=this.#i;#s=new l(!1);stalled=this.#s;#r=new l(void 0);stats=this.#r;#a=new l([]);buffered=this.#a;#o=new S;constructor(e,n){this.enabled=l.from(n?.enabled??!1),this.source=e,this.source.supported.set(zr),this.#o.run(this.#c.bind(this)),this.#o.run(this.#u.bind(this)),this.#o.run(this.#l.bind(this)),this.#o.run(this.#d.bind(this))}#c(e){const n=e.getAll([this.enabled,this.source.broadcast,this.source.track,this.source.config]);if(!n){this.#e.set(void 0);return}const[s,i,r,a]=n,o=e.get(i.active);if(!o)return;let c=new Mr({source:this.source,broadcast:o,track:r,config:a,stats:this.#r});e.cleanup(()=>c?.close()),e.run(u=>{if(!c)return;const d=u.get(this.#e);if(d){const f=u.get(c.timestamp),m=u.get(d.timestamp);if(!f||m&&m>f+Er)return}this.#e.set(c),c=void 0,u.close()})}#u(e){const n=e.get(this.#e);if(!n){this.#a.set([]);return}e.cleanup(()=>n.close()),e.run(s=>{const i=s.get(n.frame);this.#t.update(r=>(r?.close(),i?.clone()))}),e.proxy(this.#n,n.timestamp),e.proxy(this.#a,n.buffered)}#l(e){const n=e.get(this.source.catalog);if(!n)return;const s=n.display;if(s){e.set(this.#i,{width:s.width,height:s.height});return}const i=e.get(this.frame);i&&e.set(this.#i,{width:i.displayWidth,height:i.displayHeight})}#d(e){if(e.get(this.enabled)){if(!e.get(this.frame)){this.#s.set(!0);return}this.#s.set(!1),e.timer(()=>{this.#s.set(!0)},Sr)}}close(){this.#t.update(e=>{e?.close()}),this.#o.close()}}class Mr{source;broadcast;track;config;stats;timestamp=new l(void 0);frame=new l(void 0);buffered=new l([]);#e=new l([]);signals=new S;constructor(e){const{codedWidth:n,codedHeight:s,...i}=e.config;this.source=e.source,this.broadcast=e.broadcast,this.track=e.track,this.config=i,this.stats=e.stats,this.signals.run(this.#t.bind(this))}#t(e){const n=this.broadcast.subscribe(this.track,te.video);e.cleanup(()=>n.close());const s=new VideoDecoder({output:async i=>{try{const r=h.fromMicro(i.timestamp);if(r<(this.timestamp.peek()??0))return;this.frame.peek()===void 0&&this.frame.set(i.clone());const a=this.source.sync.wait(r).then(()=>!0);if(!await Promise.race([a,e.cancel])||r<(this.timestamp.peek()??0))return;this.timestamp.set(r),this.#r(r),this.frame.update(o=>(o?.close(),i.clone()))}finally{i.close()}},error:i=>{console.error(i),e.close()}});e.cleanup(()=>{s.state!=="closed"&&s.close()}),this.config.container.kind==="cmaf"?this.#i(e,n,s):this.#n(e,n,s)}#n(e,n,s){const i=new se(n,{format:new Le,latency:this.source.sync.buffer});e.cleanup(()=>i.close()),e.run(a=>{const o=a.get(i.buffered),c=a.get(this.#e);this.buffered.update(()=>Se(o,c))}),s.configure({...this.config,description:this.config.description?V(this.config.description):void 0,optimizeForLatency:this.config.optimizeForLatency??!0,flip:!1});let r;e.spawn(async()=>{for(;;){const a=await Promise.race([i.next(),e.cancel]);if(!a)break;const{frame:o,group:c}=a;if(!o){r&&(r.final=!0);continue}const u=h.fromMicro(o.timestamp);this.source.sync.received(u,"video");const d=new EncodedVideoChunk({type:o.keyframe?"key":"delta",data:o.data,timestamp:o.timestamp});if(this.stats.update(f=>({frameCount:(f?.frameCount??0)+1,bytesReceived:(f?.bytesReceived??0)+o.data.byteLength})),r?.group===c||r?.final&&r.group+1===c){const f=h.fromMicro(r.timestamp),m=h.fromMicro(o.timestamp);this.#s(f,m)}r={timestamp:o.timestamp,group:c,final:!1},s.decode(d)}})}#i(e,n,s){if(this.config.container.kind!=="cmaf")return;const i=be(this.config.container.init),r=Ie(i),a=this.config.description?V(this.config.description):void 0,o=new se(n,{format:new An(r),latency:this.source.sync.buffer});e.cleanup(()=>o.close()),e.run(u=>{const d=u.get(o.buffered),f=u.get(this.#e);this.buffered.update(()=>Se(d,f))}),s.configure({codec:this.config.codec,description:a,optimizeForLatency:this.config.optimizeForLatency??!0,flip:!1});let c;e.spawn(async()=>{for(;;){const u=await Promise.race([o.next(),e.cancel]);if(!u)break;const{frame:d,group:f}=u;if(!d){c&&(c.final=!0);continue}const m=h.fromMicro(d.timestamp);if(this.source.sync.received(m,"video"),this.stats.update(b=>({frameCount:(b?.frameCount??0)+1,bytesReceived:(b?.bytesReceived??0)+d.data.byteLength})),c?.group===f||c?.final&&c.group+1===f){const b=h.fromMicro(c.timestamp),w=h.fromMicro(d.timestamp);this.#s(b,w)}if(c={timestamp:d.timestamp,group:f,final:!1},s.state==="closed")break;s.decode(new EncodedVideoChunk({type:d.keyframe?"key":"delta",data:d.data,timestamp:d.timestamp}))}})}#s(e,n){e>n||this.#e.mutate(s=>{for(const i of s)if(i.start<=n&&i.end>=e){i.start=h.min(i.start,e),i.end=h.max(i.end,n);return}s.push({start:e,end:n}),s.sort((i,r)=>i.start-r.start)})}#r(e){this.#e.mutate(n=>{for(;n.length>0;){if(n[0].end>=e){n[0].start=h.max(n[0].start,e);break}n.shift()}})}close(){this.signals.close(),this.frame.update(e=>{e?.close()})}}async function zr(t){const e=t.description?V(t.description):void 0,{supported:n}=await VideoDecoder.isConfigSupported({codec:t.codec,description:e,optimizeForLatency:t.optimizeForLatency??!0});if(n)return!0;if(t.codec.startsWith("avc3.")){const s=`avc1.${t.codec.slice(5)}`;if((await VideoDecoder.isConfigSupported({codec:s,description:e,optimizeForLatency:t.optimizeForLatency??!0})).supported)return t.codec=s,!0}return!1}let Tr=class{muxer;source;#e=new l(void 0);stats=this.#e;#t=new l([]);buffered=this.#t;#n=new l(!1);stalled=this.#n;#i=new l(h.zero);timestamp=this.#i;signals=new S;constructor(t,e){this.muxer=t,this.source=e,this.source.supported.set(Cr),this.signals.run(this.#s.bind(this)),this.signals.run(this.#c.bind(this)),this.signals.run(this.#u.bind(this))}#s(t){const e=t.get(this.muxer.element);if(!e)return;const n=t.get(this.muxer.mediaSource);if(!n)return;const s=t.get(this.source.broadcast);if(!s)return;const i=t.get(s.active);if(!i)return;const r=t.get(this.source.track);if(!r)return;const a=t.get(this.source.config);if(!a)return;const o=`video/mp4; codecs="${a.codec}"`,c=n.addSourceBuffer(o);t.cleanup(()=>{n.removeSourceBuffer(c),c.abort()}),t.event(c,"error",u=>{console.error("[MSE] SourceBuffer error:",u)}),t.event(c,"updateend",()=>{this.#t.set(_n(c.buffered))}),a.container.kind==="cmaf"?this.#a(t,i,r,a,c,e):this.#o(t,i,r,a,c,e)}async#r(t,e){for(;t.updating;)await new Promise(n=>t.addEventListener("updateend",n,{once:!0}));for(t.appendBuffer(e);t.updating;)await new Promise(n=>t.addEventListener("updateend",n,{once:!0}))}#a(t,e,n,s,i,r){if(s.container.kind!=="cmaf")throw new Error("unreachable");const a=e.subscribe(n,te.video);t.cleanup(()=>a.close());const o=be(s.container.init),c=Ie(o);t.spawn(async()=>{for(await this.#r(i,o);;){const u=await a.readFrame();if(!u)return;const d=bn(u,c);this.source.sync.received(h.fromMicro(d),"video"),await this.#r(i,u),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0))}})}#o(t,e,n,s,i,r){const a=e.subscribe(n,te.video);t.cleanup(()=>a.close());const o=new se(a,{format:new Le,latency:this.source.sync.buffer});t.cleanup(()=>o.close()),t.spawn(async()=>{const c=sr(s);await this.#r(i,c);let u=1,d,f;for(;;){const m=await o.next();if(!m)return;if(!m.frame)continue;f=m.frame;const b=h.fromMicro(f.timestamp);this.source.sync.received(b,"video");break}for(;;){const m=await o.next();if(m&&!m.frame)continue;const b=m?.frame;if(b){d=Me.sub(b.timestamp,f.timestamp);const k=h.fromMicro(b.timestamp);this.source.sync.received(k,"video")}const w=kn({data:f.data,timestamp:f.timestamp,duration:d??0,keyframe:f.keyframe,sequence:u++});if(await this.#r(i,w),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0)),!b)return;f=b}})}#c(t){const e=t.get(this.muxer.element);if(!e)return;const n=()=>{this.#n.set(e.readyState<=HTMLMediaElement.HAVE_CURRENT_DATA)};n(),t.event(e,"waiting",n),t.event(e,"playing",n),t.event(e,"seeking",n)}#u(t){const e=t.get(this.muxer.element);if(e)if("requestVideoFrameCallback"in e){const n=e;let s;const i=()=>{const r=h.fromSecond(n.currentTime);this.#i.set(r),s=n.requestVideoFrameCallback(i)};s=n.requestVideoFrameCallback(i),t.cleanup(()=>n.cancelVideoFrameCallback(s))}else t.event(e,"timeupdate",()=>{const n=h.fromSecond(e.currentTime);this.#i.set(n)})}close(){this.source.close(),this.signals.close()}};async function Cr(t){return MediaSource.isTypeSupported(`video/mp4; codecs="${t.codec}"`)}class Ir{decoder;canvas;paused;frame=new l(void 0);timestamp=new l(void 0);#e=new l(void 0);#t=new l(!1);#n=new S;constructor(e,n){this.decoder=e,this.canvas=l.from(n?.canvas),this.paused=l.from(n?.paused??!1),this.#n.run(s=>{const i=s.get(this.canvas);this.#e.set(i?.getContext("2d")??void 0)}),this.#n.run(this.#s.bind(this)),this.#n.run(this.#r.bind(this)),this.#n.run(this.#a.bind(this)),this.#n.run(this.#i.bind(this))}#i(e){const n=e.getAll([this.canvas,this.decoder.display]);if(!n)return;const[s,i]=n;(s.width!==i.width||s.height!==i.height)&&(s.width=i.width,s.height=i.height)}#s(e){const n=e.get(this.canvas);if(!n){this.#t.set(!1);return}let s=!1;const i=()=>{this.#t.set(s&&!document.hidden)},r=new IntersectionObserver(a=>{for(const o of a)s=o.isIntersecting,i()},{threshold:.01});e.event(document,"visibilitychange",i),r.observe(n),e.cleanup(()=>r.disconnect()),e.cleanup(()=>this.#t.set(!1))}#r(e){const n=e.get(this.paused),s=e.get(this.#t);if(e.cleanup(()=>this.decoder.enabled.set(!1)),!n){this.decoder.enabled.set(s);return}const i=e.get(this.frame);this.decoder.enabled.set(!i)}#a(e){const n=e.get(this.#e);if(!n)return;const s=e.get(this.paused);let i;s||(i=e.get(this.decoder.frame));let r=requestAnimationFrame(()=>{const a=i??this.frame.peek();this.#o(n,a),i&&(this.frame.update(o=>(o?.close(),i.clone())),this.timestamp.set(h.fromMicro(i.timestamp))),r=void 0});e.cleanup(()=>{i?.close(),r&&cancelAnimationFrame(r)})}#o(e,n){if(!n){e.fillStyle="#000",e.fillRect(0,0,e.canvas.width,e.canvas.height);return}e.save(),e.fillStyle="#000",e.fillRect(0,0,e.canvas.width,e.canvas.height),this.decoder.source.catalog.peek()?.flip&&(e.scale(-1,1),e.translate(-e.canvas.width,0)),e.drawImage(n,0,0,e.canvas.width,e.canvas.height),e.restore()}close(){this.frame.update(e=>{e?.close()}),this.#n.close()}}function Lr(t){return e=>{const n=[],s=[];for(const[i,r]of e)if(r.codedWidth&&r.codedHeight){const a=r.codedWidth*r.codedHeight;a<=t?n.push({name:i,size:a}):s.push({name:i,size:a})}return n.sort((i,r)=>r.size-i.size),n.length>0?n.map(i=>i.name):s.length>0?(s.sort((i,r)=>i.size-r.size),[s[0].name]):e.map(([i])=>i)}}function Br(t){return e=>{const n=[],s=[];for(const[i,r]of e)r.bitrate!=null&&r.bitrate<=t?n.push({name:i,bitrate:r.bitrate}):r.bitrate!=null&&s.push({name:i,bitrate:r.bitrate});return n.sort((i,r)=>r.bitrate-i.bitrate),n.length>0?n.map(i=>i.name):s.length>0?(s.sort((i,r)=>i.bitrate-r.bitrate),[s[0].name]):e.map(([i])=>i)}}function Rr(t){let e=t[0];for(const n of t){const[,s]=n,[,i]=e,r=(s.codedWidth??0)*(s.codedHeight??0),a=(i.codedWidth??0)*(i.codedHeight??0);if(r!==a){r>a&&(e=n);continue}(s.bitrate??0)>(i.bitrate??0)&&(e=n)}return e[0]}let Or=class{broadcast;target;#e=new l(void 0);catalog=this.#e;#t=new l({});available=this.#t;#n=new l(void 0);track=this.#n;#i=new l(void 0);config=this.#i;sync;supported;#s=new S;constructor(t,e){this.broadcast=l.from(e?.broadcast),this.target=l.from(e?.target),this.sync=t,this.supported=l.from(e?.supported),this.#s.run(this.#r.bind(this)),this.#s.run(this.#a.bind(this)),this.#s.run(this.#o.bind(this))}#r(t){const e=t.get(this.broadcast);if(!e)return;const n=t.get(e.catalog)?.video;n&&t.set(this.#e,n)}#a(t){const e=t.get(this.supported);if(!e)return;const n=t.get(this.#e)?.renditions??{};t.spawn(async()=>{const s={};for(const[i,r]of Object.entries(n))await e(r)&&(s[i]=r);Object.keys(s).length===0&&Object.keys(n).length>0&&console.warn("[Source] No supported video renditions found:",n),this.#t.set(s)})}#o(t){const e=t.get(this.#t);if(Object.keys(e).length===0)return;const n=t.get(this.target);if(n?.name&&n.name in e){const o=e[n.name];t.set(this.#n,n.name),t.set(this.#i,o),t.set(this.sync.video,o.jitter);return}let s=n;if(!n?.bitrate){const o=t.get(this.broadcast),c=(o?t.get(o.connection):void 0)?.recvBandwidth;if(c){const u=t.get(c);if(u!=null){const d=Math.round(u*.8);s={...n,bitrate:d}}}}const i=this.#c(e,s);if(!i)return;const r=e[i];t.set(this.#n,i),t.set(this.#i,r);const a=r.jitter??(r.framerate?Math.ceil(1e3/r.framerate):void 0);t.set(this.sync.video,a)}#c(t,e){const n=Object.entries(t);if(n.length===0)return;if(n.length===1)return n[0][0];const s=[];if(e?.pixels!=null&&s.push(Lr(e.pixels)),e?.bitrate!=null&&s.push(Br(e.bitrate)),s.length===0)return Rr(n);const i=s.map(a=>a(n)),r=i.map(a=>new Set(a));for(const a of i[0])if(r.every(o=>o.has(a)))return a;console.warn("conflicting rendition filters, no rendition satisfies all criteria")}close(){this.#s.close()}};function _n(t){const e=[];for(let n=0;n<t.length;n++){const s=h.fromSecond(t.start(n)),i=h.fromSecond(t.end(n));e.push({start:s,end:i})}return e}class Dr{source;stats=new l(void 0);stalled=new l(!1);buffered=new l([]);timestamp=new l(h.zero);constructor(e){this.source=e}}class Fr{source;volume=new l(.5);muted=new l(!1);stats=new l(void 0);buffered=new l([]);context=new l(void 0);constructor(e){this.source=e}}class $r{element=new l(void 0);broadcast;latency;jitter;paused;video;#e;audio;#t;sync;signals=new S;constructor(e){this.element=l.from(e?.element),this.broadcast=l.from(e?.broadcast),this.sync=new st({latency:e?.latency,connection:e?.connection}),this.latency=this.sync.latency,this.jitter=this.sync.jitter,this.#e=new Or(this.sync,{broadcast:this.broadcast}),this.#t=new Vr(this.sync,{broadcast:this.broadcast}),this.video=new Dr(this.#e),this.audio=new Fr(this.#t),this.paused=l.from(e?.paused??!1),this.signals.run(this.#n.bind(this))}#n(e){const n=e.get(this.element);n&&(n instanceof HTMLCanvasElement?this.#i(e,n):n instanceof HTMLVideoElement&&this.#s(e,n))}#i(e,n){const s=new Ur(this.#e),i=new vr(this.#t),r=new Ar(i,{volume:this.audio.volume,muted:this.audio.muted,paused:this.paused}),a=new Ir(s,{canvas:n,paused:this.paused});e.cleanup(()=>{s.close(),i.close(),r.close(),a.close()}),e.proxy(this.video.stats,s.stats),e.proxy(this.video.buffered,s.buffered),e.proxy(this.video.stalled,s.stalled),e.proxy(this.video.timestamp,s.timestamp),e.proxy(this.audio.stats,i.stats),e.proxy(this.audio.buffered,i.buffered),e.proxy(this.audio.context,i.context)}#s(e,n){const s=new xr(this.sync,{paused:this.paused,element:n}),i=new Tr(s,this.#e),r=new Nr(s,this.#t,{volume:this.audio.volume,muted:this.audio.muted});e.cleanup(()=>{i.close(),r.close(),s.close()}),e.proxy(this.video.stats,i.stats),e.proxy(this.video.buffered,i.buffered),e.proxy(this.video.stalled,i.stalled),e.proxy(this.video.timestamp,i.timestamp),e.proxy(this.audio.stats,r.stats),e.proxy(this.audio.buffered,r.buffered),e.proxy(this.audio.context,r.context)}close(){this.signals.close()}}class Nr{muxer;source;volume;muted;#e=new l(void 0);stats=this.#e;#t=new l([]);buffered=this.#t;context=new l(void 0);#n=new S;constructor(e,n,s){this.muxer=e,this.source=n,this.source.supported.set(jr),this.volume=l.from(s?.volume??.5),this.muted=l.from(s?.muted??!1),this.#n.run(this.#i.bind(this)),this.#n.run(this.#o.bind(this))}#i(e){const n=e.get(this.muxer.element);if(!n)return;const s=e.get(this.muxer.mediaSource);if(!s)return;const i=e.get(this.source.broadcast);if(!i)return;const r=e.get(i.active);if(!r)return;const a=e.get(this.source.track);if(!a)return;const o=e.get(this.source.config);if(!o)return;const c=`audio/mp4; codecs="${o.codec}"`,u=s.addSourceBuffer(c);e.cleanup(()=>{s.removeSourceBuffer(u),u.abort()}),e.event(u,"error",f=>{console.error("[MSE] SourceBuffer error:",f)}),e.event(u,"updateend",()=>{this.#t.set(_n(u.buffered))});const d=r.subscribe(a,te.audio);e.cleanup(()=>d.close()),o.container.kind==="cmaf"?this.#r(e,d,o,u,n):this.#a(e,d,o,u,n)}async#s(e,n){for(;e.updating;)await new Promise(s=>e.addEventListener("updateend",s,{once:!0}));for(e.appendBuffer(n);e.updating;)await new Promise(s=>e.addEventListener("updateend",s,{once:!0}))}#r(e,n,s,i,r){if(s.container.kind!=="cmaf")throw new Error("unreachable");const a=be(s.container.init),o=Ie(a);e.spawn(async()=>{for(await this.#s(i,a);;){const c=await n.readFrame();if(!c)return;const u=bn(c,o);this.source.sync.received(h.fromMicro(u),"audio"),await this.#s(i,c),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0))}})}#a(e,n,s,i,r){const a=new se(n,{format:new Le,latency:this.source.sync.buffer});e.cleanup(()=>a.close()),e.spawn(async()=>{const o=ir(s);await this.#s(i,o);let c=1,u,d;for(;;){const f=await a.next();if(!f)return;if(!f.frame)continue;d=f.frame;const m=h.fromMicro(d.timestamp);this.source.sync.received(m,"audio");break}for(;;){const f=await a.next();if(f&&!f.frame)continue;const m=f?.frame;if(m){u=Me.sub(m.timestamp,d.timestamp);const w=h.fromMicro(m.timestamp);this.source.sync.received(w,"audio")}const b=kn({data:d.data,timestamp:d.timestamp,duration:u??0,keyframe:d.keyframe,sequence:c++});if(await this.#s(i,b),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0)),!m)return;d=m}})}#o(e){const n=e.get(this.muxer.element);if(!n)return;const s=e.get(this.volume),i=e.get(this.muted);i&&!n.muted?n.muted=!0:!i&&n.muted&&(n.muted=!1),s!==n.volume&&(n.volume=s),e.event(n,"volumechange",()=>{this.volume.set(n.volume)})}close(){this.#n.close()}}async function jr(t){return MediaSource.isTypeSupported(`audio/mp4; codecs="${t.codec}"`)}const Pr=128;class Vr{broadcast;target;#e=new l(void 0);catalog=this.#e;#t=new l({});available=this.#t;#n=new l(void 0);track=this.#n;#i=new l(void 0);config=this.#i;supported;sync;#s=new S;constructor(e,n){this.sync=e,this.broadcast=l.from(n?.broadcast),this.target=l.from(n?.target),this.supported=l.from(n?.supported),this.#s.run(this.#r.bind(this)),this.#s.run(this.#a.bind(this)),this.#s.run(this.#o.bind(this))}#r(e){const n=e.get(this.broadcast);if(!n)return;const s=e.get(n.catalog)?.audio;s&&e.set(this.#e,s)}#a(e){const n=e.get(this.#e)?.renditions??{},s=e.get(this.supported);s&&e.spawn(async()=>{const i={};for(const[r,a]of Object.entries(n))await s(a)&&(i[r]=a);Object.keys(i).length===0&&Object.keys(n).length>0&&console.warn("no supported audio renditions found:",n),this.#t.set(i)})}#o(e){const n=e.get(this.#t);if(Object.keys(n).length===0)return;const s=e.get(this.target);let i;if(s?.name&&s.name in n)i={track:s.name,config:n[s.name]};else if(i=this.#c(n),!i)return;e.set(this.#n,i.track),e.set(this.#i,i.config);const r=i.config.jitter??Wr(i.config)??0,a=Math.ceil(Pr/i.config.sampleRate*1e3),o=r+a;e.set(this.sync.audio,o)}#c(e){const n=Object.entries(e);if(n.length!==0){for(const[s,i]of n)if(i.container.kind==="legacy")return{track:s,config:i};for(const[s,i]of n)if(i.container.kind==="cmaf")return{track:s,config:i}}}close(){this.#s.close()}}function Wr(t){if(t.codec.startsWith("opus"))return 20;if(t.codec.startsWith("mp4a"))return Math.ceil(1024/t.sampleRate*1e3)}const Zr=Ce([an(["loc","cmaf","legacy","mediatimeline","eventtimeline"]),v()]),Hr=Ce([an(["video","audio","audiodescription","caption","subtitle","signlanguage"]),v()]),qr=A({name:v(),packaging:Zr,isLive:Ye(),role:p(Hr),codec:p(v()),width:p(M()),height:p(M()),framerate:p(M()),samplerate:p(M()),channelConfig:p(v()),bitrate:p(M()),initData:p(v()),renderGroup:p(M()),altGroup:p(M()),jitter:p(M())}),Gr=A({version:Ve(1),tracks:q(qr)});function Xr(t){const e=new TextDecoder().decode(t);try{const n=JSON.parse(e);return Gr.parse(n)}catch(n){throw console.warn("invalid MSF catalog",e),n}}async function Yr(t){const e=await t.readFrame();if(e)return Xr(e)}const Kr=48e3,Mt=2;function Jr(t){let e="";for(let n=0;n<t.length;n++)e+=t[n].toString(16).padStart(2,"0");return e}function Sn(t){let e;try{e=t.initData?be(t.initData):void 0}catch{e=void 0}return t.packaging==="cmaf"&&t.initData&&e?{container:{kind:"cmaf",init:t.initData},description:void 0}:{container:{kind:"legacy"},description:e?Jr(e):void 0}}function Qr(t){if(!t.codec)return;const{container:e,description:n}=Sn(t);return{codec:t.codec,container:e,description:n,codedWidth:t.width!=null?Z(t.width):void 0,codedHeight:t.height!=null?Z(t.height):void 0,framerate:t.framerate,bitrate:t.bitrate!=null?Z(t.bitrate):void 0,jitter:t.jitter!=null?Z(t.jitter):void 0}}function eo(t){if(!t.codec)return;const e=(()=>{if(!t.channelConfig)return Mt;const i=Number.parseInt(t.channelConfig,10);return Number.isFinite(i)?i:Mt})(),{container:n,description:s}=Sn(t);return{codec:t.codec,container:n,description:s,sampleRate:Z(t.samplerate??Kr),numberOfChannels:Z(e),bitrate:t.bitrate!=null?Z(t.bitrate):void 0,jitter:t.jitter!=null?Z(t.jitter):void 0}}function to(t){const e={},n={};for(const i of t.tracks)if(i.role==="video"){const r=Qr(i);r&&(e[i.name]=r)}else if(i.role==="audio"){const r=eo(i);r&&(n[i.name]=r)}const s={};return Object.keys(e).length>0&&(s.video={renditions:e}),Object.keys(n).length>0&&(s.audio={renditions:n}),s}const no=["hang","msf","manual"];class so{connection;enabled;name;status=new l("offline");reload;catalogFormat;#e=new l(void 0);active=this.#e;catalog;#t;#n=new l(!1);signals=new S;constructor(e){this.connection=l.from(e?.connection),this.name=l.from(e?.name??In()),this.enabled=l.from(e?.enabled??!1),this.reload=l.from(e?.reload??!1),this.catalogFormat=l.from(e?.catalogFormat??"hang"),this.catalog=l.from(e?.catalog),this.#t=e?.announced??new l(new Set),this.signals.run(this.#i.bind(this)),this.signals.run(this.#s.bind(this)),this.signals.run(this.#r.bind(this))}#i(e){if(!e.get(this.reload)){this.#n.set(!0);return}if(e.get(this.connection)?.url.hostname.endsWith("mediaoverquic.com")){console.warn("Cloudflare relay does not support broadcast discovery yet; ignoring reload signal."),this.#n.set(!0);return}const n=e.get(this.name),s=e.get(this.#t);this.#n.set(s.has(n))}#s(e){if(!e.get(this.enabled)||!e.get(this.#n))return;const n=e.get(this.connection);if(!n)return;const s=e.get(this.name),i=n.consume(s);e.cleanup(()=>i.close()),e.set(this.#e,i,void 0)}#r(e){if(!e.get(this.enabled))return;const n=e.get(this.catalogFormat);if(n==="manual"){const o=e.get(this.catalog);this.status.set(o?"live":"loading");return}const s=e.get(this.active);if(!s)return;this.status.set("loading");const i=n==="hang"?"catalog.json":"catalog",r=s.subscribe(i,te.catalog);e.cleanup(()=>r.close());const a=n==="hang"?async()=>ai(r):async()=>{const o=await Yr(r);return o?to(o):void 0};e.spawn(async()=>{try{for(;;){const o=await Promise.race([e.cancel,a()]);if(!o)break;console.debug("received catalog",n,this.name.peek(),o),this.catalog.set(o),this.status.set("live")}}catch(o){console.warn("error fetching catalog",this.name.peek(),o)}finally{this.catalog.set(void 0),this.status.set("offline")}})}close(){this.signals.close()}}const io="hang";function ro(t){return no.find(e=>e===t)??io}const oo=["url","name","paused","volume","muted","reload","latency","jitter","catalog-format"],ao=new FinalizationRegistry(t=>t.close());class co extends HTMLElement{static observedAttributes=oo;connection;broadcast;backend;#e=new l(!1);signals=new S;constructor(){super(),ao.register(this,this.signals),this.connection=new Bn({enabled:this.#e}),this.signals.cleanup(()=>this.connection.close()),this.broadcast=new so({connection:this.connection.established,announced:this.connection.announced,enabled:this.#e}),this.signals.cleanup(()=>this.broadcast.close()),this.backend=new $r({broadcast:this.broadcast,connection:this.connection.established}),this.signals.cleanup(()=>this.backend.close());const e=()=>{const s=this.querySelector("canvas"),i=this.querySelector("video");if(s&&i)throw new Error("Cannot have both canvas and video elements");this.backend.element.set(s??i)},n=new MutationObserver(e);n.observe(this,{childList:!0,subtree:!0}),this.signals.cleanup(()=>n.disconnect()),e(),this.signals.run(s=>{const i=s.get(this.connection.url);i?this.setAttribute("url",i.toString()):this.removeAttribute("url")}),this.signals.run(s=>{const i=s.get(this.broadcast.name);this.setAttribute("name",i.toString())}),this.signals.run(s=>{s.get(this.backend.audio.muted)?this.setAttribute("muted",""):this.removeAttribute("muted")}),this.signals.run(s=>{s.get(this.backend.paused)?this.setAttribute("paused","true"):this.removeAttribute("paused")}),this.signals.run(s=>{const i=s.get(this.backend.audio.volume);this.setAttribute("volume",i.toString())}),this.signals.run(s=>{if(s.get(this.backend.latency)==="real-time")this.setAttribute("latency","real-time");else{const i=Math.floor(s.get(this.backend.jitter));this.setAttribute("latency",i.toString())}})}connectedCallback(){this.#e.set(!0),this.style.display="block",this.style.position="relative"}disconnectedCallback(){this.#e.set(!1)}#t(e){const n=e?Number.parseFloat(e):Number.NaN;this.backend.latency.set(Number.isFinite(n)?n:100)}attributeChangedCallback(e,n,s){if(n!==s)if(e==="url")this.connection.url.set(s?new URL(s):void 0);else if(e==="name")this.broadcast.name.set(ct(s??""));else if(e==="paused")this.backend.paused.set(s!==null);else if(e==="volume"){const i=s?Number.parseFloat(s):.5;this.backend.audio.volume.set(i)}else if(e==="muted")this.backend.audio.muted.set(s!==null);else if(e==="reload")this.broadcast.reload.set(s!==null);else if(e==="latency")!s||s==="real-time"?this.backend.latency.set("real-time"):this.#t(s);else if(e==="jitter")this.#t(s);else if(e==="catalog-format")this.broadcast.catalogFormat.set(ro(s));else{const i=e;throw new Error(`Invalid attribute: ${i}`)}}get url(){return this.connection.url.peek()}set url(e){this.connection.url.set(e?new URL(e):void 0)}get name(){return this.broadcast.name.peek()}set name(e){this.broadcast.name.set(ct(e))}get paused(){return this.backend.paused.peek()}set paused(e){this.backend.paused.set(e)}get volume(){return this.backend.audio.volume.peek()}set volume(e){this.backend.audio.volume.set(e)}get muted(){return this.backend.audio.muted.peek()}set muted(e){this.backend.audio.muted.set(e)}get reload(){return this.broadcast.reload.peek()}set reload(e){this.broadcast.reload.set(e)}get latency(){return this.backend.latency.peek()}set latency(e){this.backend.latency.set(e)}get jitter(){return this.backend.jitter.peek()}set jitter(e){this.backend.latency.set(e)}get catalogFormat(){return this.broadcast.catalogFormat.peek()}set catalogFormat(e){this.broadcast.catalogFormat.set(e)}get catalog(){return this.broadcast.catalog.peek()}set catalog(e){this.broadcast.catalog.set(e)}}customElements.define("moq-watch",co);const En=navigator.userAgent.toLowerCase().includes("firefox"),We={aac:"mp4a.40.2",opus:"opus",av1:"av01.0.08M.08",h264:"avc1.640028",h265:"hev1.1.6.L93.B0",vp9:"vp09.00.10.08",vp8:"vp8"};async function zt(t){return globalThis.AudioDecoder?(await AudioDecoder.isConfigSupported({codec:We[t],numberOfChannels:2,sampleRate:48e3})).supported===!0:!1}async function ce(t){const e=await VideoDecoder.isConfigSupported({codec:We[t],hardwareAcceleration:"prefer-software"}),n=await VideoDecoder.isConfigSupported({codec:We[t],hardwareAcceleration:"prefer-hardware"});return{hardware:En||n.config?.hardwareAcceleration!=="prefer-hardware"?void 0:n.supported===!0,software:e.supported===!0}}async function uo(){return{webtransport:typeof WebTransport<"u"?En?"partial":"full":"partial",audio:{decoding:{aac:await zt("aac"),opus:await zt("opus")?"full":"partial"},render:typeof AudioContext<"u"&&typeof AudioBufferSourceNode<"u"},video:{decoding:typeof VideoDecoder<"u"?{h264:await ce("h264"),h265:await ce("h265"),vp8:await ce("vp8"),vp9:await ce("vp9"),av1:await ce("av1")}:void 0,render:typeof OffscreenCanvas<"u"&&typeof CanvasRenderingContext2D<"u"}}}function D(t,e,...n){const s=document.createElement(t);if(!e)return s;const{style:i,classList:r,dataset:a,attributes:o,...c}=e;return i&&Object.assign(s.style,i),r&&s.classList.add(...r),a&&Object.entries(a).forEach(([u,d])=>{s.dataset[u]=d}),o&&Object.entries(o).forEach(([u,d])=>{s.setAttribute(u,d)}),n&&n.forEach(u=>{typeof u=="string"?s.appendChild(document.createTextNode(u)):s.appendChild(u)}),Object.assign(s,c),s}const Tt=navigator.userAgent.toLowerCase().includes("firefox"),lo=["show","details"];class fo extends HTMLElement{#e=new l("warning");#t=new l(!1);#n=new l(void 0);#i=new l(!1);#s;static observedAttributes=lo;constructor(){super(),uo().then(e=>this.#n.set(e)).catch(e=>console.error("Failed to detect watch support:",e))}connectedCallback(){this.#s=new S,this.#s.run(this.#a.bind(this))}disconnectedCallback(){this.#s?.close(),this.#s=void 0}attributeChangedCallback(e,n,s){if(e==="show"){const i=s??"warning";if(i==="always"||i==="warning"||i==="error"||i==="never")this.show=i;else throw new Error(`Invalid show: ${i}`)}else if(e==="details")this.details=s!==null;else{const i=e;throw new Error(`Invalid attribute: ${i}`)}}get show(){return this.#e.peek()}set show(e){this.#e.set(e)}get details(){return this.#t.peek()}set details(e){this.#t.set(e)}#r(e){return e.webtransport==="none"||!e.audio.decoding||!e.video.decoding||!e.audio.render||!e.video.render||!Object.values(e.audio.decoding).some(n=>n===!0||n==="full"||n==="partial")||!Object.values(e.video.decoding).some(n=>n.software||n.hardware)?"none":!Object.values(e.audio.decoding).every(n=>n===!0||n==="full")||!Object.values(e.video.decoding).every(n=>n.software||n.hardware)?"partial":"full"}#a(e){const n=e.get(this.#n);if(!n||e.get(this.#i))return;const s=e.get(this.#e);if(s==="never")return;const i=this.#r(n);if(s==="warning"&&i==="full"||s==="error"&&i!=="none")return;const r=D("div",{style:{margin:"0 auto",maxWidth:"28rem",padding:"1rem"}});this.appendChild(r),e.cleanup(()=>this.removeChild(r)),this.#o(r,i,e),e.get(this.#t)&&this.#c(r,n,e)}#o(e,n,s){const i=D("div",{style:{display:"flex",flexDirection:"row",gap:"1rem",flexWrap:"wrap",justifyContent:"space-between",alignItems:"center"}}),r=D("div",{style:{fontWeight:"bold"}});n==="full"?r.textContent="🟢 Full Browser Support":n==="partial"?r.textContent="🟡 Partial Browser Support":n==="none"&&(r.textContent="🔴 No Browser Support");const a=D("button",{type:"button",style:{fontSize:"14px"}});s.event(a,"click",()=>{this.#t.update(c=>!c)}),s.run(c=>{a.textContent=c.get(this.#t)?"Details ➖":"Details ➕"});const o=D("button",{type:"button",style:{fontSize:"14px"}},"Close ❌");s.event(o,"click",()=>{this.#i.set(!0)}),i.appendChild(r),i.appendChild(a),i.appendChild(o),e.appendChild(i),s.cleanup(()=>e.removeChild(i))}#c(e,n,s){const i=D("div",{style:{display:"grid",gridTemplateColumns:"1fr 1fr 1fr",columnGap:"0.5rem",rowGap:"0.2rem",backgroundColor:"rgba(0, 0, 0, 0.6)",borderRadius:"0.5rem",padding:"1rem",fontSize:"0.875rem"}}),r=u=>u?"🟢 Yes":"🔴 No",a=u=>u?.hardware?"🟢 Hardware":u?.software?`🟡 Software${Tt?"*":""}`:"🔴 No",o=u=>u==="full"?"🟢 Full":u==="partial"?"🟡 Polyfill":"🔴 None",c=(u,d,f)=>{const m=D("div",{style:{gridColumnStart:"1",fontWeight:"bold",textAlign:"right"}},u),b=D("div",{style:{gridColumnStart:"2",textAlign:"center"}},d),w=D("div",{style:{gridColumnStart:"3"}},f);i.appendChild(m),i.appendChild(b),i.appendChild(w)};if(c("WebTransport","",o(n.webtransport)),c("Rendering","Audio",r(n.audio.render)),c("","Video",r(n.video.render)),c("Decoding","Opus",o(n.audio.decoding.opus)),c("","AAC",r(n.audio.decoding.aac)),c("","AV1",a(n.video.decoding?.av1)),c("","H.265",a(n.video.decoding?.h265)),c("","H.264",a(n.video.decoding?.h264)),c("","VP9",a(n.video.decoding?.vp9)),c("","VP8",a(n.video.decoding?.vp8)),Tt){const u=D("div",{style:{gridColumnStart:"1",gridColumnEnd:"4",textAlign:"center",fontSize:"0.875rem",fontStyle:"italic"}},"Hardware acceleration is ",D("a",{href:"https://github.com/w3c/webcodecs/issues/896"},"undetectable")," on Firefox.");i.appendChild(u)}e.appendChild(i),s.cleanup(()=>e.removeChild(i))}}customElements.define("moq-watch-support",fo);const U=document.getElementById("player"),it=document.getElementById("errors"),Fe=document.getElementById("quality-panel"),Ct=document.getElementById("quality-buttons"),ho=document.getElementById("url-rows"),Un=document.getElementById("info-name"),It=document.getElementById("info-url"),Lt=document.getElementById("info-mode"),de=document.getElementById("info-status"),mo=document.getElementById("stream-control"),rt=document.getElementById("stream-input"),po=document.querySelectorAll("input[name=mode]"),me=document.getElementById("volume-slider"),Ze=document.getElementById("volume-icon"),Bt=document.getElementById("volume-label"),Ee=document.getElementById("jitter-slider"),He=document.getElementById("jitter-label"),ve=document.getElementById("dbg-fps"),Rt=document.getElementById("dbg-buf"),Ot=document.getElementById("dbg-stalled"),go=document.getElementById("dbg-latency"),bo=document.getElementById("dbg-jitter"),Dt=document.getElementById("dbg-vbytes"),wo=document.getElementById("dbg-codec");function G(t,e){return`<span class="badge badge-${e}">${t}</span>`}function yo(t){const e=document.createElement("div");e.className="banner banner-error",e.innerHTML=`⛔ ${t}`,it.appendChild(e)}const ee=new URLSearchParams(window.location.search),ge=ee.get("name")??ee.get("broadcast")??ee.get("id")??null,vo=ee.get("url")??ee.get("host")??null,ko=ee.get("fallbackPlayer")==="true";rt.value=ge??"";Un.textContent=ge??"—";const ot=typeof VideoDecoder<"u";if(!ot){const t=document.createElement("div");t.className="banner banner-warn",t.innerHTML="⚠️ <strong>WebCodecs is not supported in this browser.</strong> Falling back to MSE player. Chrome 94+, Edge 94+, or Safari 17.4+ required for WebCodecs.",it.appendChild(t)}const $e=100,Ne=200;function Mn(t){const e=t==="webcodecs"||t==="auto"&&ot,n=U.querySelector("canvas, video");if(n&&n.remove(),e){const s=document.createElement("canvas");U.appendChild(s),U.setAttribute("jitter",$e),Ee.value=$e,He.textContent=`${$e} ms`,Lt.innerHTML=G("WebCodecs (canvas)","blue")}else{const s=document.createElement("video");s.autoplay=!0,s.muted=!0,U.appendChild(s),U.setAttribute("jitter",Ne),Ee.value=Ne,He.textContent=`${Ne} ms`,Lt.innerHTML=G("MSE (video)","yellow")}}const zn=ko?"mse":"auto";Mn(zn);document.querySelector(`input[name=mode][value="${zn}"]`).checked=!0;po.forEach(t=>{t.addEventListener("change",()=>Mn(t.value))});function qe(t){t===0?(U.setAttribute("muted",""),Ze.textContent="🔇",Bt.textContent="muted"):(U.removeAttribute("muted"),U.setAttribute("volume",(t/100).toFixed(2)),Ze.textContent=t<50?"🔉":"🔊",Bt.textContent=`${t}%`)}me.addEventListener("input",()=>qe(Number(me.value)));Ze.addEventListener("click",()=>{if(U.hasAttribute("muted")){const t=Number(me.value)||50;me.value=t,qe(t)}else me.value=0,qe(0)});Ee.addEventListener("input",()=>{const t=Number(Ee.value);He.textContent=`${t} ms`,U.setAttribute("jitter",t)});let Ft=0;function Ao(t){return t==null?"—":t<1024?`${t} B`:t<1024*1024?`${(t/1024).toFixed(1)} KB`:`${(t/(1024*1024)).toFixed(2)} MB`}setInterval(()=>{const t=U.backend;if(!t)return;const e=t.video.stats.peek();if(e){const o=e.frameCount-Ft;Ft=e.frameCount,ve.textContent=`${o} fps`,ve.className="dbg-value ok",Dt.textContent=Ao(e.bytesReceived)}else ve.textContent="N/A (MSE)",ve.className="dbg-value",Dt.textContent="—";const n=t.video.buffered.peek();if(n&&n.length>0){const o=n[n.length-1],c=o.end-o.start;Rt.textContent=`${c.toFixed(0)} ms`}else Rt.textContent="empty";const s=t.video.stalled.peek();Ot.textContent=s?"yes":"no",Ot.className=`dbg-value ${s?"stalled":"ok"}`;const i=t.video.source.sync.latency.peek();go.textContent=i!=null?`${i.toFixed(0)} ms`:"—";const r=t.jitter.peek();bo.textContent=r!=null?`${r} ms`:"—";const a=t.video.source.config.peek();wo.textContent=a?.codec??"—"},1e3);if(!ot){const t=document.querySelector("input[value=webcodecs]");t&&(t.disabled=!0,t.closest("label").style.opacity="0.4",t.closest("label").title="WebCodecs not supported in this browser")}function xo(t){if(t&&t.includes("://"))try{return new URL(t).hostname}catch{}return t?t.split(":")[0]:window.location.hostname||"localhost"}const $t=xo(vo),Ge=window.location.protocol,le=[`${Ge}//${$t}/moq`,`${Ge}//${$t}:4443/moq`];if(Ge==="http:"){const t=window.location.href.replace(/^http:/,"https:"),e=document.createElement("div");e.className="banner banner-warn",e.innerHTML=`⚠️ <strong>Page loaded over HTTP.</strong> WebTransport requires HTTPS — playback may fail. <a href="${t}" style="color:inherit;font-weight:bold;">Switch to HTTPS →</a>`,it.appendChild(e)}le.forEach((t,e)=>{const n=document.createElement("tr");n.id=`url-row-${e}`,n.innerHTML=`
-    <td>${e+1}</td>
-    <td>${t}</td>
-    <td id="url-status-${e}"><span class="status-dot dot-pending"></span>pending</td>
-  `,ho.appendChild(n)});function ke(t,e,n){const s=document.getElementById(`url-status-${t}`),i={pending:"dot-pending",trying:"dot-trying",ok:"dot-ok",fail:"dot-fail"}[e];s.innerHTML=`<span class="status-dot ${i}"></span>${n}`;const r=document.getElementById(`url-row-${t}`);r.className=e==="ok"?"url-row active":"url-row"}async function _o(t,e=4e3){const n=new AbortController,s=setTimeout(()=>n.abort(),e);try{return await fetch(t,{method:"HEAD",mode:"no-cors",signal:n.signal}),!0}catch(i){return i.name==="AbortError"?!1:!(i instanceof TypeError)}finally{clearTimeout(s)}}async function So(){for(let t=0;t<le.length;t++){const e=le[t];if(ke(t,"trying","trying…"),de.innerHTML=G(`probing ${t+1}/${le.length}`,"yellow"),await _o(e)){ke(t,"ok","✓ reachable");for(let s=t+1;s<le.length;s++)ke(s,"pending","skipped");return e}else ke(t,"fail","✗ unreachable")}return null}let Xe=ge??"",K=new Set,Tn="";function Ue(){if(Ct.innerHTML="",!Xe){Fe.style.display="none";return}const t=[...K].filter(e=>e.split("/").includes(Xe));if(t.length===0){Fe.style.display="none";return}Fe.style.display="flex";for(const e of t){const n=e===Tn,s=document.createElement("button");s.type="button",s.textContent=e.split("/").pop()??e,s.title=e,s.className="quality-btn"+(n?" active":""),s.addEventListener("click",()=>{U.name=e}),Ct.appendChild(s)}}let Nt=null;U.connection.established.subscribe(async t=>{if(K=new Set,Ue(),!t)return;Nt=t;const e=t.announced();try{for(;;){const n=await e.next();if(!n||Nt!==t)break;const s=n.path.toString();n.active?K=new Set([...K,s]):K=new Set([...K].filter(i=>i!==s)),Ue()}}catch{}});U.broadcast.name.subscribe(t=>{Tn=t.toString(),Ue()});function Eo(t){if(!t)return;Xe=t,U.setAttribute("name",t),Un.textContent=t,Ue();const e=new URL(window.location);e.searchParams.set("name",t),window.history.replaceState({},"",e)}mo.addEventListener("submit",t=>{t.preventDefault(),Eo(rt.value.trim())});(async()=>{const t=await So();if(!t){yo("Could not reach any relay URL. Check that the relay server is running and the host is correct."),It.textContent="none",de.innerHTML=G("unreachable","red");return}It.textContent=t,de.innerHTML=G("connecting…","yellow"),U.setAttribute("url",t),ge?(U.setAttribute("name",ge),de.innerHTML=G("connected","green")):(de.innerHTML=G("enter stream id","yellow"),rt.focus())})();
