+++ MoQPlugin/src/main/resources/moq-ams-player-build/assets/play-DA0pAR74.js
@@ -1,0 +1,476 @@
+import{S as d,E as _,e as zn,M as h,a as je,b as Me,d as Cn,_ as ot,R as In,f as at}from"./time-Cy1-YUvq.js";function g(t,e,n){function s(a,c){if(a._zod||Object.defineProperty(a,"_zod",{value:{def:c,constr:o,traits:new Set},enumerable:!1}),a._zod.traits.has(t))return;a._zod.traits.add(t),e(a,c);const u=o.prototype,l=Object.keys(u);for(let f=0;f<l.length;f++){const m=l[f];m in a||(a[m]=u[m].bind(a))}}const i=n?.Parent??Object;class r extends i{}Object.defineProperty(r,"name",{value:t});function o(a){var c;const u=n?.Parent?new r:this;s(u,a),(c=u._zod).deferred??(c.deferred=[]);for(const l of u._zod.deferred)l();return u}return Object.defineProperty(o,"init",{value:s}),Object.defineProperty(o,Symbol.hasInstance,{value:a=>n?.Parent&&a instanceof n.Parent?!0:a?._zod?.traits?.has(t)}),Object.defineProperty(o,"name",{value:t}),o}class Q extends Error{constructor(){super("Encountered Promise during synchronous parse. Use .parseAsync() instead.")}}class Ln extends Error{constructor(e){super(`Encountered unidirectional transform during encode: ${e}`),this.name="ZodEncodeError"}}const Rn={};function ie(t){return Rn}function Bn(t){const e=Object.values(t).filter(n=>typeof n=="number");return Object.entries(t).filter(([n,s])=>e.indexOf(+n)===-1).map(([n,s])=>s)}function On(t,e){return typeof e=="bigint"?e.toString():e}function Dt(t){return{get value(){{const e=t();return Object.defineProperty(this,"value",{value:e}),e}}}}function Nt(t){const e=t.startsWith("^")?1:0,n=t.endsWith("$")?t.length-1:t.length;return t.slice(e,n)}const ct=Symbol("evaluating");function T(t,e,n){let s;Object.defineProperty(t,e,{get(){if(s!==ct)return s===void 0&&(s=ct,s=n()),s},set(i){Object.defineProperty(t,e,{value:i})},configurable:!0})}const jt="captureStackTrace"in Error?Error.captureStackTrace:(...t)=>{};function Ae(t){return typeof t=="object"&&t!==null&&!Array.isArray(t)}function Pt(t){if(Ae(t)===!1)return!1;const e=t.constructor;if(e===void 0||typeof e!="function")return!0;const n=e.prototype;return!(Ae(n)===!1||Object.prototype.hasOwnProperty.call(n,"isPrototypeOf")===!1)}function $n(t){return Pt(t)?{...t}:Array.isArray(t)?[...t]:t}const Fn=new Set(["string","number","symbol"]);function Pe(t){return t.replace(/[.*+?^${}()|[\]\\]/g,"\\$&")}function Dn(t,e,n){const s=new t._zod.constr(e??t._zod.def);return(!e||n?.parent)&&(s._zod.parent=t),s}function I(t){return{}}function Nn(t){return Object.keys(t).filter(e=>t[e]._zod.optin==="optional"&&t[e]._zod.optout==="optional")}const jn={safeint:[Number.MIN_SAFE_INTEGER,Number.MAX_SAFE_INTEGER],int32:[-2147483648,2147483647],uint32:[0,4294967295],float32:[-34028234663852886e22,34028234663852886e22],float64:[-Number.MAX_VALUE,Number.MAX_VALUE]};function ue(t,e=0){if(t.aborted===!0)return!0;for(let n=e;n<t.issues.length;n++)if(t.issues[n]?.continue!==!0)return!0;return!1}function J(t,e){return e.map(n=>{var s;return(s=n).path??(s.path=[]),n.path.unshift(t),n})}function be(t){return typeof t=="string"?t:t?.message}function re(t,e,n){const s={...t,path:t.path??[]};if(!t.message){const i=be(t.inst?._zod.def?.error?.(t))??be(e?.error?.(t))??be(n.customError?.(t))??be(n.localeError?.(t))??"Invalid input";s.message=i}return delete s.inst,delete s.continue,e?.reportInput||delete s.input,s}const Vt=(t,e)=>{t.name="$ZodError",Object.defineProperty(t,"_zod",{value:t._zod,enumerable:!1}),Object.defineProperty(t,"issues",{value:e,enumerable:!1}),t.message=JSON.stringify(e,On,2),Object.defineProperty(t,"toString",{value:()=>t.message,enumerable:!1})},Pn=g("$ZodError",Vt),Te=g("$ZodError",Vt,{Parent:Error}),Vn=t=>(e,n,s,i)=>{const r=s?Object.assign(s,{async:!1}):{async:!1},o=e._zod.run({value:n,issues:[]},r);if(o instanceof Promise)throw new Q;if(o.issues.length){const a=new(i?.Err??t)(o.issues.map(c=>re(c,r,ie())));throw jt(a,i?.callee),a}return o.value},Wn=Vn(Te),Zn=t=>async(e,n,s,i)=>{const r=s?Object.assign(s,{async:!0}):{async:!0};let o=e._zod.run({value:n,issues:[]},r);if(o instanceof Promise&&(o=await o),o.issues.length){const a=new(i?.Err??t)(o.issues.map(c=>re(c,r,ie())));throw jt(a,i?.callee),a}return o.value},qn=Zn(Te),Hn=t=>(e,n,s)=>{const i=s?{...s,async:!1}:{async:!1},r=e._zod.run({value:n,issues:[]},i);if(r instanceof Promise)throw new Q;return r.issues.length?{success:!1,error:new(t??Pn)(r.issues.map(o=>re(o,i,ie())))}:{success:!0,data:r.value}},Wt=Hn(Te),Gn=t=>async(e,n,s)=>{const i=s?Object.assign(s,{async:!0}):{async:!0};let r=e._zod.run({value:n,issues:[]},i);return r instanceof Promise&&(r=await r),r.issues.length?{success:!1,error:new t(r.issues.map(o=>re(o,i,ie())))}:{success:!0,data:r.value}},Zt=Gn(Te),Xn=/^$|^(?:[0-9a-zA-Z+/]{4})*(?:(?:[0-9a-zA-Z+/]{2}==)|(?:[0-9a-zA-Z+/]{3}=))?$/,Yn=t=>{const e=t?`[\\s\\S]{${t?.minimum??0},${t?.maximum??""}}`:"[\\s\\S]*";return new RegExp(`^${e}$`)},Kn=/^-?\d+$/,qt=/^-?\d+(?:\.\d+)?$/,Jn=/^(?:true|false)$/i,ze=g("$ZodCheck",(t,e)=>{var n;t._zod??(t._zod={}),t._zod.def=e,(n=t._zod).onattach??(n.onattach=[])}),Ht={number:"number",bigint:"bigint",object:"date"},Qn=g("$ZodCheckLessThan",(t,e)=>{ze.init(t,e);const n=Ht[typeof e.value];t._zod.onattach.push(s=>{const i=s._zod.bag,r=(e.inclusive?i.maximum:i.exclusiveMaximum)??Number.POSITIVE_INFINITY;e.value<r&&(e.inclusive?i.maximum=e.value:i.exclusiveMaximum=e.value)}),t._zod.check=s=>{(e.inclusive?s.value<=e.value:s.value<e.value)||s.issues.push({origin:n,code:"too_big",maximum:typeof e.value=="object"?e.value.getTime():e.value,input:s.value,inclusive:e.inclusive,inst:t,continue:!e.abort})}}),es=g("$ZodCheckGreaterThan",(t,e)=>{ze.init(t,e);const n=Ht[typeof e.value];t._zod.onattach.push(s=>{const i=s._zod.bag,r=(e.inclusive?i.minimum:i.exclusiveMinimum)??Number.NEGATIVE_INFINITY;e.value>r&&(e.inclusive?i.minimum=e.value:i.exclusiveMinimum=e.value)}),t._zod.check=s=>{(e.inclusive?s.value>=e.value:s.value>e.value)||s.issues.push({origin:n,code:"too_small",minimum:typeof e.value=="object"?e.value.getTime():e.value,input:s.value,inclusive:e.inclusive,inst:t,continue:!e.abort})}}),ts=g("$ZodCheckNumberFormat",(t,e)=>{ze.init(t,e),e.format=e.format||"float64";const n=e.format?.includes("int"),s=n?"int":"number",[i,r]=jn[e.format];t._zod.onattach.push(o=>{const a=o._zod.bag;a.format=e.format,a.minimum=i,a.maximum=r,n&&(a.pattern=Kn)}),t._zod.check=o=>{const a=o.value;if(n){if(!Number.isInteger(a)){o.issues.push({expected:s,format:e.format,code:"invalid_type",continue:!1,input:a,inst:t});return}if(!Number.isSafeInteger(a)){a>0?o.issues.push({input:a,code:"too_big",maximum:Number.MAX_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:t,origin:s,inclusive:!0,continue:!e.abort}):o.issues.push({input:a,code:"too_small",minimum:Number.MIN_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:t,origin:s,inclusive:!0,continue:!e.abort});return}}a<i&&o.issues.push({origin:"number",input:a,code:"too_small",minimum:i,inclusive:!0,inst:t,continue:!e.abort}),a>r&&o.issues.push({origin:"number",input:a,code:"too_big",maximum:r,inclusive:!0,inst:t,continue:!e.abort})}}),ns=g("$ZodCheckStringFormat",(t,e)=>{var n,s;ze.init(t,e),t._zod.onattach.push(i=>{const r=i._zod.bag;r.format=e.format,e.pattern&&(r.patterns??(r.patterns=new Set),r.patterns.add(e.pattern))}),e.pattern?(n=t._zod).check??(n.check=i=>{e.pattern.lastIndex=0,!e.pattern.test(i.value)&&i.issues.push({origin:"string",code:"invalid_format",format:e.format,input:i.value,...e.pattern?{pattern:e.pattern.toString()}:{},inst:t,continue:!e.abort})}):(s=t._zod).check??(s.check=()=>{})}),ss={major:4,minor:3,patch:6},L=g("$ZodType",(t,e)=>{var n;t??(t={}),t._zod.def=e,t._zod.bag=t._zod.bag||{},t._zod.version=ss;const s=[...t._zod.def.checks??[]];t._zod.traits.has("$ZodCheck")&&s.unshift(t);for(const i of s)for(const r of i._zod.onattach)r(t);if(s.length===0)(n=t._zod).deferred??(n.deferred=[]),t._zod.deferred?.push(()=>{t._zod.run=t._zod.parse});else{const i=(o,a,c)=>{let u=ue(o),l;for(const f of a){if(f._zod.def.when){if(!f._zod.def.when(o))continue}else if(u)continue;const m=o.issues.length,w=f._zod.check(o);if(w instanceof Promise&&c?.async===!1)throw new Q;if(l||w instanceof Promise)l=(l??Promise.resolve()).then(async()=>{await w,o.issues.length!==m&&(u||(u=ue(o,m)))});else{if(o.issues.length===m)continue;u||(u=ue(o,m))}}return l?l.then(()=>o):o},r=(o,a,c)=>{if(ue(o))return o.aborted=!0,o;const u=i(a,s,c);if(u instanceof Promise){if(c.async===!1)throw new Q;return u.then(l=>t._zod.parse(l,c))}return t._zod.parse(u,c)};t._zod.run=(o,a)=>{if(a.skipChecks)return t._zod.parse(o,a);if(a.direction==="backward"){const u=t._zod.parse({value:o.value,issues:[]},{...a,skipChecks:!0});return u instanceof Promise?u.then(l=>r(l,o,a)):r(u,o,a)}const c=t._zod.parse(o,a);if(c instanceof Promise){if(a.async===!1)throw new Q;return c.then(u=>i(u,s,a))}return i(c,s,a)}}T(t,"~standard",()=>({validate:i=>{try{const r=Wt(t,i);return r.success?{value:r.data}:{issues:r.error?.issues}}catch{return Zt(t,i).then(r=>r.success?{value:r.data}:{issues:r.error?.issues})}},vendor:"zod",version:1}))}),Gt=g("$ZodString",(t,e)=>{L.init(t,e),t._zod.pattern=[...t?._zod.bag?.patterns??[]].pop()??Yn(t._zod.bag),t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=String(n.value)}catch{}return typeof n.value=="string"||n.issues.push({expected:"string",code:"invalid_type",input:n.value,inst:t}),n}}),Xt=g("$ZodStringFormat",(t,e)=>{ns.init(t,e),Gt.init(t,e)});function is(t){if(t==="")return!0;if(t.length%4!==0)return!1;try{return atob(t),!0}catch{return!1}}const rs=g("$ZodBase64",(t,e)=>{e.pattern??(e.pattern=Xn),Xt.init(t,e),t._zod.bag.contentEncoding="base64",t._zod.check=n=>{is(n.value)||n.issues.push({code:"invalid_format",format:"base64",input:n.value,inst:t,continue:!e.abort})}}),Yt=g("$ZodNumber",(t,e)=>{L.init(t,e),t._zod.pattern=t._zod.bag.pattern??qt,t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=Number(n.value)}catch{}const i=n.value;if(typeof i=="number"&&!Number.isNaN(i)&&Number.isFinite(i))return n;const r=typeof i=="number"?Number.isNaN(i)?"NaN":Number.isFinite(i)?void 0:"Infinity":void 0;return n.issues.push({expected:"number",code:"invalid_type",input:i,inst:t,...r?{received:r}:{}}),n}}),os=g("$ZodNumberFormat",(t,e)=>{ts.init(t,e),Yt.init(t,e)}),as=g("$ZodBoolean",(t,e)=>{L.init(t,e),t._zod.pattern=Jn,t._zod.parse=(n,s)=>{if(e.coerce)try{n.value=!!n.value}catch{}const i=n.value;return typeof i=="boolean"||n.issues.push({expected:"boolean",code:"invalid_type",input:i,inst:t}),n}});function ut(t,e,n){t.issues.length&&e.issues.push(...J(n,t.issues)),e.value[n]=t.value}const cs=g("$ZodArray",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{const i=n.value;if(!Array.isArray(i))return n.issues.push({expected:"array",code:"invalid_type",input:i,inst:t}),n;n.value=Array(i.length);const r=[];for(let o=0;o<i.length;o++){const a=i[o],c=e.element._zod.run({value:a,issues:[]},s);c instanceof Promise?r.push(c.then(u=>ut(u,n,o))):ut(c,n,o)}return r.length?Promise.all(r).then(()=>n):n}});function xe(t,e,n,s,i){if(t.issues.length){if(i&&!(n in s))return;e.issues.push(...J(n,t.issues))}t.value===void 0?n in s&&(e.value[n]=void 0):e.value[n]=t.value}function us(t){const e=Object.keys(t.shape);for(const s of e)if(!t.shape?.[s]?._zod?.traits?.has("$ZodType"))throw new Error(`Invalid element at key "${s}": expected a Zod schema`);const n=Nn(t.shape);return{...t,keys:e,keySet:new Set(e),numKeys:e.length,optionalKeys:new Set(n)}}function ls(t,e,n,s,i,r){const o=[],a=i.keySet,c=i.catchall._zod,u=c.def.type,l=c.optout==="optional";for(const f in e){if(a.has(f))continue;if(u==="never"){o.push(f);continue}const m=c.run({value:e[f],issues:[]},s);m instanceof Promise?t.push(m.then(w=>xe(w,n,f,e,l))):xe(m,n,f,e,l)}return o.length&&n.issues.push({code:"unrecognized_keys",keys:o,input:e,inst:r}),t.length?Promise.all(t).then(()=>n):n}const ds=g("$ZodObject",(t,e)=>{if(L.init(t,e),!Object.getOwnPropertyDescriptor(e,"shape")?.get){const o=e.shape;Object.defineProperty(e,"shape",{get:()=>{const a={...o};return Object.defineProperty(e,"shape",{value:a}),a}})}const n=Dt(()=>us(e));T(t._zod,"propValues",()=>{const o=e.shape,a={};for(const c in o){const u=o[c]._zod;if(u.values){a[c]??(a[c]=new Set);for(const l of u.values)a[c].add(l)}}return a});const s=Ae,i=e.catchall;let r;t._zod.parse=(o,a)=>{r??(r=n.value);const c=o.value;if(!s(c))return o.issues.push({expected:"object",code:"invalid_type",input:c,inst:t}),o;o.value={};const u=[],l=r.shape;for(const f of r.keys){const m=l[f],w=m._zod.optout==="optional",b=m._zod.run({value:c[f],issues:[]},a);b instanceof Promise?u.push(b.then(k=>xe(k,o,f,c,w))):xe(b,o,f,c,w)}return i?ls(u,c,o,a,n.value,t):u.length?Promise.all(u).then(()=>o):o}});function lt(t,e,n,s){for(const r of t)if(r.issues.length===0)return e.value=r.value,e;const i=t.filter(r=>!ue(r));return i.length===1?(e.value=i[0].value,i[0]):(e.issues.push({code:"invalid_union",input:e.value,inst:n,errors:t.map(r=>r.issues.map(o=>re(o,s,ie())))}),e)}const Kt=g("$ZodUnion",(t,e)=>{L.init(t,e),T(t._zod,"optin",()=>e.options.some(i=>i._zod.optin==="optional")?"optional":void 0),T(t._zod,"optout",()=>e.options.some(i=>i._zod.optout==="optional")?"optional":void 0),T(t._zod,"values",()=>{if(e.options.every(i=>i._zod.values))return new Set(e.options.flatMap(i=>Array.from(i._zod.values)))}),T(t._zod,"pattern",()=>{if(e.options.every(i=>i._zod.pattern)){const i=e.options.map(r=>r._zod.pattern);return new RegExp(`^(${i.map(r=>Nt(r.source)).join("|")})$`)}});const n=e.options.length===1,s=e.options[0]._zod.run;t._zod.parse=(i,r)=>{if(n)return s(i,r);let o=!1;const a=[];for(const c of e.options){const u=c._zod.run({value:i.value,issues:[]},r);if(u instanceof Promise)a.push(u),o=!0;else{if(u.issues.length===0)return u;a.push(u)}}return o?Promise.all(a).then(c=>lt(c,i,t,r)):lt(a,i,t,r)}}),fs=g("$ZodDiscriminatedUnion",(t,e)=>{e.inclusive=!1,Kt.init(t,e);const n=t._zod.parse;T(t._zod,"propValues",()=>{const i={};for(const r of e.options){const o=r._zod.propValues;if(!o||Object.keys(o).length===0)throw new Error(`Invalid discriminated union option at index "${e.options.indexOf(r)}"`);for(const[a,c]of Object.entries(o)){i[a]||(i[a]=new Set);for(const u of c)i[a].add(u)}}return i});const s=Dt(()=>{const i=e.options,r=new Map;for(const o of i){const a=o._zod.propValues?.[e.discriminator];if(!a||a.size===0)throw new Error(`Invalid discriminated union option at index "${e.options.indexOf(o)}"`);for(const c of a){if(r.has(c))throw new Error(`Duplicate discriminator value "${String(c)}"`);r.set(c,o)}}return r});t._zod.parse=(i,r)=>{const o=i.value;if(!Ae(o))return i.issues.push({code:"invalid_type",expected:"object",input:o,inst:t}),i;const a=s.value.get(o?.[e.discriminator]);return a?a._zod.run(i,r):e.unionFallback?n(i,r):(i.issues.push({code:"invalid_union",errors:[],note:"No matching discriminator",discriminator:e.discriminator,input:o,path:[e.discriminator],inst:t}),i)}}),hs=g("$ZodRecord",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{const i=n.value;if(!Pt(i))return n.issues.push({expected:"record",code:"invalid_type",input:i,inst:t}),n;const r=[],o=e.keyType._zod.values;if(o){n.value={};const a=new Set;for(const u of o)if(typeof u=="string"||typeof u=="number"||typeof u=="symbol"){a.add(typeof u=="number"?u.toString():u);const l=e.valueType._zod.run({value:i[u],issues:[]},s);l instanceof Promise?r.push(l.then(f=>{f.issues.length&&n.issues.push(...J(u,f.issues)),n.value[u]=f.value})):(l.issues.length&&n.issues.push(...J(u,l.issues)),n.value[u]=l.value)}let c;for(const u in i)a.has(u)||(c=c??[],c.push(u));c&&c.length>0&&n.issues.push({code:"unrecognized_keys",input:i,inst:t,keys:c})}else{n.value={};for(const a of Reflect.ownKeys(i)){if(a==="__proto__")continue;let c=e.keyType._zod.run({value:a,issues:[]},s);if(c instanceof Promise)throw new Error("Async schemas not supported in object keys currently");if(typeof a=="string"&&qt.test(a)&&c.issues.length){const l=e.keyType._zod.run({value:Number(a),issues:[]},s);if(l instanceof Promise)throw new Error("Async schemas not supported in object keys currently");l.issues.length===0&&(c=l)}if(c.issues.length){e.mode==="loose"?n.value[a]=i[a]:n.issues.push({code:"invalid_key",origin:"record",issues:c.issues.map(l=>re(l,s,ie())),input:a,path:[a],inst:t});continue}const u=e.valueType._zod.run({value:i[a],issues:[]},s);u instanceof Promise?r.push(u.then(l=>{l.issues.length&&n.issues.push(...J(a,l.issues)),n.value[c.value]=l.value})):(u.issues.length&&n.issues.push(...J(a,u.issues)),n.value[c.value]=u.value)}}return r.length?Promise.all(r).then(()=>n):n}}),ms=g("$ZodEnum",(t,e)=>{L.init(t,e);const n=Bn(e.entries),s=new Set(n);t._zod.values=s,t._zod.pattern=new RegExp(`^(${n.filter(i=>Fn.has(typeof i)).map(i=>typeof i=="string"?Pe(i):i.toString()).join("|")})$`),t._zod.parse=(i,r)=>{const o=i.value;return s.has(o)||i.issues.push({code:"invalid_value",values:n,input:o,inst:t}),i}}),ps=g("$ZodLiteral",(t,e)=>{if(L.init(t,e),e.values.length===0)throw new Error("Cannot create literal schema with no valid values");const n=new Set(e.values);t._zod.values=n,t._zod.pattern=new RegExp(`^(${e.values.map(s=>typeof s=="string"?Pe(s):s?Pe(s.toString()):String(s)).join("|")})$`),t._zod.parse=(s,i)=>{const r=s.value;return n.has(r)||s.issues.push({code:"invalid_value",values:e.values,input:r,inst:t}),s}}),gs=g("$ZodTransform",(t,e)=>{L.init(t,e),t._zod.parse=(n,s)=>{if(s.direction==="backward")throw new Ln(t.constructor.name);const i=e.transform(n.value,n);if(s.async)return(i instanceof Promise?i:Promise.resolve(i)).then(r=>(n.value=r,n));if(i instanceof Promise)throw new Q;return n.value=i,n}});function dt(t,e){return t.issues.length&&e===void 0?{issues:[],value:void 0}:t}const ws=g("$ZodOptional",(t,e)=>{L.init(t,e),t._zod.optin="optional",t._zod.optout="optional",T(t._zod,"values",()=>e.innerType._zod.values?new Set([...e.innerType._zod.values,void 0]):void 0),T(t._zod,"pattern",()=>{const n=e.innerType._zod.pattern;return n?new RegExp(`^(${Nt(n.source)})?$`):void 0}),t._zod.parse=(n,s)=>{if(e.innerType._zod.optin==="optional"){const i=e.innerType._zod.run(n,s);return i instanceof Promise?i.then(r=>dt(r,n.value)):dt(i,n.value)}return n.value===void 0?n:e.innerType._zod.run(n,s)}}),bs=g("$ZodDefault",(t,e)=>{L.init(t,e),t._zod.optin="optional",T(t._zod,"values",()=>e.innerType._zod.values),t._zod.parse=(n,s)=>{if(s.direction==="backward")return e.innerType._zod.run(n,s);if(n.value===void 0)return n.value=e.defaultValue,n;const i=e.innerType._zod.run(n,s);return i instanceof Promise?i.then(r=>ft(r,e)):ft(i,e)}});function ft(t,e){return t.value===void 0&&(t.value=e.defaultValue),t}const ys=g("$ZodPipe",(t,e)=>{L.init(t,e),T(t._zod,"values",()=>e.in._zod.values),T(t._zod,"optin",()=>e.in._zod.optin),T(t._zod,"optout",()=>e.out._zod.optout),T(t._zod,"propValues",()=>e.in._zod.propValues),t._zod.parse=(n,s)=>{if(s.direction==="backward"){const r=e.out._zod.run(n,s);return r instanceof Promise?r.then(o=>ye(o,e.in,s)):ye(r,e.in,s)}const i=e.in._zod.run(n,s);return i instanceof Promise?i.then(r=>ye(r,e.out,s)):ye(i,e.out,s)}});function ye(t,e,n){return t.issues.length?(t.aborted=!0,t):e._zod.run({value:t.value,issues:t.issues},n)}var ht;class vs{constructor(){this._map=new WeakMap,this._idmap=new Map}add(e,...n){const s=n[0];return this._map.set(e,s),s&&typeof s=="object"&&"id"in s&&this._idmap.set(s.id,e),this}clear(){return this._map=new WeakMap,this._idmap=new Map,this}remove(e){const n=this._map.get(e);return n&&typeof n=="object"&&"id"in n&&this._idmap.delete(n.id),this._map.delete(e),this}get(e){const n=e._zod.parent;if(n){const s={...this.get(n)??{}};delete s.id;const i={...s,...this._map.get(e)};return Object.keys(i).length?i:void 0}return this._map.get(e)}has(e){return this._map.has(e)}}function ks(){return new vs}(ht=globalThis).__zod_globalRegistry??(ht.__zod_globalRegistry=ks());function As(t,e){return new t({type:"string",...I()})}function xs(t,e){return new t({type:"string",format:"base64",check:"string_format",abort:!1,...I()})}function Ss(t,e){return new t({type:"number",checks:[],...I()})}function _s(t,e){return new t({type:"number",check:"number_format",abort:!1,format:"safeint",...I()})}function Es(t,e){return new t({type:"boolean",...I()})}function Jt(t,e){return new Qn({check:"less_than",...I(),value:t,inclusive:!0})}function Us(t,e){return new es({check:"greater_than",...I(),value:t,inclusive:!0})}function Qt(t){return Us(0)}const R=g("ZodMiniType",(t,e)=>{if(!t._zod)throw new Error("Uninitialized schema in ZodMiniType.");L.init(t,e),t.def=e,t.type=e.type,t.parse=(n,s)=>Wn(t,n,s,{callee:t.parse}),t.safeParse=(n,s)=>Wt(t,n,s),t.parseAsync=async(n,s)=>qn(t,n,s,{callee:t.parseAsync}),t.safeParseAsync=async(n,s)=>Zt(t,n,s),t.check=(...n)=>t.clone({...e,checks:[...e.checks??[],...n.map(s=>typeof s=="function"?{_zod:{check:s,def:{check:"custom"},onattach:[]}}:s)]},{parent:!0}),t.with=t.check,t.clone=(n,s)=>Dn(t,n,s),t.brand=()=>t,t.register=((n,s)=>(n.add(t,s),t)),t.apply=n=>n(t)}),en=g("ZodMiniString",(t,e)=>{Gt.init(t,e),R.init(t,e)});function v(t){return As(en)}const Ms=g("ZodMiniStringFormat",(t,e)=>{Xt.init(t,e),en.init(t,e)}),Ts=g("ZodMiniBase64",(t,e)=>{rs.init(t,e),Ms.init(t,e)});function zs(t){return xs(Ts)}const tn=g("ZodMiniNumber",(t,e)=>{Yt.init(t,e),R.init(t,e)});function M(t){return Ss(tn)}const Cs=g("ZodMiniNumberFormat",(t,e)=>{os.init(t,e),tn.init(t,e)});function nn(t){return _s(Cs)}const Is=g("ZodMiniBoolean",(t,e)=>{as.init(t,e),R.init(t,e)});function Xe(t){return Es(Is)}const Ls=g("ZodMiniArray",(t,e)=>{cs.init(t,e),R.init(t,e)});function H(t,e){return new Ls({type:"array",element:t,...I()})}const Rs=g("ZodMiniObject",(t,e)=>{ds.init(t,e),R.init(t,e),T(t,"shape",()=>e.shape)});function A(t,e){const n={type:"object",shape:t??{},...I()};return new Rs(n)}const Bs=g("ZodMiniUnion",(t,e)=>{Kt.init(t,e),R.init(t,e)});function Ce(t,e){return new Bs({type:"union",options:t,...I()})}const Os=g("ZodMiniDiscriminatedUnion",(t,e)=>{fs.init(t,e),R.init(t,e)});function $s(t,e,n){return new Os({type:"union",options:e,discriminator:t,...I()})}const Fs=g("ZodMiniRecord",(t,e)=>{hs.init(t,e),R.init(t,e)});function sn(t,e,n){return new Fs({type:"record",keyType:t,valueType:e,...I()})}const Ds=g("ZodMiniEnum",(t,e)=>{ms.init(t,e),R.init(t,e),t.options=Object.values(e.entries)});function rn(t,e){const n=Array.isArray(t)?Object.fromEntries(t.map(s=>[s,s])):t;return new Ds({type:"enum",entries:n,...I()})}const Ns=g("ZodMiniLiteral",(t,e)=>{ps.init(t,e),R.init(t,e)});function Ve(t,e){return new Ns({type:"literal",values:Array.isArray(t)?t:[t],...I()})}const js=g("ZodMiniTransform",(t,e)=>{gs.init(t,e),R.init(t,e)});function on(t){return new js({type:"transform",transform:t})}const Ps=g("ZodMiniOptional",(t,e)=>{ws.init(t,e),R.init(t,e)});function p(t){return new Ps({type:"optional",innerType:t})}const Vs=g("ZodMiniDefault",(t,e)=>{bs.init(t,e),R.init(t,e)});function Ws(t,e){return new Vs({type:"default",innerType:t,get defaultValue(){return typeof e=="function"?e():$n(e)}})}const Zs=g("ZodMiniPipe",(t,e)=>{ys.init(t,e),R.init(t,e)});function an(t,e){return new Zs({type:"pipe",in:t,out:e})}const cn=Ws($s("kind",[A({kind:Ve("legacy")}),A({kind:Ve("cmaf"),init:zs()})]),{kind:"legacy"});M().check(nn(),Qt(),Jt(255)).brand("u8");const B=M().check(nn(),Qt(),Jt(Number.MAX_SAFE_INTEGER)).brand("u53");function Z(t){return B.parse(t)}const qs=A({name:v()}),mt=A({codec:v(),container:cn,description:p(v()),sampleRate:B,numberOfChannels:B,bitrate:p(B),jitter:p(B)}),Hs=Ce([A({renditions:sn(v(),mt)}),an(A({track:qs,config:mt}),on(t=>({renditions:{[t.track.name]:t.config}})))]),Gs=A({hardware:p(H(v())),software:p(H(v())),unsupported:p(H(v()))}),Xs=A({hardware:p(H(v())),software:p(H(v())),unsupported:p(H(v()))}),Ys=A({video:p(Gs),audio:p(Xs)}),pe=A({name:v()}),Ks=A({message:p(pe),typing:p(pe)}),Js=A({x:p(M()),y:p(M()),z:p(M()),s:p(M())}),Qs=A({initial:p(Js),track:p(pe),handle:p(v()),peers:p(pe)}),te={catalog:100,audio:80,video:60},ei=A({id:p(v()),name:p(v()),avatar:p(v()),color:p(v())}),ti=A({name:v()}),pt=A({codec:v(),container:cn,description:p(v()),codedWidth:p(B),codedHeight:p(B),displayAspectWidth:p(B),displayAspectHeight:p(B),framerate:p(M()),bitrate:p(B),optimizeForLatency:p(Xe()),jitter:p(B)}),ni=Ce([A({renditions:sn(v(),pt),display:p(A({width:B,height:B})),rotation:p(M()),flip:p(Xe())}),an(H(A({track:ti,config:pt})),on(t=>{const e=t[0]?.config;return{renditions:Object.fromEntries(t.map(n=>[n.track.name,n.config])),display:e?.displayAspectWidth!==void 0&&e?.displayAspectHeight!==void 0?{width:e.displayAspectWidth,height:e.displayAspectHeight}:void 0,rotation:void 0,flip:void 0}}))]),si=A({video:p(ni),audio:p(Hs),location:p(Qs),user:p(ei),chat:p(Ks),capabilities:p(Ys),preview:p(pe)});function ii(t){const e=new TextDecoder().decode(t);try{const n=JSON.parse(e);return si.parse(n)}catch(n){throw console.warn("invalid catalog",e),n}}async function ri(t){const e=await t.readFrame();if(e)return ii(e)}function un(t){return t instanceof ArrayBuffer||typeof SharedArrayBuffer<"u"&&t instanceof SharedArrayBuffer}const oi="utf-16",Re="utf-16be",gt="utf-16le",fe="utf-8";function ln(t,e={}){let n;un(t)?n=new DataView(t):n=new DataView(t.buffer,t.byteOffset,t.byteLength);let s=0,{encoding:i}=e;if(!i){const u=n.getUint8(0),l=n.getUint8(1);u==239&&l==187&&n.getUint8(2)==191?(i=fe,s=3):u==254&&l==255?(i=Re,s=2):u==255&&l==254?(i=gt,s=2):i=fe}if(typeof TextDecoder<"u")return new TextDecoder(i).decode(n);const{byteLength:r}=n,o=i!==Re;let a="",c;for(;s<r;){switch(i){case fe:if(c=n.getUint8(s),c<128)s++;else if(c>=194&&c<=223)if(s+1<r){const u=n.getUint8(s+1);u>=128&&u<=191?(c=(c&31)<<6|u&63,s+=2):s++}else s++;else if(c>=224&&c<=239)if(s+2<=r-1){const u=n.getUint8(s+1),l=n.getUint8(s+2);u>=128&&u<=191&&l>=128&&l<=191?(c=(c&15)<<12|(u&63)<<6|l&63,s+=3):s++}else s++;else if(c>=240&&c<=244)if(s+3<=r-1){const u=n.getUint8(s+1),l=n.getUint8(s+2),f=n.getUint8(s+3);u>=128&&u<=191&&l>=128&&l<=191&&f>=128&&f<=191?(c=(c&7)<<18|(u&63)<<12|(l&63)<<6|f&63,s+=4):s++}else s++;else s++;break;case Re:case oi:case gt:c=n.getUint16(s,o),s+=2;break}a+=String.fromCodePoint(c)}return a}function ai(t){return new TextEncoder().encode(t)}function ci(t){return{writers:t?.writers??{}}}const ui=["dinf","edts","grpl","mdia","meco","mfra","minf","moof","moov","mvex","schi","sinf","stbl","strk","traf","trak","tref","udta","vttc"];function dn(t){return"boxes"in t||ui.includes(t.type)}const wt="utf8",C="uint",ne="template",bt="string",yt="int",vt="data";var E=class{constructor(e,n){this.writeUint=(s,i)=>{const{dataView:r,cursor:o}=this;switch(i){case 1:r.setUint8(o,s);break;case 2:r.setUint16(o,s);break;case 3:{const a=(s&16776960)>>8,c=s&255;r.setUint16(o,a),r.setUint8(o+2,c);break}case 4:r.setUint32(o,s);break;case 8:{const a=Math.floor(s/Math.pow(2,32)),c=s-a*Math.pow(2,32);r.setUint32(o,a),r.setUint32(o+4,c);break}}this.cursor+=i},this.writeInt=(s,i)=>{const{dataView:r,cursor:o}=this;switch(i){case 1:r.setInt8(o,s);break;case 2:r.setInt16(o,s);break;case 4:r.setInt32(o,s);break;case 8:const a=Math.floor(s/Math.pow(2,32)),c=s-a*Math.pow(2,32);r.setUint32(o,a),r.setUint32(o+4,c);break}this.cursor+=i},this.writeString=s=>{for(let i=0,r=s.length;i<r;i++)this.writeUint(s.charCodeAt(i),1)},this.writeTerminatedString=s=>{if(s.length!==0){for(let i=0,r=s.length;i<r;i++)this.writeUint(s.charCodeAt(i),1);this.writeUint(0,1)}},this.writeUtf8TerminatedString=s=>{const i=ai(s);new Uint8Array(this.dataView.buffer).set(i,this.cursor),this.cursor+=i.length,this.writeUint(0,1)},this.writeBytes=s=>{Array.isArray(s)||(s=[s]);for(const i of s)new Uint8Array(this.dataView.buffer).set(i,this.cursor),this.cursor+=i.length},this.writeArray=(s,i,r,o)=>{const a=i===C?this.writeUint:i===ne?this.writeTemplate:this.writeInt;for(let c=0;c<o;c++)a(s[c]??0,r)},this.writeTemplate=(s,i)=>{const r=Math.round(s*Math.pow(2,i===4?16:8));this.writeUint(r,i)},this.writeBoxHeader=(s,i)=>{i>4294967295?(this.writeUint(1,4),this.writeString(s),this.writeUint(i,8)):(this.writeUint(i,4),this.writeString(s))},this.dataView=new DataView(new ArrayBuffer(n)),this.cursor=0,this.writeBoxHeader(e,n)}get buffer(){return this.dataView.buffer}get byteLength(){return this.dataView.byteLength}get byteOffset(){return this.dataView.byteOffset}writeFullBox(e,n){this.writeUint(e,1),this.writeUint(n,3)}};function fn(t,e){return Array.from(t,n=>di(n,e))}function Ye(t,e){const n=fn(t,e);return{bytes:n,size:n.reduce((s,i)=>s+i.byteLength,0)}}function li(t,e){const{bytes:n,size:s}=Ye(t.boxes,e),i=8+s,r=new E(t.type,i);return r.writeBytes(n),r}function di(t,e){let n=null;if("type"in t){const{type:s}=t,i=e.writers?.[s];if(i?n=i(t,e):dn(t)?n=li(t,e):"view"in t&&(n=t.view),!n)throw new Error(`No writer found for box type: ${s}`)}if("buffer"in t&&(n=t),!n)throw new Error("Invalid box");return new Uint8Array(n.buffer,n.byteOffset,n.byteLength)}function fi(t,e,n){const s=n>0?n:t.byteLength-(e-t.byteOffset);return new Uint8Array(t.buffer,e,Math.max(s,0))}function hi(t,e,n){let s=NaN;const i=e-t.byteOffset;switch(n){case 1:s=t.getInt8(i);break;case 2:s=t.getInt16(i);break;case 4:s=t.getInt32(i);break;case 8:const r=t.getInt32(i),o=t.getInt32(i+4);s=r*Math.pow(2,32)+o;break}return s}function X(t,e,n){const s=e-t.byteOffset;let i=NaN,r,o;switch(n){case 1:i=t.getUint8(s);break;case 2:i=t.getUint16(s);break;case 3:r=t.getUint16(s),o=t.getUint8(s+2),i=(r<<8)+o;break;case 4:i=t.getUint32(s);break;case 8:r=t.getUint32(s),o=t.getUint32(s+4),i=r*Math.pow(2,32)+o;break}return i}function kt(t,e,n){let s="";for(let i=0;i<n;i++){const r=X(t,e+i,1);s+=String.fromCharCode(r)}return s}function mi(t,e,n){const s=n/2;return X(t,e,s)+X(t,e+s,s)/Math.pow(2,s)}function pi(t,e){let n="",s=e;for(;s-t.byteOffset<t.byteLength;){const i=X(t,s,1);if(i===0)break;n+=String.fromCharCode(i),s++}return n}function gi(t,e){const n=t.byteLength-(e-t.byteOffset);return n>0?ln(new DataView(t.buffer,e,n),{encoding:fe}):""}function wi(t,e){const n=t.byteLength-(e-t.byteOffset);let s="";if(n>0){const i=new DataView(t.buffer,e,n);let r=0;for(;r<n&&i.getUint8(r)!==0;r++);s=ln(new DataView(t.buffer,e,r),{encoding:fe})}return s}var bi=class hn{constructor(e,n){this.truncated=!1,this.slice=(s,i)=>{const r=new hn(new DataView(this.dataView.buffer,s,i),this.config),o=this.offset-s,a=i-o;return this.offset+=a,r.jump(o),r},this.read=(s,i=0)=>{const{dataView:r,offset:o}=this;let a,c=i;switch(s){case C:a=X(r,o,i);break;case yt:a=hi(r,o,i);break;case ne:a=mi(r,o,i);break;case bt:i===-1?(a=pi(r,o),c=a.length+1):a=kt(r,o,i);break;case vt:a=fi(r,o,i),c=a.length;break;case wt:i===-1?(a=wi(r,o),c=a.length+1):a=gi(r,o);break;default:a=-1}return this.offset+=c,a},this.readUint=s=>this.read(C,s),this.readInt=s=>this.read(yt,s),this.readString=s=>this.read(bt,s),this.readTemplate=s=>this.read(ne,s),this.readData=s=>this.read(vt,s),this.readUtf8=s=>this.read(wt,s),this.readFullBox=()=>({version:this.readUint(1),flags:this.readUint(3)}),this.readArray=(s,i,r)=>{const o=[];for(let a=0;a<r;a++)o.push(this.read(s,i));return o},this.jump=s=>{this.offset+=s},this.readBox=()=>{const{dataView:s,offset:i}=this;let r=0;const o=X(s,i,4),a=kt(s,i+4,4),c={size:o,type:a};r+=8,c.size===1&&(c.largesize=X(s,i+r,8),r+=8);const u=c.size===0?this.bytesRemaining:c.largesize??c.size;if(this.cursor+u>s.byteLength)throw this.truncated=!0,new Error("Truncated box");return this.jump(r),a==="uuid"&&(c.usertype=this.readArray("uint",1,16)),c.view=this.slice(i,u),c},this.readBoxes=(s=-1)=>{const i=[];for(const r of this)if(i.push(r),s>0&&i.length>=s)break;return i},this.readEntries=(s,i)=>{const r=[];for(let o=0;o<s;o++)r.push(i());return r},this.dataView=un(e)?new DataView(e):e instanceof DataView?e:new DataView(e.buffer,e.byteOffset,e.byteLength),this.offset=this.dataView.byteOffset,this.config=n||{}}get buffer(){return this.dataView.buffer}get byteOffset(){return this.dataView.byteOffset}get byteLength(){return this.dataView.byteLength}get cursor(){return this.offset-this.dataView.byteOffset}get done(){return this.cursor>=this.dataView.byteLength||this.truncated}get bytesRemaining(){return this.dataView.byteLength-this.cursor}*[Symbol.iterator](){const{readers:e={}}=this.config;for(;!this.done;)try{const n=this.readBox(),{type:s,view:i}=n,r=e[s]||e[s.trim()];if(r&&Object.assign(n,r(i,s)),dn(n)&&!n.boxes){const o=[];for(const a of i)o.push(a);n.boxes=o}yield n}catch(n){if(n instanceof Error&&n.message==="Truncated box")break;throw n}}};function Ke(t,e){const n=[];for(const s of new bi(t,e))n.push(s);return n}function yi(t,e){return fn(t,ci(e))}function vi(t,e){const{readArray:n,readUint:s,readTemplate:i,readBoxes:r}=e;return{type:t,reserved1:n(C,1,6),dataReferenceIndex:s(2),reserved2:n(C,4,2),channelcount:s(2),samplesize:s(2),preDefined:s(2),reserved3:s(2),samplerate:i(4),boxes:r()}}function Je(t,e){const{readArray:n,readUint:s,readInt:i,readTemplate:r,readBoxes:o}=e;return{type:t,reserved1:n(C,1,6),dataReferenceIndex:s(2),preDefined1:s(2),reserved2:s(2),preDefined2:n(C,4,3),width:s(2),height:s(2),horizresolution:r(4),vertresolution:r(4),reserved3:s(4),frameCount:s(2),compressorName:n(C,1,32),depth:s(2),preDefined3:i(2),boxes:o()}}function At(t){return Je("avc1",t)}function ki(t){return Je("hev1",t)}function Ai(t){return Je("hvc1",t)}function xi(t){return{type:"mdat",data:t.readData(-1)}}function Si(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(e==1?8:4),i=t.readUint(e==1?8:4),r=t.readUint(4),o=t.readUint(e==1?8:4),a=t.readUint(2);return{type:"mdhd",version:e,flags:n,creationTime:s,modificationTime:i,timescale:r,duration:o,language:String.fromCharCode((a>>10&31)+96,(a>>5&31)+96,(a&31)+96),preDefined:t.readUint(2)}}function _i(t){return{type:"mfhd",...t.readFullBox(),sequenceNumber:t.readUint(4)}}function Ei(t){return vi("mp4a",t)}function Ui(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(4);return{type:"stsd",version:e,flags:n,entryCount:s,entries:t.readBoxes(s)}}function Mi(t){const{version:e,flags:n}=t.readFullBox();return{type:"tfdt",version:e,flags:n,baseMediaDecodeTime:t.readUint(e==1?8:4)}}function Ti(t){const{version:e,flags:n}=t.readFullBox();return{type:"tfhd",version:e,flags:n,trackId:t.readUint(4),baseDataOffset:n&1?t.readUint(8):void 0,sampleDescriptionIndex:n&2?t.readUint(4):void 0,defaultSampleDuration:n&8?t.readUint(4):void 0,defaultSampleSize:n&16?t.readUint(4):void 0,defaultSampleFlags:n&32?t.readUint(4):void 0}}function zi(t){const{version:e,flags:n}=t.readFullBox(),s=e===1?8:4;return{type:"tkhd",version:e,flags:n,creationTime:t.readUint(s),modificationTime:t.readUint(s),trackId:t.readUint(4),reserved1:t.readUint(4),duration:t.readUint(s),reserved2:t.readArray(C,4,2),layer:t.readUint(2),alternateGroup:t.readUint(2),volume:t.readTemplate(2),reserved3:t.readUint(2),matrix:t.readArray(ne,4,9),width:t.readTemplate(4),height:t.readTemplate(4)}}function Ci(t){return{type:"trex",...t.readFullBox(),trackId:t.readUint(4),defaultSampleDescriptionIndex:t.readUint(4),defaultSampleDuration:t.readUint(4),defaultSampleSize:t.readUint(4),defaultSampleFlags:t.readUint(4)}}function Ii(t){const{version:e,flags:n}=t.readFullBox(),s=t.readUint(4);let i,r;n&1&&(i=t.readInt(4)),n&4&&(r=t.readUint(4));const o=t.readEntries(s,()=>{const a={};return n&256&&(a.sampleDuration=t.readUint(4)),n&512&&(a.sampleSize=t.readUint(4)),n&1024&&(a.sampleFlags=t.readUint(4)),n&2048&&(a.sampleCompositionTimeOffset=e===1?t.readInt(4):t.readUint(4)),a});return{type:"trun",version:e,flags:n,sampleCount:s,dataOffset:i,firstSampleFlags:r,samples:o}}function Li(t,e){const n=t.entries.length,{bytes:s,size:i}=Ye(t.entries,e),r=new E("dref",16+i);return r.writeFullBox(t.version,t.flags),r.writeUint(n,4),r.writeBytes(s),r}function Ri(t){const e=t.compatibleBrands.length*4,n=new E("ftyp",16+e);n.writeString(t.majorBrand),n.writeUint(t.minorVersion,4);for(const s of t.compatibleBrands)n.writeString(s);return n}function Bi(t){const e=t.name.length+1,n=new E("hdlr",32+e);return n.writeFullBox(t.version,t.flags),n.writeUint(t.preDefined,4),n.writeString(t.handlerType),n.writeArray(t.reserved,C,4,3),n.writeTerminatedString(t.name),n}function Oi(t){const e=new E("mdat",8+t.data.length);return e.writeBytes(t.data),e}function $i(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("mdhd",n+s+i+4+2+2);r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.timescale,4),r.writeUint(t.duration,e);const o=t.language.length>=3?(t.language.charCodeAt(0)-96&31)<<10|(t.language.charCodeAt(1)-96&31)<<5|t.language.charCodeAt(2)-96&31:0;return r.writeUint(o,2),r.writeUint(t.preDefined,2),r}function Fi(t){const e=new E("mfhd",16);return e.writeFullBox(t.version,t.flags),e.writeUint(t.sequenceNumber,4),e}function Di(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("mvhd",n+s+i+4+4+2+2+8+36+24+4);return r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.timescale,4),r.writeUint(t.duration,e),r.writeTemplate(t.rate,4),r.writeTemplate(t.volume,2),r.writeUint(t.reserved1,2),r.writeArray(t.reserved2,C,4,2),r.writeArray(t.matrix,ne,4,9),r.writeArray(t.preDefined,C,4,6),r.writeUint(t.nextTrackId,4),r}function Ni(t){const e=new E("smhd",16);return e.writeFullBox(t.version,t.flags),e.writeUint(t.balance,2),e.writeUint(t.reserved,2),e}function ji(t,e){const n=t.entries.length,{bytes:s,size:i}=Ye(t.entries,e),r=new E("stsd",16+i);return r.writeFullBox(t.version,t.flags),r.writeUint(n,4),r.writeBytes(s),r}function Pi(t){const e=t.entryCount*8,n=new E("stts",16+e);n.writeFullBox(t.version,t.flags),n.writeUint(t.entryCount,4);for(const s of t.entries)n.writeUint(s.sampleCount,4),n.writeUint(s.sampleDelta,4);return n}function Vi(t){const e=t.version===1?8:4,n=8,s=4,i=e,r=new E("tfdt",n+s+i);return r.writeFullBox(t.version,t.flags),r.writeUint(t.baseMediaDecodeTime,e),r}function Wi(t){const e=t.flags&1?8:0,n=t.flags&2?4:0,s=t.flags&8?4:0,i=t.flags&16?4:0,r=t.flags&32?4:0,o=new E("tfhd",16+e+n+s+i+r);return o.writeFullBox(t.version,t.flags),o.writeUint(t.trackId,4),t.flags&1&&o.writeUint(t.baseDataOffset??0,8),t.flags&2&&o.writeUint(t.sampleDescriptionIndex??0,4),t.flags&8&&o.writeUint(t.defaultSampleDuration??0,4),t.flags&16&&o.writeUint(t.defaultSampleSize??0,4),t.flags&32&&o.writeUint(t.defaultSampleFlags??0,4),o}function Zi(t){const e=t.version===1?8:4,n=8,s=4,i=e*3,r=new E("tkhd",n+s+i+4+4+8+2+2+2+2+36+4+4);return r.writeFullBox(t.version,t.flags),r.writeUint(t.creationTime,e),r.writeUint(t.modificationTime,e),r.writeUint(t.trackId,4),r.writeUint(t.reserved1,4),r.writeUint(t.duration,e),r.writeArray(t.reserved2,C,4,2),r.writeUint(t.layer,2),r.writeUint(t.alternateGroup,2),r.writeTemplate(t.volume,2),r.writeUint(t.reserved3,2),r.writeArray(t.matrix,ne,4,9),r.writeTemplate(t.width,4),r.writeTemplate(t.height,4),r}function qi(t){const e=new E("trex",32);return e.writeFullBox(t.version,t.flags),e.writeUint(t.trackId,4),e.writeUint(t.defaultSampleDescriptionIndex,4),e.writeUint(t.defaultSampleDuration,4),e.writeUint(t.defaultSampleSize,4),e.writeUint(t.defaultSampleFlags,4),e}function Hi(t){const e=t.flags&1?4:0,n=t.flags&4?4:0;let s=0;t.flags&256&&(s+=4),t.flags&512&&(s+=4),t.flags&1024&&(s+=4),t.flags&2048&&(s+=4);const i=s*t.sampleCount,r=new E("trun",16+e+n+i);r.writeFullBox(t.version,t.flags),r.writeUint(t.sampleCount,4),t.flags&1&&r.writeUint(t.dataOffset??0,4),t.flags&4&&r.writeUint(t.firstSampleFlags??0,4);for(const o of t.samples)t.flags&256&&r.writeUint(o.sampleDuration??0,4),t.flags&512&&r.writeUint(o.sampleSize??0,4),t.flags&1024&&r.writeUint(o.sampleFlags??0,4),t.flags&2048&&r.writeUint(o.sampleCompositionTimeOffset??0,4);return r}function Gi(t){const e=t.location.length+1,n=new E("url ",12+e);return n.writeFullBox(t.version,t.flags),n.writeTerminatedString(t.location),n}function Xi(t){const e=new E("vmhd",20);return e.writeFullBox(t.version,t.flags),e.writeUint(t.graphicsmode,2),e.writeArray(t.opcolor,C,2,3),e}const Yi={avc1:At,avc3:At,hvc1:Ai,hev1:ki,mp4a:Ei,stsd:Ui,mdhd:Si,tkhd:zi,trex:Ci},mn={mfhd:_i,tfhd:Ti,tfdt:Mi,trun:Ii,mdat:xi};function j(t,e){for(const n of t){if(e(n))return n;const s=n.boxes;if(s&&Array.isArray(s)){const i=j(s,e);if(i)return i}}}function Qe(t){const e=new ArrayBuffer(t.byteLength);return new Uint8Array(e).set(t),e}function q(t){return e=>e.type===t}function Ie(t){const e=Ke(Qe(t),{readers:Yi}),n=j(e,q("mdhd"));if(!n)throw new Error("No mdhd box found in init segment");const s=j(e,q("tkhd"))?.trackId??1,i=j(e,q("stsd"));if(!i?.entries||i.entries.length===0)throw new Error("No stsd box found in init segment");const r=i.entries[0],o=Ki(r),a=j(e,c=>c.type==="trex"&&c.trackId===s);return{description:o,timescale:n.timescale,trackId:s,defaultSampleDuration:a?.defaultSampleDuration??0,defaultSampleSize:a?.defaultSampleSize??0,defaultSampleFlags:a?.defaultSampleFlags??0}}function Ki(t){if(!(!t.boxes||!Array.isArray(t.boxes)))for(const e of t.boxes){if(e instanceof Uint8Array){if(e.length>8){const s=String.fromCharCode(e[4],e[5],e[6],e[7]);if(s==="avcC"||s==="hvcC"||s==="dOps")return new Uint8Array(e.slice(8));if(s==="esds")return xt(new Uint8Array(e.slice(8)))}continue}const n=e.type;if(n==="avcC"||n==="hvcC"||n==="dOps"){if(e.view){const s=e.view,i=8,r=s.byteOffset+i,o=e.size-i;return new Uint8Array(s.buffer,r,o)}if(e.data instanceof Uint8Array)return new Uint8Array(e.data);if(e.raw instanceof Uint8Array)return new Uint8Array(e.raw.slice(8))}if(n==="esds"){let s;if(e.view){const i=e.view,r=8;s=new Uint8Array(i.buffer,i.byteOffset+r,e.size-r)}else e.data instanceof Uint8Array?s=new Uint8Array(e.data):e.raw instanceof Uint8Array&&(s=new Uint8Array(e.raw.slice(8)));if(s)return xt(s)}}}function xt(t){let e=4;for(;e<t.length;){const n=t[e++];let s=0;for(let i=0;i<4&&e<t.length;i++){const r=t[e++];if(s=s<<7|r&127,(r&128)===0)break}if(n===5)return e+s<=t.length?new Uint8Array(t.buffer,t.byteOffset+e,s):void 0;n===3?e+=3:n===4?e+=13:e+=s}}function pn(t,e){const n=Ke(Qe(t),{readers:mn});return(j(n,q("tfdt"))?.baseMediaDecodeTime??0)*1e6/e.timescale}function Ji(t,e){const n=Ke(Qe(t),{readers:mn}),s=j(n,q("tfdt"))?.baseMediaDecodeTime??0,i=j(n,q("tfhd")),r=i?.defaultSampleDuration??e.defaultSampleDuration,o=i?.defaultSampleSize??e.defaultSampleSize,a=i?.defaultSampleFlags??e.defaultSampleFlags,c=j(n,q("trun"));if(!c)throw new Error("No trun box found in data segment");const u=j(n,q("mdat"));if(!u)throw new Error("No mdat box found in data segment");const l=u.data;if(!l)throw new Error("No data in mdat box");const f=[];let m=0,w=s;for(let b=0;b<c.sampleCount;b++){const k=c.samples[b]??{},S=k.sampleSize??o,z=k.sampleDuration??r;if(S<=0)throw new Error(`Invalid sample size ${S} for sample ${b} in trun`);if(z<0)throw new Error(`Invalid sample duration ${z} for sample ${b} in trun`);if(m+S>l.length)throw new Error(`Sample ${b} would overflow mdat: offset=${m}, size=${S}, mdatLength=${l.length}`);const x=b===0&&c.firstSampleFlags!==void 0?c.firstSampleFlags:k.sampleFlags??a,P=k.sampleCompositionTimeOffset??0,F=new Uint8Array(l.slice(m,m+S));m+=S;const O=w+P,y=Math.round(O*1e6/e.timescale),D=x===0||(x&65536)===0;f.push({data:F,timestamp:y,keyframe:D}),w+=z}return f}function V(t){if(t=t.startsWith("0x")?t.slice(2):t,t.length%2)throw new Error("invalid hex string length");const e=t.match(/.{2}/g);if(!e)throw new Error("invalid hex string format");return new Uint8Array(e.map(n=>parseInt(n,16)))}const Se=[65536,0,0,0,65536,0,0,0,1073741824],Qi={ftyp:Ri,mvhd:Di,tkhd:Zi,mdhd:$i,hdlr:Bi,vmhd:Xi,smhd:Ni,"url ":Gi,dref:Li,stsd:ji,stts:Pi,trex:qi,mfhd:Fi,tfhd:Wi,tfdt:Vi,trun:Hi,mdat:Oi};function he(t){return yi(t,{writers:Qi})}function et(t,e,n,s){const i=12+s.length,r=new Uint8Array(i),o=new DataView(r.buffer);return o.setUint32(0,i,!1),r[4]=t.charCodeAt(0),r[5]=t.charCodeAt(1),r[6]=t.charCodeAt(2),r[7]=t.charCodeAt(3),o.setUint32(8,e<<24|n,!1),r.set(s,12),r}function gn(){const t=new Uint8Array(4);return et("stsc",0,0,t)}function wn(){const t=new Uint8Array(8);return et("stsz",0,0,t)}function bn(){const t=new Uint8Array(4);return et("stco",0,0,t)}function er(t,e,n){const s=8+n.length,i=8+(78+s),r=new Uint8Array(i),o=new DataView(r.buffer);let a=0;return o.setUint32(a,i,!1),a+=4,r[a++]=97,r[a++]=118,r[a++]=99,r[a++]=49,a+=6,o.setUint16(a,1,!1),a+=2,o.setUint16(a,0,!1),a+=2,o.setUint16(a,0,!1),a+=2,a+=12,o.setUint16(a,t,!1),a+=2,o.setUint16(a,e,!1),a+=2,o.setUint32(a,4718592,!1),a+=4,o.setUint32(a,4718592,!1),a+=4,o.setUint32(a,0,!1),a+=4,o.setUint16(a,1,!1),a+=2,a+=32,o.setUint16(a,24,!1),a+=2,o.setUint16(a,65535,!1),a+=2,o.setUint32(a,s,!1),a+=4,r[a++]=97,r[a++]=118,r[a++]=99,r[a++]=67,r.set(n,a),r}function tr(t){const{codedWidth:e,codedHeight:n,description:s}=t;if(!e||!n||!s)throw new Error("Missing required fields to create video init segment");const i=1e6,r=1,o={type:"ftyp",majorBrand:"isom",minorVersion:512,compatibleBrands:["isom","iso6","mp41"]},a={type:"mvhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:i,duration:0,rate:65536,volume:256,reserved1:0,reserved2:[0,0],matrix:Se,preDefined:[0,0,0,0,0,0],nextTrackId:r+1},c={type:"tkhd",version:0,flags:3,creationTime:0,modificationTime:0,trackId:r,reserved1:0,duration:0,reserved2:[0,0],layer:0,alternateGroup:0,volume:0,reserved3:0,matrix:Se,width:e*65536,height:n*65536},u={type:"mdhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:i,duration:0,language:"und",preDefined:0},l={type:"hdlr",version:0,flags:0,preDefined:0,handlerType:"vide",reserved:[0,0,0],name:"VideoHandler"},f={type:"vmhd",version:0,flags:1,graphicsmode:0,opcolor:[0,0,0]},m={type:"dinf",boxes:[{type:"dref",version:0,flags:0,entryCount:1,entries:[{type:"url ",version:0,flags:1,location:""}]}]},w={type:"stsd",version:0,flags:0,entryCount:1,entries:[er(e,n,V(s))]},b={type:"stts",version:0,flags:0,entryCount:0,entries:[]},k=gn(),S=wn(),z=bn(),x=he([o,{type:"moov",boxes:[a,{type:"trak",boxes:[c,{type:"mdia",boxes:[u,l,{type:"minf",boxes:[f,m,{type:"stbl",boxes:[w,b,k,S,z]}]}]}]},{type:"mvex",boxes:[{type:"trex",version:0,flags:0,trackId:r,defaultSampleDescriptionIndex:1,defaultSampleDuration:0,defaultSampleSize:0,defaultSampleFlags:0}]}]}]),P=x.reduce((y,D)=>y+D.byteLength,0),F=new Uint8Array(P);let O=0;for(const y of x)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;return F}function nr(t){const{sampleRate:e,numberOfChannels:n,description:s,codec:i}=t,r=1e6,o=1,a={type:"ftyp",majorBrand:"isom",minorVersion:512,compatibleBrands:["isom","iso6","mp41"]},c={type:"mvhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:r,duration:0,rate:65536,volume:256,reserved1:0,reserved2:[0,0],matrix:Se,preDefined:[0,0,0,0,0,0],nextTrackId:o+1},u={type:"tkhd",version:0,flags:3,creationTime:0,modificationTime:0,trackId:o,reserved1:0,duration:0,reserved2:[0,0],layer:0,alternateGroup:0,volume:256,reserved3:0,matrix:Se,width:0,height:0},l={type:"mdhd",version:0,flags:0,creationTime:0,modificationTime:0,timescale:r,duration:0,language:"und",preDefined:0},f={type:"hdlr",version:0,flags:0,preDefined:0,handlerType:"soun",reserved:[0,0,0],name:"SoundHandler"},m={type:"smhd",version:0,flags:0,balance:0,reserved:0},w={type:"dinf",boxes:[{type:"dref",version:0,flags:0,entryCount:1,entries:[{type:"url ",version:0,flags:1,location:""}]}]},b={type:"stsd",version:0,flags:0,entryCount:1,entries:[sr(i,e,n,s)]},k={type:"stts",version:0,flags:0,entryCount:0,entries:[]},S=gn(),z=wn(),x=bn(),P=he([a,{type:"moov",boxes:[c,{type:"trak",boxes:[u,{type:"mdia",boxes:[l,f,{type:"minf",boxes:[m,w,{type:"stbl",boxes:[b,k,S,z,x]}]}]}]},{type:"mvex",boxes:[{type:"trex",version:0,flags:0,trackId:o,defaultSampleDescriptionIndex:1,defaultSampleDuration:0,defaultSampleSize:0,defaultSampleFlags:0}]}]}]),F=P.reduce((D,Tn)=>D+Tn.byteLength,0),O=new Uint8Array(F);let y=0;for(const D of P)O.set(new Uint8Array(D.buffer,D.byteOffset,D.byteLength),y),y+=D.byteLength;return O}function sr(t,e,n,s){if(t.startsWith("mp4a"))return ir(e,n,s);if(t==="opus")return rr(e,n,s);throw new Error(`Unsupported audio codec: ${t}`)}function ir(t,e,n){const s=ar(t,e,n),i=8+(28+s.length),r=new Uint8Array(i),o=new DataView(r.buffer);let a=0;return o.setUint32(a,i,!1),a+=4,r[a++]=109,r[a++]=112,r[a++]=52,r[a++]=97,a+=6,o.setUint16(a,1,!1),a+=2,a+=8,o.setUint16(a,e,!1),a+=2,o.setUint16(a,16,!1),a+=2,o.setUint16(a,0,!1),a+=2,o.setUint16(a,0,!1),a+=2,o.setUint32(a,t*65536,!1),a+=4,r.set(s,a),r}function rr(t,e,n){const s=cr(e,t,n),i=8+(28+s.length),r=new Uint8Array(i),o=new DataView(r.buffer);let a=0;return o.setUint32(a,i,!1),a+=4,r[a++]=79,r[a++]=112,r[a++]=117,r[a++]=115,a+=6,o.setUint16(a,1,!1),a+=2,a+=8,o.setUint16(a,e,!1),a+=2,o.setUint16(a,16,!1),a+=2,o.setUint16(a,0,!1),a+=2,o.setUint16(a,0,!1),a+=2,o.setUint32(a,t*65536,!1),a+=4,r.set(s,a),r}function or(t,e){const n={96e3:0,88200:1,64e3:2,48e3:3,44100:4,32e3:5,24e3:6,22050:7,16e3:8,12e3:9,11025:10,8e3:11,7350:12}[t]??4,s=16|n>>1,i=(n&1)<<7|e<<3;return new Uint8Array([s,i])}function ar(t,e,n){const s=n?V(n):or(t,e),i=s.length,r=15+i,o=5+r+3,a=14+o,c=new Uint8Array(a),u=new DataView(c.buffer);let l=0;return u.setUint32(l,a,!1),l+=4,c[l++]=101,c[l++]=115,c[l++]=100,c[l++]=115,u.setUint32(l,0,!1),l+=4,c[l++]=3,c[l++]=o,u.setUint16(l,0,!1),l+=2,c[l++]=0,c[l++]=4,c[l++]=r,c[l++]=64,c[l++]=21,c[l++]=0,c[l++]=0,c[l++]=0,u.setUint32(l,0,!1),l+=4,u.setUint32(l,0,!1),l+=4,c[l++]=5,c[l++]=i,c.set(s,l),l+=i,c[l++]=6,c[l++]=1,c[l++]=2,c}function cr(t,e,n){if(n){const a=V(n),c=8+a.length,u=new Uint8Array(c);return new DataView(u.buffer).setUint32(0,c,!1),u[4]=100,u[5]=79,u[6]=112,u[7]=115,u.set(a,8),u}const s=19,i=new Uint8Array(s),r=new DataView(i.buffer);let o=0;return r.setUint32(o,s,!1),o+=4,i[o++]=100,i[o++]=79,i[o++]=112,i[o++]=115,i[o++]=0,i[o++]=t,r.setUint16(o,312,!1),o+=2,r.setUint32(o,e,!1),o+=4,r.setInt16(o,0,!1),o+=2,i[o++]=0,i}function yn(t){const{data:e,timestamp:n,duration:s,keyframe:i,sequence:r,trackId:o=1}=t,a=i?33554432:16842752,c={type:"mfhd",version:0,flags:0,sequenceNumber:r},u={type:"tfhd",version:0,flags:131072,trackId:o},l={type:"tfdt",version:1,flags:0,baseMediaDecodeTime:n},f={type:"trun",version:0,flags:1793,sampleCount:1,dataOffset:0,samples:[{sampleDuration:s,sampleSize:e.byteLength,sampleFlags:a}]},m={type:"moof",boxes:[c,{type:"traf",boxes:[u,l,f]}]},w=he([m]);let b=0;for(const y of w)b+=y.byteLength;f.dataOffset=b+8;const k=he([m]);b=0;for(const y of k)b+=y.byteLength;const S=new ArrayBuffer(e.byteLength),z=new Uint8Array(S);z.set(e);const x=he([{type:"mdat",data:z}]);let P=0;for(const y of x)P+=y.byteLength;const F=new Uint8Array(b+P);let O=0;for(const y of k)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;for(const y of x)F.set(new Uint8Array(y.buffer,y.byteOffset,y.byteLength),O),O+=y.byteLength;return F}let vn=class{#e;constructor(t){this.#e=t}decode(t){return Ji(t,this.#e).map(e=>({data:e.data,timestamp:e.timestamp,keyframe:e.keyframe}))}};class se{#e;#t;#n;#i=[];#s;#r;#a=new d([]);buffered=this.#a;#o=new _;constructor(e,n){this.#e=e,this.#t=n.format,this.#n=d.from(n.latency??h.zero),this.#o.spawn(this.#c.bind(this)),this.#o.cleanup(()=>{this.#e.close();for(const s of this.#i)s.consumer.close();this.#i.length=0})}async#c(){for(;;){const e=await this.#e.recvGroup();if(!e)break;if(this.#s===void 0&&(this.#s=e.sequence),e.sequence<this.#s){console.warn(`skipping old group: ${e.sequence} < ${this.#s}`),e.close();continue}const n={consumer:e,frames:[]};this.#i.push(n),this.#i.sort((s,i)=>s.consumer.sequence-i.consumer.sequence),this.#o.spawn(this.#u.bind(this,n))}}async#u(e){try{let n=0;for(;;){const s=await e.consumer.readFrame();if(!s)break;const i=this.#t.decode(s);for(const r of i){const o={data:r.data,timestamp:r.timestamp,keyframe:n===0?!0:r.keyframe};n++,e.frames.push(o),(e.latest===void 0||o.timestamp>e.latest)&&(e.latest=o.timestamp),this.#l(),e.consumer.sequence===this.#s?(this.#r?.(),this.#r=void 0):this.#d()}}}catch{}finally{e.done=!0,e.consumer.sequence===this.#s&&(this.#s+=1),this.#l(),this.#r?.(),this.#r=void 0,e.consumer.close()}}#d(){if(this.#s===void 0)return;let e=!1;for(;this.#i.length>=2;){const n=Me.fromMilli(this.#n.peek());let s,i;for(const o of this.#i){if(o.latest===void 0)continue;const a=o.frames.at(0)?.timestamp??o.latest;(s===void 0||a<s)&&(s=a),(i===void 0||o.latest>i)&&(i=o.latest)}if(s===void 0||i===void 0||i-s<=n)break;const r=this.#i.shift();if(!r)break;this.#s=this.#i[0]?.consumer.sequence,console.warn(`skipping slow group: ${r.consumer.sequence} -> ${this.#s}`),r.consumer.close(),r.frames.length=0,e=!0}e&&(this.#l(),this.#r?.(),this.#r=void 0)}async next(){for(;;){if(this.#i.length>0&&this.#s!==void 0&&this.#i[0].consumer.sequence<=this.#s){const n=this.#i[0].frames.shift();if(n)return this.#l(),{frame:n,group:this.#i[0].consumer.sequence};if(this.#s>this.#i[0].consumer.sequence||this.#i[0].done){this.#i[0].consumer.sequence===this.#s&&(this.#s+=1);const s=this.#i.shift();if(s)return this.#l(),{frame:void 0,group:s.consumer.sequence}}}if(this.#r)throw new Error("multiple calls to next not supported");const e=new Promise(n=>{this.#r=n}).then(()=>!0);if(!await Promise.race([e,this.#o.closed])){this.#r=void 0;return}}}#l(){const e=[];let n;for(const s of this.#i){const i=s.frames.at(0);if(!i||s.latest===void 0)continue;const r=h.fromMicro(i.timestamp),o=h.fromMicro(s.latest),a=e.at(-1),c=n?.done&&n.consumer.sequence+1===s.consumer.sequence;a&&(a.end>=r||c)?a.end=h.max(a.end,o):e.push({start:r,end:o}),n=s}this.#a.set(e)}close(){this.#o.close()}}class Le{decode(e){const[n,s]=Cn(e);return[{data:s,timestamp:n,keyframe:!1}]}}function _e(t,e){if(t.length===0)return e;if(e.length===0)return t;const n=[],s=[...t,...e].sort((i,r)=>i.start-r.start);for(const i of s){const r=n.at(-1);r&&r.end>=i.start?r.end=h.max(r.end,i.end):n.push({...i})}return n}let Be;async function St(){return globalThis.AudioEncoder&&globalThis.AudioDecoder?!0:(Be||(console.warn("using Opus polyfill; performance may be degraded"),Be=Promise.all([ot(()=>import("./libav-opus-af-BlMWboA7-CFTeN5TA.js"),[],import.meta.url),ot(()=>import("./main-DGBFe0O7-CIZu5tmC.js"),[],import.meta.url)]).then(async([t,e])=>(await e.load({LibAV:t,polyfill:!0}),!0))),await Be)}function we(t){const e=atob(t),n=new Uint8Array(e.length);for(let s=0;s<e.length;s++)n[s]=e.charCodeAt(s);return n}const W=0,N=1,oe=2,Y=3,ur=4;function kn(t,e,n){if(t<=0)throw new Error("invalid channels");if(e<=0)throw new Error("invalid capacity");if(n<=0)throw new Error("invalid sample rate");const s=new SharedArrayBuffer(t*e*Float32Array.BYTES_PER_ELEMENT),i=new SharedArrayBuffer(ur*Int32Array.BYTES_PER_ELEMENT),r=new Int32Array(i);return Atomics.store(r,Y,1),{channels:t,capacity:e,rate:n,samples:s,control:i}}function lr(t,e){return(t-e|0)>0?t:e}function ae(t,e){return(t%e+e)%e}function Oe(t,e,n){for(;;){const s=Atomics.load(t,e);if((n-s|0)<=0)return s;if(Atomics.compareExchange(t,e,s,n)===s)return n}}class tt{channels;capacity;rate;init;#e;#t;constructor(e){this.channels=e.channels,this.capacity=e.capacity,this.rate=e.rate,this.init=e,this.#e=new Int32Array(e.control),this.#t=[];for(let n=0;n<this.channels;n++)this.#t.push(new Float32Array(e.samples,n*this.capacity*Float32Array.BYTES_PER_ELEMENT,this.capacity))}insert(e,n){if(n.length!==this.channels)throw new Error("wrong number of channels");let s=Math.round(je.fromMicro(e)*this.rate);const i=n[0].length;let r=0;const o=s+i|0,a=Atomics.load(this.#e,N),c=a-s|0;if(c>0){if(c>=i)return;r=c,s=s+c|0}const u=i-r;(o-a|0)>this.capacity&&Oe(this.#e,N,o-this.capacity|0);const l=Atomics.load(this.#e,W),f=s-l|0;if(f>0){const k=Math.min(f,this.capacity);for(let S=0;S<this.channels;S++){const z=this.#t[S];for(let x=0;x<k;x++)z[ae(l+x|0,this.capacity)]=0}}for(let k=0;k<this.channels;k++){const S=n[k],z=this.#t[k];for(let x=0;x<u;x++)z[ae(s+x|0,this.capacity)]=S[r+x]}Atomics.store(this.#e,W,lr(Atomics.load(this.#e,W),o));const m=Atomics.load(this.#e,N),w=Atomics.load(this.#e,W),b=Atomics.load(this.#e,oe);(w-m|0)>=b&&b>0&&Atomics.store(this.#e,Y,0)}read(e){if(Atomics.load(this.#e,Y)===1)return 0;let n=Atomics.load(this.#e,N);const s=Atomics.load(this.#e,W),i=Atomics.load(this.#e,oe),r=s-n|0;if(i>0&&r>i){const c=s-i|0;n=Oe(this.#e,N,c)}const o=s-n|0,a=Math.min(o,e[0].length);if(a<=0)return 0;for(let c=0;c<this.channels;c++){const u=this.#t[c],l=e[c];for(let f=0;f<a;f++)l[f]=u[ae(n+f|0,this.capacity)]}return Oe(this.#e,N,n+a|0),a}setLatency(e){Atomics.store(this.#e,oe,e)}resize(e){const n=kn(this.channels,e,this.rate),s=new tt(n),i=Atomics.load(this.#e,N),r=Atomics.load(this.#e,W),o=Atomics.load(this.#e,oe),a=Atomics.load(this.#e,Y),c=r-i|0,u=Math.max(0,Math.min(c,s.capacity)),l=r-u|0;for(let f=0;f<this.channels;f++){const m=this.#t[f],w=s.#t[f];for(let b=0;b<u;b++){const k=l+b|0;w[ae(k,s.capacity)]=m[ae(k,this.capacity)]}}return Atomics.store(s.#e,N,l),Atomics.store(s.#e,W,r),Atomics.store(s.#e,oe,o),Atomics.store(s.#e,Y,a),s}get timestamp(){const e=Atomics.load(this.#e,N);return Me.fromSecond(e/this.rate)}get stalled(){return Atomics.load(this.#e,Y)===1}get length(){return Atomics.load(this.#e,W)-Atomics.load(this.#e,N)|0}}function dr(){return!(typeof SharedArrayBuffer>"u"||typeof crossOriginIsolated<"u"&&!crossOriginIsolated)}function fr(t,e,n,s){return dr()?(console.log("[audio] using SharedArrayBuffer audio buffer"),new hr(t,e,n,s)):(console.log("[audio] using postMessage audio buffer (SharedArrayBuffer unavailable)"),new mr(t,e,n,s))}class hr{rate;channels;#e;#t;#n=new d(0);timestamp=this.#n;#i=new d(!0);stalled=this.#i;#s=new _;constructor(e,n,s,i){this.#e=e,this.channels=n,this.rate=s;const r=Math.max(s,i*2),o=kn(n,r,s);this.#t=new tt(o),this.#t.setLatency(i);const a={type:"init-shared",...o};e.port.postMessage(a),this.#s.interval(()=>{this.#n.set(this.#t.timestamp),this.#i.set(this.#t.stalled)},50)}insert(e,n){this.#t.insert(e,n)}setLatency(e){if(this.#t.capacity<e*1.5){const n=Math.max(this.rate,e*2);this.#t=this.#t.resize(n),this.#t.setLatency(e);const s={type:"init-shared",...this.#t.init};this.#e.port.postMessage(s)}else this.#t.setLatency(e)}close(){this.#s.close()}}class mr{rate;channels;#e;#t=new d(0);timestamp=this.#t;#n=new d(!0);stalled=this.#n;#i=new _;constructor(e,n,s,i){this.#e=e,this.channels=n,this.rate=s;const r=h.fromSecond(i/s),o={type:"init-post",channels:n,rate:s,latency:r};e.port.postMessage(o),this.#i.event(e.port,"message",a=>{const c=a.data;c?.type==="state"&&(this.#t.set(c.timestamp),this.#n.set(c.stalled))}),e.port.start()}insert(e,n){const s={type:"data",data:n,timestamp:e};this.#e.port.postMessage(s,n.map(i=>i.buffer))}setLatency(e){const n={type:"latency",latency:h.fromSecond(e/this.rate)};this.#e.port.postMessage(n)}close(){this.#i.close()}}const pr=`var __defProp = Object.defineProperty;
+var __getOwnPropNames = Object.getOwnPropertyNames;
+var __esm = (fn, res) => function __init() {
+  return fn && (res = (0, fn[__getOwnPropNames(fn)[0]])(fn = 0)), res;
+};
+var __export = (target, all) => {
+  for (var name in all)
+    __defProp(target, name, { get: all[name], enumerable: true });
+};
+
+// ../lite/src/path.ts
+var init_path = __esm({
+  "../lite/src/path.ts"() {
+    "use strict";
+  }
+});
+
+// ../lite/src/varint.ts
+var MAX_U6, MAX_U14, MAX_U30, MAX_U53, MAX_U64, MAX_U62;
+var init_varint = __esm({
+  "../lite/src/varint.ts"() {
+    "use strict";
+    MAX_U6 = 2 ** 6 - 1;
+    MAX_U14 = 2 ** 14 - 1;
+    MAX_U30 = 2 ** 30 - 1;
+    MAX_U53 = Number.MAX_SAFE_INTEGER;
+    MAX_U64 = (1n << 64n) - 1n;
+    MAX_U62 = 2n ** 62n - 1n;
+  }
+});
+
+// ../lite/src/index.ts
+init_path();
+
+// ../lite/src/time.ts
+var time_exports = {};
+__export(time_exports, {
+  Micro: () => Micro,
+  Milli: () => Milli,
+  Nano: () => Nano,
+  Second: () => Second
+});
+var Nano = {
+  zero: 0,
+  fromMicro: (us) => us * 1e3,
+  fromMilli: (ms) => ms * 1e6,
+  fromSecond: (s) => s * 1e9,
+  toMicro: (ns) => ns / 1e3,
+  toMilli: (ns) => ns / 1e6,
+  toSecond: (ns) => ns / 1e9,
+  now: () => performance.now() * 1e6,
+  add: (a, b) => a + b,
+  sub: (a, b) => a - b,
+  mul: (a, b) => a * b,
+  div: (a, b) => a / b,
+  max: (a, b) => Math.max(a, b),
+  min: (a, b) => Math.min(a, b)
+};
+var Micro = {
+  zero: 0,
+  fromNano: (ns) => ns / 1e3,
+  fromMilli: (ms) => ms * 1e3,
+  fromSecond: (s) => s * 1e6,
+  toNano: (us) => us * 1e3,
+  toMilli: (us) => us / 1e3,
+  toSecond: (us) => us / 1e6,
+  now: () => performance.now() * 1e3,
+  add: (a, b) => a + b,
+  sub: (a, b) => a - b,
+  mul: (a, b) => a * b,
+  div: (a, b) => a / b,
+  max: (a, b) => Math.max(a, b),
+  min: (a, b) => Math.min(a, b)
+};
+var Milli = {
+  zero: 0,
+  fromNano: (ns) => ns / 1e6,
+  fromMicro: (us) => us / 1e3,
+  fromSecond: (s) => s * 1e3,
+  toNano: (ms) => ms * 1e6,
+  toMicro: (ms) => ms * 1e3,
+  toSecond: (ms) => ms / 1e3,
+  now: () => performance.now(),
+  add: (a, b) => a + b,
+  sub: (a, b) => a - b,
+  mul: (a, b) => a * b,
+  div: (a, b) => a / b,
+  max: (a, b) => Math.max(a, b),
+  min: (a, b) => Math.min(a, b)
+};
+var Second = {
+  zero: 0,
+  fromNano: (ns) => ns / 1e9,
+  fromMicro: (us) => us / 1e6,
+  fromMilli: (ms) => ms / 1e3,
+  toNano: (s) => s * 1e9,
+  toMicro: (s) => s * 1e6,
+  toMilli: (s) => s * 1e3,
+  now: () => performance.now() / 1e3,
+  add: (a, b) => a + b,
+  sub: (a, b) => a - b,
+  mul: (a, b) => a * b,
+  div: (a, b) => a / b,
+  max: (a, b) => Math.max(a, b),
+  min: (a, b) => Math.min(a, b)
+};
+
+// ../lite/src/index.ts
+init_varint();
+
+// src/audio/ring-buffer.ts
+var AudioRingBuffer = class {
+  #buffer;
+  #writeIndex = 0;
+  #readIndex = 0;
+  rate;
+  channels;
+  #stalled = true;
+  constructor(props) {
+    if (props.channels <= 0) throw new Error("invalid channels");
+    if (props.rate <= 0) throw new Error("invalid sample rate");
+    if (props.latency <= 0) throw new Error("invalid latency");
+    const samples = Math.ceil(props.rate * time_exports.Second.fromMilli(props.latency));
+    if (samples === 0) throw new Error("empty buffer");
+    this.rate = props.rate;
+    this.channels = props.channels;
+    this.#buffer = [];
+    for (let i = 0; i < this.channels; i++) {
+      this.#buffer[i] = new Float32Array(samples);
+    }
+  }
+  get stalled() {
+    return this.#stalled;
+  }
+  get timestamp() {
+    return time_exports.Micro.fromSecond(this.#readIndex / this.rate);
+  }
+  get length() {
+    return this.#writeIndex - this.#readIndex;
+  }
+  get capacity() {
+    return this.#buffer[0]?.length;
+  }
+  resize(latency) {
+    const newCapacity = Math.ceil(this.rate * time_exports.Second.fromMilli(latency));
+    if (newCapacity === this.capacity) return;
+    if (newCapacity === 0) throw new Error("empty buffer");
+    const newBuffer = [];
+    for (let i = 0; i < this.channels; i++) {
+      newBuffer[i] = new Float32Array(newCapacity);
+    }
+    const samplesToKeep = Math.min(this.length, newCapacity);
+    if (samplesToKeep > 0) {
+      const copyStart = this.#writeIndex - samplesToKeep;
+      for (let channel = 0; channel < this.channels; channel++) {
+        const src = this.#buffer[channel];
+        const dst = newBuffer[channel];
+        for (let i = 0; i < samplesToKeep; i++) {
+          const srcPos = (copyStart + i) % src.length;
+          const dstPos = (copyStart + i) % dst.length;
+          dst[dstPos] = src[srcPos];
+        }
+      }
+    }
+    this.#buffer = newBuffer;
+    this.#readIndex = this.#writeIndex - samplesToKeep;
+    if (samplesToKeep === 0) this.#stalled = true;
+  }
+  write(timestamp, data) {
+    if (data.length !== this.channels) throw new Error("wrong number of channels");
+    let start = Math.round(time_exports.Second.fromMicro(timestamp) * this.rate);
+    let samples = data[0].length;
+    let offset = this.#readIndex - start;
+    if (offset > samples) {
+      return;
+    } else if (offset > 0) {
+      samples -= offset;
+      start += offset;
+    } else {
+      offset = 0;
+    }
+    const end = start + samples;
+    const overflow = end - this.#readIndex - this.#buffer[0].length;
+    if (overflow >= 0) {
+      this.#stalled = false;
+      this.#readIndex += overflow;
+    }
+    if (start > this.#writeIndex) {
+      const gapSize = Math.min(start - this.#writeIndex, this.#buffer[0].length);
+      if (gapSize === 1) {
+        console.warn("floating point inaccuracy detected");
+      }
+      for (let channel = 0; channel < this.channels; channel++) {
+        const dst = this.#buffer[channel];
+        for (let i = 0; i < gapSize; i++) {
+          const writePos = (this.#writeIndex + i) % dst.length;
+          dst[writePos] = 0;
+        }
+      }
+    }
+    for (let channel = 0; channel < this.channels; channel++) {
+      let src = data[channel];
+      src = src.subarray(src.length - samples);
+      const dst = this.#buffer[channel];
+      if (src.length !== samples) throw new Error("mismatching number of samples");
+      for (let i = 0; i < samples; i++) {
+        const writePos = (start + i) % dst.length;
+        dst[writePos] = src[i];
+      }
+    }
+    if (end > this.#writeIndex) {
+      this.#writeIndex = end;
+    }
+  }
+  read(output) {
+    if (output.length !== this.channels) throw new Error("wrong number of channels");
+    if (this.#stalled) return 0;
+    const samples = Math.min(this.#writeIndex - this.#readIndex, output[0].length);
+    if (samples === 0) return 0;
+    for (let channel = 0; channel < this.channels; channel++) {
+      const dst = output[channel];
+      const src = this.#buffer[channel];
+      if (dst.length !== output[0].length) throw new Error("mismatching number of samples");
+      for (let i = 0; i < samples; i++) {
+        const readPos = (this.#readIndex + i) % src.length;
+        dst[i] = src[readPos];
+      }
+    }
+    this.#readIndex += samples;
+    return samples;
+  }
+};
+
+// src/audio/shared-ring-buffer.ts
+var WRITE = 0;
+var READ = 1;
+var LATENCY = 2;
+var STALLED = 3;
+var CONTROL_SLOTS = 4;
+function allocSharedRingBuffer(channels, capacity, rate) {
+  if (channels <= 0) throw new Error("invalid channels");
+  if (capacity <= 0) throw new Error("invalid capacity");
+  if (rate <= 0) throw new Error("invalid sample rate");
+  const samples = new SharedArrayBuffer(channels * capacity * Float32Array.BYTES_PER_ELEMENT);
+  const control = new SharedArrayBuffer(CONTROL_SLOTS * Int32Array.BYTES_PER_ELEMENT);
+  const ctrl = new Int32Array(control);
+  Atomics.store(ctrl, STALLED, 1);
+  return { channels, capacity, rate, samples, control };
+}
+function i32Max(a, b) {
+  return (a - b | 0) > 0 ? a : b;
+}
+function slot(idx, capacity) {
+  return (idx % capacity + capacity) % capacity;
+}
+function casAdvance(arr, idx, candidate) {
+  for (; ; ) {
+    const current = Atomics.load(arr, idx);
+    if ((candidate - current | 0) <= 0) return current;
+    const witnessed = Atomics.compareExchange(arr, idx, current, candidate);
+    if (witnessed === current) return candidate;
+  }
+}
+var SharedRingBuffer = class _SharedRingBuffer {
+  channels;
+  capacity;
+  rate;
+  init;
+  #control;
+  #samples;
+  constructor(init) {
+    this.channels = init.channels;
+    this.capacity = init.capacity;
+    this.rate = init.rate;
+    this.init = init;
+    this.#control = new Int32Array(init.control);
+    this.#samples = [];
+    for (let i = 0; i < this.channels; i++) {
+      this.#samples.push(
+        new Float32Array(init.samples, i * this.capacity * Float32Array.BYTES_PER_ELEMENT, this.capacity)
+      );
+    }
+  }
+  /**
+   * Insert audio samples at the given timestamp.
+   * Main thread only. Handles out-of-order writes, gap filling, and overflow.
+   */
+  insert(timestamp, data) {
+    if (data.length !== this.channels) throw new Error("wrong number of channels");
+    let start = Math.round(time_exports.Second.fromMicro(timestamp) * this.rate);
+    const originalLength = data[0].length;
+    let offset = 0;
+    const end = start + originalLength | 0;
+    const read = Atomics.load(this.#control, READ);
+    const behind = read - start | 0;
+    if (behind > 0) {
+      if (behind >= originalLength) {
+        return;
+      }
+      offset = behind;
+      start = start + behind | 0;
+    }
+    const samples = originalLength - offset;
+    if ((end - read | 0) > this.capacity) {
+      casAdvance(this.#control, READ, end - this.capacity | 0);
+    }
+    const write = Atomics.load(this.#control, WRITE);
+    const gap = start - write | 0;
+    if (gap > 0) {
+      const gapSize = Math.min(gap, this.capacity);
+      for (let channel = 0; channel < this.channels; channel++) {
+        const dst = this.#samples[channel];
+        for (let i = 0; i < gapSize; i++) {
+          dst[slot(write + i | 0, this.capacity)] = 0;
+        }
+      }
+    }
+    for (let channel = 0; channel < this.channels; channel++) {
+      const src = data[channel];
+      const dst = this.#samples[channel];
+      for (let i = 0; i < samples; i++) {
+        dst[slot(start + i | 0, this.capacity)] = src[offset + i];
+      }
+    }
+    Atomics.store(this.#control, WRITE, i32Max(Atomics.load(this.#control, WRITE), end));
+    const currentRead = Atomics.load(this.#control, READ);
+    const currentWrite = Atomics.load(this.#control, WRITE);
+    const latency = Atomics.load(this.#control, LATENCY);
+    if ((currentWrite - currentRead | 0) >= latency && latency > 0) {
+      Atomics.store(this.#control, STALLED, 0);
+    }
+  }
+  /**
+   * Read audio samples into the output buffers.
+   * AudioWorklet only. Returns the number of samples read.
+   */
+  read(output) {
+    if (Atomics.load(this.#control, STALLED) === 1) return 0;
+    let read = Atomics.load(this.#control, READ);
+    const write = Atomics.load(this.#control, WRITE);
+    const latency = Atomics.load(this.#control, LATENCY);
+    const buffered = write - read | 0;
+    if (latency > 0 && buffered > latency) {
+      const skipTo = write - latency | 0;
+      read = casAdvance(this.#control, READ, skipTo);
+    }
+    const available = write - read | 0;
+    const count = Math.min(available, output[0].length);
+    if (count <= 0) return 0;
+    for (let channel = 0; channel < this.channels; channel++) {
+      const src = this.#samples[channel];
+      const dst = output[channel];
+      for (let i = 0; i < count; i++) {
+        dst[i] = src[slot(read + i | 0, this.capacity)];
+      }
+    }
+    casAdvance(this.#control, READ, read + count | 0);
+    return count;
+  }
+  /** Update the target latency in samples. */
+  setLatency(samples) {
+    Atomics.store(this.#control, LATENCY, samples);
+  }
+  /**
+   * Allocate a new ring with \`newCapacity\` samples and copy the unread window
+   * [READ, WRITE) plus control state into it. Used when growing capacity so
+   * we don't drop buffered audio. If \`newCapacity\` is smaller than the unread
+   * span, the oldest samples are truncated.
+   *
+   * Main thread only. \`resize()\` reads from the source \`SharedRingBuffer\` and
+   * writes into a freshly allocated buffer from \`allocSharedRingBuffer\`, so it
+   * relies on the same invariant as \`insert()\`: no concurrent main-thread
+   * writers. The AudioWorklet reader is tolerated via the CAS discipline used
+   * by READ/WRITE elsewhere.
+   */
+  resize(newCapacity) {
+    const init = allocSharedRingBuffer(this.channels, newCapacity, this.rate);
+    const dst = new _SharedRingBuffer(init);
+    const read = Atomics.load(this.#control, READ);
+    const write = Atomics.load(this.#control, WRITE);
+    const latency = Atomics.load(this.#control, LATENCY);
+    const stalled = Atomics.load(this.#control, STALLED);
+    const available = write - read | 0;
+    const copyCount = Math.max(0, Math.min(available, dst.capacity));
+    const copyStart = write - copyCount | 0;
+    for (let channel = 0; channel < this.channels; channel++) {
+      const src = this.#samples[channel];
+      const out = dst.#samples[channel];
+      for (let i = 0; i < copyCount; i++) {
+        const idx = copyStart + i | 0;
+        out[slot(idx, dst.capacity)] = src[slot(idx, this.capacity)];
+      }
+    }
+    Atomics.store(dst.#control, READ, copyStart);
+    Atomics.store(dst.#control, WRITE, write);
+    Atomics.store(dst.#control, LATENCY, latency);
+    Atomics.store(dst.#control, STALLED, stalled);
+    return dst;
+  }
+  /** Current playback timestamp derived from READ position. */
+  get timestamp() {
+    const read = Atomics.load(this.#control, READ);
+    return time_exports.Micro.fromSecond(read / this.rate);
+  }
+  /** Whether the buffer is stalled (waiting to fill). */
+  get stalled() {
+    return Atomics.load(this.#control, STALLED) === 1;
+  }
+  /**
+   * Number of buffered samples (WRITE - READ).
+   *
+   * Non-atomic: WRITE and READ are loaded separately, so a concurrent
+   * writer/reader can make the two loads inconsistent. Intended for
+   * tests and diagnostics, not control-flow decisions.
+   */
+  get length() {
+    return Atomics.load(this.#control, WRITE) - Atomics.load(this.#control, READ) | 0;
+  }
+};
+
+// src/audio/render-worklet.ts
+var Render = class extends AudioWorkletProcessor {
+  // Set after init, depending on which path the main thread chose.
+  #backend;
+  #underflow = 0;
+  #stateCounter = 0;
+  constructor() {
+    super();
+    this.port.onmessage = (event) => {
+      const msg = event.data;
+      if (msg.type === "init-shared") {
+        console.log("[audio-worklet] init-shared: using SharedArrayBuffer path");
+        this.#backend = new SharedRingBuffer(msg);
+        this.#underflow = 0;
+      } else if (msg.type === "init-post") {
+        console.log("[audio-worklet] init-post: using postMessage path");
+        this.#backend = new AudioRingBuffer(msg);
+        this.#underflow = 0;
+      } else if (msg.type === "data") {
+        if (this.#backend instanceof AudioRingBuffer) this.#backend.write(msg.timestamp, msg.data);
+      } else if (msg.type === "latency") {
+        if (this.#backend instanceof AudioRingBuffer) this.#backend.resize(msg.latency);
+      }
+    };
+  }
+  process(_inputs, outputs, _parameters) {
+    const output = outputs[0];
+    const backend = this.#backend;
+    const samplesRead = backend?.read(output) ?? 0;
+    if (samplesRead < output[0].length) {
+      this.#underflow += output[0].length - samplesRead;
+    } else if (this.#underflow > 0 && backend) {
+      console.debug(\`audio underflow: \${Math.round(1e3 * this.#underflow / backend.rate)}ms\`);
+      this.#underflow = 0;
+    }
+    if (backend instanceof AudioRingBuffer) {
+      this.#stateCounter++;
+      if (this.#stateCounter >= 5) {
+        this.#stateCounter = 0;
+        const state = {
+          type: "state",
+          timestamp: backend.timestamp,
+          stalled: backend.stalled
+        };
+        this.port.postMessage(state);
+      }
+    }
+    return true;
+  }
+};
+registerProcessor("render", Render);
+`,gr=new Blob([pr],{type:"application/javascript"}),wr=URL.createObjectURL(gr);let br=class{source;enabled;#e=new d(void 0);context=this.#e;#t=new d(void 0);root=this.#t;#n=new d(void 0);sampleRate=this.#n;#i=new d(void 0);stats=this.#i;#s=new d(void 0);timestamp=this.#s;#r=new d(!0);stalled=this.#r;#a=new d([]);#o=new d([]);buffered=this.#o;#c;#u=new _;constructor(t,e){this.source=t,this.source.supported.set(yr),this.enabled=d.from(e?.enabled??!1),this.#u.run(this.#d.bind(this)),this.#u.run(this.#l.bind(this)),this.#u.run(this.#h.bind(this)),this.#u.run(this.#m.bind(this))}#d(t){const e=t.get(this.source.config);if(!e)return;const n=e.sampleRate,s=e.numberOfChannels,i=new AudioContext({latencyHint:"interactive",sampleRate:n});t.set(this.#e,i),t.cleanup(()=>i.close()),t.spawn(async()=>{if(await i.audioWorklet.addModule(wr),i.state==="closed")return;const r=new AudioWorkletNode(i,"render",{channelCount:s,channelCountMode:"explicit",outputChannelCount:[s]});t.cleanup(()=>r.disconnect());const o=this.source.sync.buffer.peek(),a=Math.ceil(n*je.fromMilli(o)),c=fr(r,s,n,a);this.#c=c,t.cleanup(()=>{c.close(),this.#c=void 0}),t.run(u=>{const l=h.fromMicro(u.get(c.timestamp));this.#s.set(l),this.#b(l)}),t.run(u=>{this.#r.set(u.get(c.stalled))}),t.set(this.#t,r)})}#l(t){const e=t.getAll([this.enabled,this.#e]);if(!e)return;const[n,s]=e;s.resume()}#h(t){if(!t.get(this.#t))return;const e=this.#c;if(!e)return;const n=t.get(this.source.sync.buffer),s=Math.ceil(e.rate*je.fromMilli(n));e.setLatency(s)}#m(t){if(!t.get(this.enabled))return;const e=t.get(this.source.broadcast);if(!e)return;const n=t.get(this.source.track);if(!n)return;const s=t.get(this.source.config);if(!s)return;const i=t.get(e.active);if(!i)return;const r=i.subscribe(n,te.audio);t.cleanup(()=>r.close()),s.container.kind==="cmaf"?this.#g(t,r,s):this.#p(t,r,s)}#p(t,e,n){const s=new se(e,{format:new Le,latency:this.source.sync.buffer});t.cleanup(()=>s.close()),t.run(i=>{const r=i.get(s.buffered),o=i.get(this.#a);this.#o.update(()=>_e(r,o))}),t.spawn(async()=>{if(!await St())return;let i=0;const r=new AudioDecoder({output:a=>{if(i++,i<=3){a.close();return}this.#f(a)},error:a=>console.error(a)});t.cleanup(()=>{r.state!=="closed"&&r.close()});const o=n.codec==="opus"?void 0:n.description?V(n.description):void 0;for(r.configure({...n,description:o});;){const a=await s.next();if(!a)break;const{frame:c}=a;if(!c)continue;const u=h.fromMicro(c.timestamp);this.source.sync.received(u,"audio"),this.#i.update(f=>({bytesReceived:(f?.bytesReceived??0)+c.data.byteLength}));const l=new EncodedAudioChunk({type:c.keyframe?"key":"delta",data:c.data,timestamp:c.timestamp});r.decode(l)}})}#g(t,e,n){if(n.container.kind!=="cmaf")return;const s=we(n.container.init),i=Ie(s),r=n.codec==="opus"?void 0:n.description?V(n.description):void 0,o=new se(e,{format:new vn(i),latency:this.source.sync.buffer});t.cleanup(()=>o.close()),t.run(a=>{const c=a.get(o.buffered),u=a.get(this.#a);this.#o.update(()=>_e(c,u))}),t.spawn(async()=>{if(!await St())return;const a=new AudioDecoder({output:c=>this.#f(c),error:c=>console.error(c)});for(t.cleanup(()=>{a.state!=="closed"&&a.close()}),a.configure({codec:n.codec,sampleRate:n.sampleRate,numberOfChannels:n.numberOfChannels,description:r});;){const c=await Promise.race([o.next(),t.cancel]);if(!c)break;const{frame:u}=c;if(!u)continue;const l=h.fromMicro(u.timestamp);if(this.source.sync.received(l,"audio"),this.#i.update(f=>({bytesReceived:(f?.bytesReceived??0)+u.data.byteLength})),a.state==="closed")break;a.decode(new EncodedAudioChunk({type:u.keyframe?"key":"delta",data:u.data,timestamp:u.timestamp}))}})}#f(t){const e=t.timestamp,n=h.fromMicro(e),s=this.#c;if(!s){t.close();return}const i=t.numberOfFrames/t.sampleRate*1e6,r=h.fromMicro(i),o=h.add(n,r);this.#w(n,o);const a=Math.min(t.numberOfChannels,s.channels),c=[];for(let u=0;u<a;u++){const l=new Float32Array(t.numberOfFrames);t.copyTo(l,{format:"f32-planar",planeIndex:u}),c.push(l)}s.insert(e,c),t.close()}#w(t,e){t>e||this.#a.mutate(n=>{for(const s of n)if(t<=s.end+1&&e>=s.start){s.start=h.min(s.start,t),s.end=h.max(s.end,e);return}n.push({start:t,end:e}),n.sort((s,i)=>s.start-i.start)})}#b(t){this.#a.mutate(e=>{for(;e.length>0;){if(e[0].end>=t){e[0].start=h.max(e[0].start,t);break}e.shift()}})}close(){this.#u.close()}};async function yr(t){const e=t.codec==="opus"?void 0:t.description?V(t.description):void 0;return(await AudioDecoder.isConfigSupported({...t,description:e})).supported??!1}const _t=.001,$e=.2;class vr{source;volume;muted;paused;#e=new _;#t=.5;#n=new d(void 0);constructor(e,n){this.source=e,this.volume=d.from(n?.volume??.5),this.muted=d.from(n?.muted??!1),this.paused=d.from(n?.paused??n?.muted??!1),this.#e.run(s=>{s.get(this.muted)?(this.#t=this.volume.peek()||.5,this.volume.set(0)):this.volume.set(this.#t)}),this.#e.run(s=>{const i=!s.get(this.paused)&&!s.get(this.muted);this.source.enabled.set(i)}),this.#e.run(s=>{const i=s.get(this.volume);this.muted.set(i===0)}),this.#e.run(s=>{const i=s.get(this.source.root);if(!i)return;const r=new GainNode(i.context,{gain:s.get(this.volume)});i.connect(r),s.set(this.#n,r),s.run(o=>{o.get(this.source.enabled)&&(r.connect(i.context.destination),o.cleanup(()=>r.disconnect()))})}),this.#e.run(s=>{const i=s.get(this.#n);if(!i)return;s.cleanup(()=>i.gain.cancelScheduledValues(i.context.currentTime));const r=s.get(this.volume);r<_t?(i.gain.exponentialRampToValueAtTime(_t,i.context.currentTime+$e),i.gain.setValueAtTime(0,i.context.currentTime+$e+.01)):i.gain.exponentialRampToValueAtTime(r,i.context.currentTime+$e)})}close(){this.#e.close()}}class kr{element;paused;#e;#t=new d(void 0);mediaSource=this.#t;#n=new _;constructor(e,n){this.element=d.from(n?.element),this.paused=d.from(n?.paused??!1),this.#e=e,this.#n.run(this.#i.bind(this)),this.#n.run(this.#s.bind(this)),this.#n.run(this.#r.bind(this)),this.#n.run(this.#a.bind(this)),this.#n.run(this.#o.bind(this))}#i(e){const n=e.get(this.element);if(!n)return;const s=new MediaSource;n.src=URL.createObjectURL(s),e.cleanup(()=>URL.revokeObjectURL(n.src)),e.event(s,"sourceopen",()=>{e.set(this.#t,s)},{once:!0}),e.event(s,"error",i=>{console.error("[MSE] MediaSource error event:",i)})}#s(e){const n=e.get(this.element);if(!n||e.get(this.paused))return;const s=h.toSecond(e.get(this.#e.buffer));e.interval(()=>{const i=n.buffered;if(i.length===0)return;const r=i.end(i.length-1)-s,o=r-n.currentTime;(o>.1||o<-.1)&&(console.warn("seeking",o>0?"forward":"backward",Math.abs(o).toFixed(3),"seconds"),n.currentTime=r)},100)}#r(e){const n=e.get(this.element);if(!n)return;const s=e.get(this.mediaSource);s&&e.interval(async()=>{for(const i of s.sourceBuffers){for(;i.updating;)await new Promise(r=>i.addEventListener("updateend",r,{once:!0}));n.currentTime>10&&i.remove(0,n.currentTime-10)}},1e3)}#a(e){const n=e.get(this.element);if(!n)return;const s=e.get(this.paused);s&&!n.paused?n.pause():!s&&n.paused&&n.play().catch(i=>{console.error("[MSE] MediaElement play error:",i)})}#o(e){const n=e.get(this.element);if(!n||e.get(this.paused))return;const s=e.get(this.#e.reference);if(s===void 0)return;const i=e.get(this.#e.buffer),r=h.sub(h.sub(h.now(),s),i);n.currentTime=h.toSecond(r)}close(){this.#n.close()}}const Ar=20,Et=100;class nt{#e=new d(void 0);reference=this.#e;latency;jitter;audio;video;#t=new d(h.zero);buffer=this.#t;#n;timestamp=new d(void 0);#i=new Map;#s;#r;signals=new _;constructor(e){this.latency=d.from(e?.latency??"real-time"),this.jitter=new d(Et),this.#s=e?.connection,this.audio=d.from(e?.audio),this.video=d.from(e?.video),this.#n=Promise.withResolvers(),this.signals.run(this.#a.bind(this)),this.signals.run(this.#o.bind(this))}#a(e){const n=e.get(this.latency);if(typeof n=="number"){this.#r=void 0,this.jitter.set(n);return}const s=(this.#s?e.get(this.#s):void 0)?.rtt,i=s?e.get(s):void 0;if(i!==void 0){this.#r=this.#r!==void 0?Math.min(this.#r,i):i;const r=Math.max(Ar,this.#r*1.25);this.jitter.set(r);return}this.#r=void 0,this.jitter.set(Et)}#o(e){const n=e.get(this.jitter),s=e.get(this.video)??h.zero,i=e.get(this.audio)??h.zero,r=h.add(h.max(s,i),n);this.#t.set(r),this.#n.resolve(),this.#n=Promise.withResolvers()}received(e,n=""){this.timestamp.update(o=>o===void 0||e>o?e:o);const s=h.now(),i=h.sub(s,e),r=this.#e.peek();if(r!==void 0){const o=h.add(h.sub(r,i),this.#t.peek());if(o<0){const a=this.#i.get(n);a?(a.count++,a.maxMs=Math.max(a.maxMs,-o)):this.#i.set(n,{count:1,maxMs:-o})}else{const a=this.#i.get(n);if(a){const c=n?`sync[${n}]`:"sync",u=nt.#c(a.maxMs);console.debug(`${c}: ${a.count} late frame(s), max ${u} behind`),this.#i.delete(n)}}if(i>=r)return}this.#e.set(i),this.#n.resolve(),this.#n=Promise.withResolvers()}now(){const e=this.#e.peek();if(e!==void 0)return h.sub(h.sub(h.now(),e),this.#t.peek())}async wait(e){if(this.#e.peek()===void 0)throw new Error("reference not set; call update() first");for(;;){const n=h.now(),s=h.sub(n,e),i=this.#e.peek();if(i===void 0)return;const r=h.add(h.sub(i,s),this.#t.peek());if(r<=0||r<5)return;const o=new Promise(a=>setTimeout(a,r)).then(()=>!0);if(await Promise.race([this.#n.promise,o]))return}}static#c(e){if(e=Math.round(e),e<1e3)return`${e}ms`;const n=e/1e3;if(n<60)return`${Math.round(n*10)/10}s`;const s=n/60;return`${Math.round(s*10)/10}m`}close(){this.signals.close()}}const xr=500,Sr=100;class _r{enabled;source;#e=new d(void 0);#t=new d(void 0);frame=this.#t;#n=new d(void 0);timestamp=this.#n;#i=new d(void 0);display=this.#i;#s=new d(!1);stalled=this.#s;#r=new d(void 0);stats=this.#r;#a=new d([]);buffered=this.#a;#o=new _;constructor(e,n){this.enabled=d.from(n?.enabled??!1),this.source=e,this.source.supported.set(Ur),this.#o.run(this.#c.bind(this)),this.#o.run(this.#u.bind(this)),this.#o.run(this.#d.bind(this)),this.#o.run(this.#l.bind(this))}#c(e){const n=e.getAll([this.enabled,this.source.broadcast,this.source.track,this.source.config]);if(!n){this.#e.set(void 0);return}const[s,i,r,o]=n,a=e.get(i.active);if(!a)return;let c=new Er({source:this.source,broadcast:a,track:r,config:o,stats:this.#r});e.cleanup(()=>c?.close()),e.run(u=>{if(!c)return;const l=u.get(this.#e);if(l){const f=u.get(c.timestamp),m=u.get(l.timestamp);if(!f||m&&m>f+Sr)return}this.#e.set(c),c=void 0,u.close()})}#u(e){const n=e.get(this.#e);if(!n){this.#a.set([]);return}e.cleanup(()=>n.close()),e.run(s=>{const i=s.get(n.frame);this.#t.update(r=>(r?.close(),i?.clone()))}),e.proxy(this.#n,n.timestamp),e.proxy(this.#a,n.buffered)}#d(e){const n=e.get(this.source.catalog);if(!n)return;const s=n.display;if(s){e.set(this.#i,{width:s.width,height:s.height});return}const i=e.get(this.frame);i&&e.set(this.#i,{width:i.displayWidth,height:i.displayHeight})}#l(e){if(e.get(this.enabled)){if(!e.get(this.frame)){this.#s.set(!0);return}this.#s.set(!1),e.timer(()=>{this.#s.set(!0)},xr)}}close(){this.#t.update(e=>{e?.close()}),this.#o.close()}}class Er{source;broadcast;track;config;stats;timestamp=new d(void 0);frame=new d(void 0);buffered=new d([]);#e=new d([]);signals=new _;constructor(e){const{codedWidth:n,codedHeight:s,...i}=e.config;this.source=e.source,this.broadcast=e.broadcast,this.track=e.track,this.config=i,this.stats=e.stats,this.signals.run(this.#t.bind(this))}#t(e){const n=this.broadcast.subscribe(this.track,te.video);e.cleanup(()=>n.close());const s=new VideoDecoder({output:async i=>{try{const r=h.fromMicro(i.timestamp);if(r<(this.timestamp.peek()??0))return;this.frame.peek()===void 0&&this.frame.set(i.clone());const o=this.source.sync.wait(r).then(()=>!0);if(!await Promise.race([o,e.cancel])||r<(this.timestamp.peek()??0))return;this.timestamp.set(r),this.#r(r),this.frame.update(a=>(a?.close(),i.clone()))}finally{i.close()}},error:i=>{console.error(i),e.close()}});e.cleanup(()=>{s.state!=="closed"&&s.close()}),this.config.container.kind==="cmaf"?this.#i(e,n,s):this.#n(e,n,s)}#n(e,n,s){const i=new se(n,{format:new Le,latency:this.source.sync.buffer});e.cleanup(()=>i.close()),e.run(o=>{const a=o.get(i.buffered),c=o.get(this.#e);this.buffered.update(()=>_e(a,c))}),s.configure({...this.config,description:this.config.description?V(this.config.description):void 0,optimizeForLatency:this.config.optimizeForLatency??!0,flip:!1});let r;e.spawn(async()=>{for(;;){const o=await Promise.race([i.next(),e.cancel]);if(!o)break;const{frame:a,group:c}=o;if(!a){r&&(r.final=!0);continue}const u=h.fromMicro(a.timestamp);this.source.sync.received(u,"video");const l=new EncodedVideoChunk({type:a.keyframe?"key":"delta",data:a.data,timestamp:a.timestamp});if(this.stats.update(f=>({frameCount:(f?.frameCount??0)+1,bytesReceived:(f?.bytesReceived??0)+a.data.byteLength})),r?.group===c||r?.final&&r.group+1===c){const f=h.fromMicro(r.timestamp),m=h.fromMicro(a.timestamp);this.#s(f,m)}r={timestamp:a.timestamp,group:c,final:!1},s.decode(l)}})}#i(e,n,s){if(this.config.container.kind!=="cmaf")return;const i=we(this.config.container.init),r=Ie(i),o=this.config.description?V(this.config.description):void 0,a=new se(n,{format:new vn(r),latency:this.source.sync.buffer});e.cleanup(()=>a.close()),e.run(u=>{const l=u.get(a.buffered),f=u.get(this.#e);this.buffered.update(()=>_e(l,f))}),s.configure({codec:this.config.codec,description:o,optimizeForLatency:this.config.optimizeForLatency??!0,flip:!1});let c;e.spawn(async()=>{for(;;){const u=await Promise.race([a.next(),e.cancel]);if(!u)break;const{frame:l,group:f}=u;if(!l){c&&(c.final=!0);continue}const m=h.fromMicro(l.timestamp);if(this.source.sync.received(m,"video"),this.stats.update(w=>({frameCount:(w?.frameCount??0)+1,bytesReceived:(w?.bytesReceived??0)+l.data.byteLength})),c?.group===f||c?.final&&c.group+1===f){const w=h.fromMicro(c.timestamp),b=h.fromMicro(l.timestamp);this.#s(w,b)}if(c={timestamp:l.timestamp,group:f,final:!1},s.state==="closed")break;s.decode(new EncodedVideoChunk({type:l.keyframe?"key":"delta",data:l.data,timestamp:l.timestamp}))}})}#s(e,n){e>n||this.#e.mutate(s=>{for(const i of s)if(i.start<=n&&i.end>=e){i.start=h.min(i.start,e),i.end=h.max(i.end,n);return}s.push({start:e,end:n}),s.sort((i,r)=>i.start-r.start)})}#r(e){this.#e.mutate(n=>{for(;n.length>0;){if(n[0].end>=e){n[0].start=h.max(n[0].start,e);break}n.shift()}})}close(){this.signals.close(),this.frame.update(e=>{e?.close()})}}async function Ur(t){const e=t.description?V(t.description):void 0,{supported:n}=await VideoDecoder.isConfigSupported({codec:t.codec,description:e,optimizeForLatency:t.optimizeForLatency??!0});if(n)return!0;if(t.codec.startsWith("avc3.")){const s=`avc1.${t.codec.slice(5)}`;if((await VideoDecoder.isConfigSupported({codec:s,description:e,optimizeForLatency:t.optimizeForLatency??!0})).supported)return t.codec=s,!0}return!1}let Mr=class{muxer;source;#e=new d(void 0);stats=this.#e;#t=new d([]);buffered=this.#t;#n=new d(!1);stalled=this.#n;#i=new d(h.zero);timestamp=this.#i;signals=new _;constructor(t,e){this.muxer=t,this.source=e,this.source.supported.set(Tr),this.signals.run(this.#s.bind(this)),this.signals.run(this.#c.bind(this)),this.signals.run(this.#u.bind(this))}#s(t){const e=t.get(this.muxer.element);if(!e)return;const n=t.get(this.muxer.mediaSource);if(!n)return;const s=t.get(this.source.broadcast);if(!s)return;const i=t.get(s.active);if(!i)return;const r=t.get(this.source.track);if(!r)return;const o=t.get(this.source.config);if(!o)return;const a=`video/mp4; codecs="${o.codec}"`,c=n.addSourceBuffer(a);t.cleanup(()=>{n.removeSourceBuffer(c),c.abort()}),t.event(c,"error",u=>{console.error("[MSE] SourceBuffer error:",u)}),t.event(c,"updateend",()=>{this.#t.set(An(c.buffered))}),o.container.kind==="cmaf"?this.#a(t,i,r,o,c,e):this.#o(t,i,r,o,c,e)}async#r(t,e){for(;t.updating;)await new Promise(n=>t.addEventListener("updateend",n,{once:!0}));for(t.appendBuffer(e);t.updating;)await new Promise(n=>t.addEventListener("updateend",n,{once:!0}))}#a(t,e,n,s,i,r){if(s.container.kind!=="cmaf")throw new Error("unreachable");const o=e.subscribe(n,te.video);t.cleanup(()=>o.close());const a=we(s.container.init),c=Ie(a);t.spawn(async()=>{for(await this.#r(i,a);;){const u=await o.readFrame();if(!u)return;const l=pn(u,c);this.source.sync.received(h.fromMicro(l),"video"),await this.#r(i,u),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0))}})}#o(t,e,n,s,i,r){const o=e.subscribe(n,te.video);t.cleanup(()=>o.close());const a=new se(o,{format:new Le,latency:this.source.sync.buffer});t.cleanup(()=>a.close()),t.spawn(async()=>{const c=tr(s);await this.#r(i,c);let u=1,l,f;for(;;){const m=await a.next();if(!m)return;if(!m.frame)continue;f=m.frame;const w=h.fromMicro(f.timestamp);this.source.sync.received(w,"video");break}for(;;){const m=await a.next();if(m&&!m.frame)continue;const w=m?.frame;if(w){l=Me.sub(w.timestamp,f.timestamp);const k=h.fromMicro(w.timestamp);this.source.sync.received(k,"video")}const b=yn({data:f.data,timestamp:f.timestamp,duration:l??0,keyframe:f.keyframe,sequence:u++});if(await this.#r(i,b),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0)),!w)return;f=w}})}#c(t){const e=t.get(this.muxer.element);if(!e)return;const n=()=>{this.#n.set(e.readyState<=HTMLMediaElement.HAVE_CURRENT_DATA)};n(),t.event(e,"waiting",n),t.event(e,"playing",n),t.event(e,"seeking",n)}#u(t){const e=t.get(this.muxer.element);if(e)if("requestVideoFrameCallback"in e){const n=e;let s;const i=()=>{const r=h.fromSecond(n.currentTime);this.#i.set(r),s=n.requestVideoFrameCallback(i)};s=n.requestVideoFrameCallback(i),t.cleanup(()=>n.cancelVideoFrameCallback(s))}else t.event(e,"timeupdate",()=>{const n=h.fromSecond(e.currentTime);this.#i.set(n)})}close(){this.source.close(),this.signals.close()}};async function Tr(t){return MediaSource.isTypeSupported(`video/mp4; codecs="${t.codec}"`)}class zr{decoder;canvas;paused;frame=new d(void 0);timestamp=new d(void 0);#e=new d(void 0);#t=new d(!1);#n=new _;constructor(e,n){this.decoder=e,this.canvas=d.from(n?.canvas),this.paused=d.from(n?.paused??!1),this.#n.run(s=>{const i=s.get(this.canvas);this.#e.set(i?.getContext("2d")??void 0)}),this.#n.run(this.#s.bind(this)),this.#n.run(this.#r.bind(this)),this.#n.run(this.#a.bind(this)),this.#n.run(this.#i.bind(this))}#i(e){const n=e.getAll([this.canvas,this.decoder.display]);if(!n)return;const[s,i]=n;(s.width!==i.width||s.height!==i.height)&&(s.width=i.width,s.height=i.height)}#s(e){const n=e.get(this.canvas);if(!n){this.#t.set(!1);return}let s=!1;const i=()=>{this.#t.set(s&&!document.hidden)},r=new IntersectionObserver(o=>{for(const a of o)s=a.isIntersecting,i()},{threshold:.01});e.event(document,"visibilitychange",i),r.observe(n),e.cleanup(()=>r.disconnect()),e.cleanup(()=>this.#t.set(!1))}#r(e){const n=e.get(this.paused),s=e.get(this.#t);if(e.cleanup(()=>this.decoder.enabled.set(!1)),!n){this.decoder.enabled.set(s);return}const i=e.get(this.frame);this.decoder.enabled.set(!i)}#a(e){const n=e.get(this.#e);if(!n)return;const s=e.get(this.paused);let i;s||(i=e.get(this.decoder.frame));let r=requestAnimationFrame(()=>{const o=i??this.frame.peek();this.#o(n,o),i&&(this.frame.update(a=>(a?.close(),i.clone())),this.timestamp.set(h.fromMicro(i.timestamp))),r=void 0});e.cleanup(()=>{i?.close(),r&&cancelAnimationFrame(r)})}#o(e,n){if(!n){e.fillStyle="#000",e.fillRect(0,0,e.canvas.width,e.canvas.height);return}e.save(),e.fillStyle="#000",e.fillRect(0,0,e.canvas.width,e.canvas.height),this.decoder.source.catalog.peek()?.flip&&(e.scale(-1,1),e.translate(-e.canvas.width,0)),e.drawImage(n,0,0,e.canvas.width,e.canvas.height),e.restore()}close(){this.frame.update(e=>{e?.close()}),this.#n.close()}}function Cr(t){return e=>{const n=[],s=[];for(const[i,r]of e)if(r.codedWidth&&r.codedHeight){const o=r.codedWidth*r.codedHeight;o<=t?n.push({name:i,size:o}):s.push({name:i,size:o})}return n.sort((i,r)=>r.size-i.size),n.length>0?n.map(i=>i.name):s.length>0?(s.sort((i,r)=>i.size-r.size),[s[0].name]):e.map(([i])=>i)}}function Ir(t){return e=>{const n=[],s=[];for(const[i,r]of e)r.bitrate!=null&&r.bitrate<=t?n.push({name:i,bitrate:r.bitrate}):r.bitrate!=null&&s.push({name:i,bitrate:r.bitrate});return n.sort((i,r)=>r.bitrate-i.bitrate),n.length>0?n.map(i=>i.name):s.length>0?(s.sort((i,r)=>i.bitrate-r.bitrate),[s[0].name]):e.map(([i])=>i)}}function Lr(t){let e=t[0];for(const n of t){const[,s]=n,[,i]=e,r=(s.codedWidth??0)*(s.codedHeight??0),o=(i.codedWidth??0)*(i.codedHeight??0);if(r!==o){r>o&&(e=n);continue}(s.bitrate??0)>(i.bitrate??0)&&(e=n)}return e[0]}let Rr=class{broadcast;target;#e=new d(void 0);catalog=this.#e;#t=new d({});available=this.#t;#n=new d(void 0);track=this.#n;#i=new d(void 0);config=this.#i;sync;supported;#s=new _;constructor(t,e){this.broadcast=d.from(e?.broadcast),this.target=d.from(e?.target),this.sync=t,this.supported=d.from(e?.supported),this.#s.run(this.#r.bind(this)),this.#s.run(this.#a.bind(this)),this.#s.run(this.#o.bind(this))}#r(t){const e=t.get(this.broadcast);if(!e)return;const n=t.get(e.catalog)?.video;n&&t.set(this.#e,n)}#a(t){const e=t.get(this.supported);if(!e)return;const n=t.get(this.#e)?.renditions??{};t.spawn(async()=>{const s={};for(const[i,r]of Object.entries(n))await e(r)&&(s[i]=r);Object.keys(s).length===0&&Object.keys(n).length>0&&console.warn("[Source] No supported video renditions found:",n),this.#t.set(s)})}#o(t){const e=t.get(this.#t);if(Object.keys(e).length===0)return;const n=t.get(this.target);if(n?.name&&n.name in e){const a=e[n.name];t.set(this.#n,n.name),t.set(this.#i,a),t.set(this.sync.video,a.jitter);return}let s=n;if(!n?.bitrate){const a=t.get(this.broadcast),c=(a?t.get(a.connection):void 0)?.recvBandwidth;if(c){const u=t.get(c);if(u!=null){const l=Math.round(u*.8);s={...n,bitrate:l}}}}const i=this.#c(e,s);if(!i)return;const r=e[i];t.set(this.#n,i),t.set(this.#i,r);const o=r.jitter??(r.framerate?Math.ceil(1e3/r.framerate):void 0);t.set(this.sync.video,o)}#c(t,e){const n=Object.entries(t);if(n.length===0)return;if(n.length===1)return n[0][0];const s=[];if(e?.pixels!=null&&s.push(Cr(e.pixels)),e?.bitrate!=null&&s.push(Ir(e.bitrate)),s.length===0)return Lr(n);const i=s.map(o=>o(n)),r=i.map(o=>new Set(o));for(const o of i[0])if(r.every(a=>a.has(o)))return o;console.warn("conflicting rendition filters, no rendition satisfies all criteria")}close(){this.#s.close()}};function An(t){const e=[];for(let n=0;n<t.length;n++){const s=h.fromSecond(t.start(n)),i=h.fromSecond(t.end(n));e.push({start:s,end:i})}return e}class Br{source;stats=new d(void 0);stalled=new d(!1);buffered=new d([]);timestamp=new d(h.zero);constructor(e){this.source=e}}class Or{source;volume=new d(.5);muted=new d(!1);stats=new d(void 0);buffered=new d([]);context=new d(void 0);constructor(e){this.source=e}}class $r{element=new d(void 0);broadcast;latency;jitter;paused;video;#e;audio;#t;sync;signals=new _;constructor(e){this.element=d.from(e?.element),this.broadcast=d.from(e?.broadcast),this.sync=new nt({latency:e?.latency,connection:e?.connection}),this.latency=this.sync.latency,this.jitter=this.sync.jitter,this.#e=new Rr(this.sync,{broadcast:this.broadcast}),this.#t=new jr(this.sync,{broadcast:this.broadcast}),this.video=new Br(this.#e),this.audio=new Or(this.#t),this.paused=d.from(e?.paused??!1),this.signals.run(this.#n.bind(this))}#n(e){const n=e.get(this.element);n&&(n instanceof HTMLCanvasElement?this.#i(e,n):n instanceof HTMLVideoElement&&this.#s(e,n))}#i(e,n){const s=new _r(this.#e),i=new br(this.#t),r=new vr(i,{volume:this.audio.volume,muted:this.audio.muted,paused:this.paused}),o=new zr(s,{canvas:n,paused:this.paused});e.cleanup(()=>{s.close(),i.close(),r.close(),o.close()}),e.proxy(this.video.stats,s.stats),e.proxy(this.video.buffered,s.buffered),e.proxy(this.video.stalled,s.stalled),e.proxy(this.video.timestamp,s.timestamp),e.proxy(this.audio.stats,i.stats),e.proxy(this.audio.buffered,i.buffered),e.proxy(this.audio.context,i.context)}#s(e,n){const s=new kr(this.sync,{paused:this.paused,element:n}),i=new Mr(s,this.#e),r=new Fr(s,this.#t,{volume:this.audio.volume,muted:this.audio.muted});e.cleanup(()=>{i.close(),r.close(),s.close()}),e.proxy(this.video.stats,i.stats),e.proxy(this.video.buffered,i.buffered),e.proxy(this.video.stalled,i.stalled),e.proxy(this.video.timestamp,i.timestamp),e.proxy(this.audio.stats,r.stats),e.proxy(this.audio.buffered,r.buffered),e.proxy(this.audio.context,r.context)}close(){this.signals.close()}}class Fr{muxer;source;volume;muted;#e=new d(void 0);stats=this.#e;#t=new d([]);buffered=this.#t;context=new d(void 0);#n=new _;constructor(e,n,s){this.muxer=e,this.source=n,this.source.supported.set(Dr),this.volume=d.from(s?.volume??.5),this.muted=d.from(s?.muted??!1),this.#n.run(this.#i.bind(this)),this.#n.run(this.#o.bind(this))}#i(e){const n=e.get(this.muxer.element);if(!n)return;const s=e.get(this.muxer.mediaSource);if(!s)return;const i=e.get(this.source.broadcast);if(!i)return;const r=e.get(i.active);if(!r)return;const o=e.get(this.source.track);if(!o)return;const a=e.get(this.source.config);if(!a)return;const c=`audio/mp4; codecs="${a.codec}"`,u=s.addSourceBuffer(c);e.cleanup(()=>{s.removeSourceBuffer(u),u.abort()}),e.event(u,"error",f=>{console.error("[MSE] SourceBuffer error:",f)}),e.event(u,"updateend",()=>{this.#t.set(An(u.buffered))});const l=r.subscribe(o,te.audio);e.cleanup(()=>l.close()),a.container.kind==="cmaf"?this.#r(e,l,a,u,n):this.#a(e,l,a,u,n)}async#s(e,n){for(;e.updating;)await new Promise(s=>e.addEventListener("updateend",s,{once:!0}));for(e.appendBuffer(n);e.updating;)await new Promise(s=>e.addEventListener("updateend",s,{once:!0}))}#r(e,n,s,i,r){if(s.container.kind!=="cmaf")throw new Error("unreachable");const o=we(s.container.init),a=Ie(o);e.spawn(async()=>{for(await this.#s(i,o);;){const c=await n.readFrame();if(!c)return;const u=pn(c,a);this.source.sync.received(h.fromMicro(u),"audio"),await this.#s(i,c),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0))}})}#a(e,n,s,i,r){const o=new se(n,{format:new Le,latency:this.source.sync.buffer});e.cleanup(()=>o.close()),e.spawn(async()=>{const a=nr(s);await this.#s(i,a);let c=1,u,l;for(;;){const f=await o.next();if(!f)return;if(!f.frame)continue;l=f.frame;const m=h.fromMicro(l.timestamp);this.source.sync.received(m,"audio");break}for(;;){const f=await o.next();if(f&&!f.frame)continue;const m=f?.frame;if(m){u=Me.sub(m.timestamp,l.timestamp);const b=h.fromMicro(m.timestamp);this.source.sync.received(b,"audio")}const w=yn({data:l.data,timestamp:l.timestamp,duration:u??0,keyframe:l.keyframe,sequence:c++});if(await this.#s(i,w),r.buffered.length>0&&r.currentTime<r.buffered.start(0)&&(r.currentTime=r.buffered.start(0)),!m)return;l=m}})}#o(e){const n=e.get(this.muxer.element);if(!n)return;const s=e.get(this.volume),i=e.get(this.muted);i&&!n.muted?n.muted=!0:!i&&n.muted&&(n.muted=!1),s!==n.volume&&(n.volume=s),e.event(n,"volumechange",()=>{this.volume.set(n.volume)})}close(){this.#n.close()}}async function Dr(t){return MediaSource.isTypeSupported(`audio/mp4; codecs="${t.codec}"`)}const Nr=128;class jr{broadcast;target;#e=new d(void 0);catalog=this.#e;#t=new d({});available=this.#t;#n=new d(void 0);track=this.#n;#i=new d(void 0);config=this.#i;supported;sync;#s=new _;constructor(e,n){this.sync=e,this.broadcast=d.from(n?.broadcast),this.target=d.from(n?.target),this.supported=d.from(n?.supported),this.#s.run(this.#r.bind(this)),this.#s.run(this.#a.bind(this)),this.#s.run(this.#o.bind(this))}#r(e){const n=e.get(this.broadcast);if(!n)return;const s=e.get(n.catalog)?.audio;s&&e.set(this.#e,s)}#a(e){const n=e.get(this.#e)?.renditions??{},s=e.get(this.supported);s&&e.spawn(async()=>{const i={};for(const[r,o]of Object.entries(n))await s(o)&&(i[r]=o);Object.keys(i).length===0&&Object.keys(n).length>0&&console.warn("no supported audio renditions found:",n),this.#t.set(i)})}#o(e){const n=e.get(this.#t);if(Object.keys(n).length===0)return;const s=e.get(this.target);let i;if(s?.name&&s.name in n)i={track:s.name,config:n[s.name]};else if(i=this.#c(n),!i)return;e.set(this.#n,i.track),e.set(this.#i,i.config);const r=i.config.jitter??Pr(i.config)??0,o=Math.ceil(Nr/i.config.sampleRate*1e3),a=r+o;e.set(this.sync.audio,a)}#c(e){const n=Object.entries(e);if(n.length!==0){for(const[s,i]of n)if(i.container.kind==="legacy")return{track:s,config:i};for(const[s,i]of n)if(i.container.kind==="cmaf")return{track:s,config:i}}}close(){this.#s.close()}}function Pr(t){if(t.codec.startsWith("opus"))return 20;if(t.codec.startsWith("mp4a"))return Math.ceil(1024/t.sampleRate*1e3)}const Vr=Ce([rn(["loc","cmaf","legacy","mediatimeline","eventtimeline"]),v()]),Wr=Ce([rn(["video","audio","audiodescription","caption","subtitle","signlanguage"]),v()]),Zr=A({name:v(),packaging:Vr,isLive:Xe(),role:p(Wr),codec:p(v()),width:p(M()),height:p(M()),framerate:p(M()),samplerate:p(M()),channelConfig:p(v()),bitrate:p(M()),initData:p(v()),renderGroup:p(M()),altGroup:p(M()),jitter:p(M())}),qr=A({version:Ve(1),tracks:H(Zr)});function Hr(t){const e=new TextDecoder().decode(t);try{const n=JSON.parse(e);return qr.parse(n)}catch(n){throw console.warn("invalid MSF catalog",e),n}}async function Gr(t){const e=await t.readFrame();if(e)return Hr(e)}const Xr=48e3,Ut=2;function Yr(t){let e="";for(let n=0;n<t.length;n++)e+=t[n].toString(16).padStart(2,"0");return e}function xn(t){let e;try{e=t.initData?we(t.initData):void 0}catch{e=void 0}return t.packaging==="cmaf"&&t.initData&&e?{container:{kind:"cmaf",init:t.initData},description:void 0}:{container:{kind:"legacy"},description:e?Yr(e):void 0}}function Kr(t){if(!t.codec)return;const{container:e,description:n}=xn(t);return{codec:t.codec,container:e,description:n,codedWidth:t.width!=null?Z(t.width):void 0,codedHeight:t.height!=null?Z(t.height):void 0,framerate:t.framerate,bitrate:t.bitrate!=null?Z(t.bitrate):void 0,jitter:t.jitter!=null?Z(t.jitter):void 0}}function Jr(t){if(!t.codec)return;const e=(()=>{if(!t.channelConfig)return Ut;const i=Number.parseInt(t.channelConfig,10);return Number.isFinite(i)?i:Ut})(),{container:n,description:s}=xn(t);return{codec:t.codec,container:n,description:s,sampleRate:Z(t.samplerate??Xr),numberOfChannels:Z(e),bitrate:t.bitrate!=null?Z(t.bitrate):void 0,jitter:t.jitter!=null?Z(t.jitter):void 0}}function Qr(t){const e={},n={};for(const i of t.tracks)if(i.role==="video"){const r=Kr(i);r&&(e[i.name]=r)}else if(i.role==="audio"){const r=Jr(i);r&&(n[i.name]=r)}const s={};return Object.keys(e).length>0&&(s.video={renditions:e}),Object.keys(n).length>0&&(s.audio={renditions:n}),s}const eo=["hang","msf","manual"];class to{connection;enabled;name;status=new d("offline");reload;catalogFormat;#e=new d(void 0);active=this.#e;catalog;#t;#n=new d(!1);signals=new _;constructor(e){this.connection=d.from(e?.connection),this.name=d.from(e?.name??zn()),this.enabled=d.from(e?.enabled??!1),this.reload=d.from(e?.reload??!1),this.catalogFormat=d.from(e?.catalogFormat??"hang"),this.catalog=d.from(e?.catalog),this.#t=e?.announced??new d(new Set),this.signals.run(this.#i.bind(this)),this.signals.run(this.#s.bind(this)),this.signals.run(this.#r.bind(this))}#i(e){if(!e.get(this.reload)){this.#n.set(!0);return}if(e.get(this.connection)?.url.hostname.endsWith("mediaoverquic.com")){console.warn("Cloudflare relay does not support broadcast discovery yet; ignoring reload signal."),this.#n.set(!0);return}const n=e.get(this.name),s=e.get(this.#t);this.#n.set(s.has(n))}#s(e){if(!e.get(this.enabled)||!e.get(this.#n))return;const n=e.get(this.connection);if(!n)return;const s=e.get(this.name),i=n.consume(s);e.cleanup(()=>i.close()),e.set(this.#e,i,void 0)}#r(e){if(!e.get(this.enabled))return;const n=e.get(this.catalogFormat);if(n==="manual"){const a=e.get(this.catalog);this.status.set(a?"live":"loading");return}const s=e.get(this.active);if(!s)return;this.status.set("loading");const i=n==="hang"?"catalog.json":"catalog",r=s.subscribe(i,te.catalog);e.cleanup(()=>r.close());const o=n==="hang"?async()=>ri(r):async()=>{const a=await Gr(r);return a?Qr(a):void 0};e.spawn(async()=>{try{for(;;){const a=await Promise.race([e.cancel,o()]);if(!a)break;console.debug("received catalog",n,this.name.peek(),a),this.catalog.set(a),this.status.set("live")}}catch(a){console.warn("error fetching catalog",this.name.peek(),a)}finally{this.catalog.set(void 0),this.status.set("offline")}})}close(){this.signals.close()}}const no="hang";function so(t){return eo.find(e=>e===t)??no}const io=["url","name","paused","volume","muted","reload","latency","jitter","catalog-format"],ro=new FinalizationRegistry(t=>t.close());class oo extends HTMLElement{static observedAttributes=io;connection;broadcast;backend;#e=new d(!1);signals=new _;constructor(){super(),ro.register(this,this.signals),this.connection=new In({enabled:this.#e}),this.signals.cleanup(()=>this.connection.close()),this.broadcast=new to({connection:this.connection.established,announced:this.connection.announced,enabled:this.#e}),this.signals.cleanup(()=>this.broadcast.close()),this.backend=new $r({broadcast:this.broadcast,connection:this.connection.established}),this.signals.cleanup(()=>this.backend.close());const e=()=>{const s=this.querySelector("canvas"),i=this.querySelector("video");if(s&&i)throw new Error("Cannot have both canvas and video elements");this.backend.element.set(s??i)},n=new MutationObserver(e);n.observe(this,{childList:!0,subtree:!0}),this.signals.cleanup(()=>n.disconnect()),e(),this.signals.run(s=>{const i=s.get(this.connection.url);i?this.setAttribute("url",i.toString()):this.removeAttribute("url")}),this.signals.run(s=>{const i=s.get(this.broadcast.name);this.setAttribute("name",i.toString())}),this.signals.run(s=>{s.get(this.backend.audio.muted)?this.setAttribute("muted",""):this.removeAttribute("muted")}),this.signals.run(s=>{s.get(this.backend.paused)?this.setAttribute("paused","true"):this.removeAttribute("paused")}),this.signals.run(s=>{const i=s.get(this.backend.audio.volume);this.setAttribute("volume",i.toString())}),this.signals.run(s=>{if(s.get(this.backend.latency)==="real-time")this.setAttribute("latency","real-time");else{const i=Math.floor(s.get(this.backend.jitter));this.setAttribute("latency",i.toString())}})}connectedCallback(){this.#e.set(!0),this.style.display="block",this.style.position="relative"}disconnectedCallback(){this.#e.set(!1)}#t(e){const n=e?Number.parseFloat(e):Number.NaN;this.backend.latency.set(Number.isFinite(n)?n:100)}attributeChangedCallback(e,n,s){if(n!==s)if(e==="url")this.connection.url.set(s?new URL(s):void 0);else if(e==="name")this.broadcast.name.set(at(s??""));else if(e==="paused")this.backend.paused.set(s!==null);else if(e==="volume"){const i=s?Number.parseFloat(s):.5;this.backend.audio.volume.set(i)}else if(e==="muted")this.backend.audio.muted.set(s!==null);else if(e==="reload")this.broadcast.reload.set(s!==null);else if(e==="latency")!s||s==="real-time"?this.backend.latency.set("real-time"):this.#t(s);else if(e==="jitter")this.#t(s);else if(e==="catalog-format")this.broadcast.catalogFormat.set(so(s));else{const i=e;throw new Error(`Invalid attribute: ${i}`)}}get url(){return this.connection.url.peek()}set url(e){this.connection.url.set(e?new URL(e):void 0)}get name(){return this.broadcast.name.peek()}set name(e){this.broadcast.name.set(at(e))}get paused(){return this.backend.paused.peek()}set paused(e){this.backend.paused.set(e)}get volume(){return this.backend.audio.volume.peek()}set volume(e){this.backend.audio.volume.set(e)}get muted(){return this.backend.audio.muted.peek()}set muted(e){this.backend.audio.muted.set(e)}get reload(){return this.broadcast.reload.peek()}set reload(e){this.broadcast.reload.set(e)}get latency(){return this.backend.latency.peek()}set latency(e){this.backend.latency.set(e)}get jitter(){return this.backend.jitter.peek()}set jitter(e){this.backend.latency.set(e)}get catalogFormat(){return this.broadcast.catalogFormat.peek()}set catalogFormat(e){this.broadcast.catalogFormat.set(e)}get catalog(){return this.broadcast.catalog.peek()}set catalog(e){this.broadcast.catalog.set(e)}}customElements.define("moq-watch",oo);const Sn=navigator.userAgent.toLowerCase().includes("firefox"),We={aac:"mp4a.40.2",opus:"opus",av1:"av01.0.08M.08",h264:"avc1.640028",h265:"hev1.1.6.L93.B0",vp9:"vp09.00.10.08",vp8:"vp8"};async function Mt(t){return globalThis.AudioDecoder?(await AudioDecoder.isConfigSupported({codec:We[t],numberOfChannels:2,sampleRate:48e3})).supported===!0:!1}async function ce(t){const e=await VideoDecoder.isConfigSupported({codec:We[t],hardwareAcceleration:"prefer-software"}),n=await VideoDecoder.isConfigSupported({codec:We[t],hardwareAcceleration:"prefer-hardware"});return{hardware:Sn||n.config?.hardwareAcceleration!=="prefer-hardware"?void 0:n.supported===!0,software:e.supported===!0}}async function ao(){return{webtransport:typeof WebTransport<"u"?Sn?"partial":"full":"partial",audio:{decoding:{aac:await Mt("aac"),opus:await Mt("opus")?"full":"partial"},render:typeof AudioContext<"u"&&typeof AudioBufferSourceNode<"u"},video:{decoding:typeof VideoDecoder<"u"?{h264:await ce("h264"),h265:await ce("h265"),vp8:await ce("vp8"),vp9:await ce("vp9"),av1:await ce("av1")}:void 0,render:typeof OffscreenCanvas<"u"&&typeof CanvasRenderingContext2D<"u"}}}function $(t,e,...n){const s=document.createElement(t);if(!e)return s;const{style:i,classList:r,dataset:o,attributes:a,...c}=e;return i&&Object.assign(s.style,i),r&&s.classList.add(...r),o&&Object.entries(o).forEach(([u,l])=>{s.dataset[u]=l}),a&&Object.entries(a).forEach(([u,l])=>{s.setAttribute(u,l)}),n&&n.forEach(u=>{typeof u=="string"?s.appendChild(document.createTextNode(u)):s.appendChild(u)}),Object.assign(s,c),s}const Tt=navigator.userAgent.toLowerCase().includes("firefox"),co=["show","details"];class uo extends HTMLElement{#e=new d("warning");#t=new d(!1);#n=new d(void 0);#i=new d(!1);#s;static observedAttributes=co;constructor(){super(),ao().then(e=>this.#n.set(e)).catch(e=>console.error("Failed to detect watch support:",e))}connectedCallback(){this.#s=new _,this.#s.run(this.#a.bind(this))}disconnectedCallback(){this.#s?.close(),this.#s=void 0}attributeChangedCallback(e,n,s){if(e==="show"){const i=s??"warning";if(i==="always"||i==="warning"||i==="error"||i==="never")this.show=i;else throw new Error(`Invalid show: ${i}`)}else if(e==="details")this.details=s!==null;else{const i=e;throw new Error(`Invalid attribute: ${i}`)}}get show(){return this.#e.peek()}set show(e){this.#e.set(e)}get details(){return this.#t.peek()}set details(e){this.#t.set(e)}#r(e){return e.webtransport==="none"||!e.audio.decoding||!e.video.decoding||!e.audio.render||!e.video.render||!Object.values(e.audio.decoding).some(n=>n===!0||n==="full"||n==="partial")||!Object.values(e.video.decoding).some(n=>n.software||n.hardware)?"none":!Object.values(e.audio.decoding).every(n=>n===!0||n==="full")||!Object.values(e.video.decoding).every(n=>n.software||n.hardware)?"partial":"full"}#a(e){const n=e.get(this.#n);if(!n||e.get(this.#i))return;const s=e.get(this.#e);if(s==="never")return;const i=this.#r(n);if(s==="warning"&&i==="full"||s==="error"&&i!=="none")return;const r=$("div",{style:{margin:"0 auto",maxWidth:"28rem",padding:"1rem"}});this.appendChild(r),e.cleanup(()=>this.removeChild(r)),this.#o(r,i,e),e.get(this.#t)&&this.#c(r,n,e)}#o(e,n,s){const i=$("div",{style:{display:"flex",flexDirection:"row",gap:"1rem",flexWrap:"wrap",justifyContent:"space-between",alignItems:"center"}}),r=$("div",{style:{fontWeight:"bold"}});n==="full"?r.textContent="🟢 Full Browser Support":n==="partial"?r.textContent="🟡 Partial Browser Support":n==="none"&&(r.textContent="🔴 No Browser Support");const o=$("button",{type:"button",style:{fontSize:"14px"}});s.event(o,"click",()=>{this.#t.update(c=>!c)}),s.run(c=>{o.textContent=c.get(this.#t)?"Details ➖":"Details ➕"});const a=$("button",{type:"button",style:{fontSize:"14px"}},"Close ❌");s.event(a,"click",()=>{this.#i.set(!0)}),i.appendChild(r),i.appendChild(o),i.appendChild(a),e.appendChild(i),s.cleanup(()=>e.removeChild(i))}#c(e,n,s){const i=$("div",{style:{display:"grid",gridTemplateColumns:"1fr 1fr 1fr",columnGap:"0.5rem",rowGap:"0.2rem",backgroundColor:"rgba(0, 0, 0, 0.6)",borderRadius:"0.5rem",padding:"1rem",fontSize:"0.875rem"}}),r=u=>u?"🟢 Yes":"🔴 No",o=u=>u?.hardware?"🟢 Hardware":u?.software?`🟡 Software${Tt?"*":""}`:"🔴 No",a=u=>u==="full"?"🟢 Full":u==="partial"?"🟡 Polyfill":"🔴 None",c=(u,l,f)=>{const m=$("div",{style:{gridColumnStart:"1",fontWeight:"bold",textAlign:"right"}},u),w=$("div",{style:{gridColumnStart:"2",textAlign:"center"}},l),b=$("div",{style:{gridColumnStart:"3"}},f);i.appendChild(m),i.appendChild(w),i.appendChild(b)};if(c("WebTransport","",a(n.webtransport)),c("Rendering","Audio",r(n.audio.render)),c("","Video",r(n.video.render)),c("Decoding","Opus",a(n.audio.decoding.opus)),c("","AAC",r(n.audio.decoding.aac)),c("","AV1",o(n.video.decoding?.av1)),c("","H.265",o(n.video.decoding?.h265)),c("","H.264",o(n.video.decoding?.h264)),c("","VP9",o(n.video.decoding?.vp9)),c("","VP8",o(n.video.decoding?.vp8)),Tt){const u=$("div",{style:{gridColumnStart:"1",gridColumnEnd:"4",textAlign:"center",fontSize:"0.875rem",fontStyle:"italic"}},"Hardware acceleration is ",$("a",{href:"https://github.com/w3c/webcodecs/issues/896"},"undetectable")," on Firefox.");i.appendChild(u)}e.appendChild(i),s.cleanup(()=>e.removeChild(i))}}customElements.define("moq-watch-support",uo);const U=document.getElementById("player"),st=document.getElementById("errors"),Fe=document.getElementById("quality-panel"),zt=document.getElementById("quality-buttons"),lo=document.getElementById("url-rows"),_n=document.getElementById("info-name"),Ct=document.getElementById("info-url"),It=document.getElementById("info-mode"),le=document.getElementById("info-status"),fo=document.getElementById("stream-control"),it=document.getElementById("stream-input"),ho=document.querySelectorAll("input[name=mode]"),me=document.getElementById("volume-slider"),Ze=document.getElementById("volume-icon"),Lt=document.getElementById("volume-label"),Ee=document.getElementById("jitter-slider"),qe=document.getElementById("jitter-label"),ve=document.getElementById("dbg-fps"),Rt=document.getElementById("dbg-buf"),Bt=document.getElementById("dbg-stalled"),mo=document.getElementById("dbg-latency"),po=document.getElementById("dbg-jitter"),Ot=document.getElementById("dbg-vbytes"),go=document.getElementById("dbg-codec");function G(t,e){return`<span class="badge badge-${e}">${t}</span>`}function wo(t){const e=document.createElement("div");e.className="banner banner-error",e.innerHTML=`⛔ ${t}`,st.appendChild(e)}const ee=new URLSearchParams(window.location.search),ge=ee.get("name")??ee.get("broadcast")??ee.get("id")??null,bo=ee.get("url")??ee.get("host")??null,yo=ee.get("fallbackPlayer")==="true";it.value=ge??"";_n.textContent=ge??"—";const rt=typeof VideoDecoder<"u";if(!rt){const t=document.createElement("div");t.className="banner banner-warn",t.innerHTML="⚠️ <strong>WebCodecs is not supported in this browser.</strong> Falling back to MSE player. Chrome 94+, Edge 94+, or Safari 17.4+ required for WebCodecs.",st.appendChild(t)}const De=100,Ne=200;function En(t){const e=t==="webcodecs"||t==="auto"&&rt,n=U.querySelector("canvas, video");if(n&&n.remove(),e){const s=document.createElement("canvas");U.appendChild(s),U.setAttribute("jitter",De),Ee.value=De,qe.textContent=`${De} ms`,It.innerHTML=G("WebCodecs (canvas)","blue")}else{const s=document.createElement("video");s.autoplay=!0,s.muted=!0,U.appendChild(s),U.setAttribute("jitter",Ne),Ee.value=Ne,qe.textContent=`${Ne} ms`,It.innerHTML=G("MSE (video)","yellow")}}const Un=yo?"mse":"auto";En(Un);document.querySelector(`input[name=mode][value="${Un}"]`).checked=!0;ho.forEach(t=>{t.addEventListener("change",()=>En(t.value))});function He(t){t===0?(U.setAttribute("muted",""),Ze.textContent="🔇",Lt.textContent="muted"):(U.removeAttribute("muted"),U.setAttribute("volume",(t/100).toFixed(2)),Ze.textContent=t<50?"🔉":"🔊",Lt.textContent=`${t}%`)}me.addEventListener("input",()=>He(Number(me.value)));Ze.addEventListener("click",()=>{if(U.hasAttribute("muted")){const t=Number(me.value)||50;me.value=t,He(t)}else me.value=0,He(0)});Ee.addEventListener("input",()=>{const t=Number(Ee.value);qe.textContent=`${t} ms`,U.setAttribute("jitter",t)});let $t=0;function vo(t){return t==null?"—":t<1024?`${t} B`:t<1024*1024?`${(t/1024).toFixed(1)} KB`:`${(t/(1024*1024)).toFixed(2)} MB`}setInterval(()=>{const t=U.backend;if(!t)return;const e=t.video.stats.peek();if(e){const a=e.frameCount-$t;$t=e.frameCount,ve.textContent=`${a} fps`,ve.className="dbg-value ok",Ot.textContent=vo(e.bytesReceived)}else ve.textContent="N/A (MSE)",ve.className="dbg-value",Ot.textContent="—";const n=t.video.buffered.peek();if(n&&n.length>0){const a=n[n.length-1],c=a.end-a.start;Rt.textContent=`${c.toFixed(0)} ms`}else Rt.textContent="empty";const s=t.video.stalled.peek();Bt.textContent=s?"yes":"no",Bt.className=`dbg-value ${s?"stalled":"ok"}`;const i=t.video.source.sync.latency.peek();mo.textContent=i!=null?`${i.toFixed(0)} ms`:"—";const r=t.jitter.peek();po.textContent=r!=null?`${r} ms`:"—";const o=t.video.source.config.peek();go.textContent=o?.codec??"—"},1e3);if(!rt){const t=document.querySelector("input[value=webcodecs]");t&&(t.disabled=!0,t.closest("label").style.opacity="0.4",t.closest("label").title="WebCodecs not supported in this browser")}function ko(t){return t.includes(":")?`[${t}]`:t}function Ao(t){if(!t)return{};if(t.includes("://"))try{const e=new URL(t);return{hostname:e.hostname,port:e.port||null,fullUrl:t}}catch{}try{const e=new URL(`https://${t}/`);return{hostname:e.hostname,port:e.port||null}}catch{return{}}}function xo(t){const e=Ao(t),n=e.hostname||window.location.hostname||"localhost",s=ko(n),i=window.location.protocol==="https:",r=[];if(e.fullUrl)try{const o=new URL(e.fullUrl);(!o.pathname||o.pathname==="/")&&(o.pathname="/moq"),r.push(o.toString().replace(/\/$/,""))}catch{}else e.port&&r.push(`${i?"https:":"http:"}//${s}:${e.port}/moq`);return i?(r.push(`https://${s}/moq`),r.push(`https://${s}:4443/moq`)):(r.push(`http://${s}/moq`),r.push(`http://${s}:4443/moq`),r.push(`https://${s}/moq`),r.push(`https://${s}:4443/moq`)),[...new Set(r)]}const de=xo(bo);if(typeof WebTransport>"u"){const t=document.createElement("div");if(t.className="banner banner-warn",window.location.protocol!=="https:"){const e=window.location.href.replace(/^http:/,"https:");t.innerHTML=`⚠️ <strong>WebTransport unavailable on this page.</strong> Playback will fall back to WebSocket (higher latency). <a href="${e}" style="color:inherit;font-weight:bold;">Switch to HTTPS →</a> for low-latency.`}else t.innerHTML="⚠️ <strong>WebTransport not supported in this browser.</strong> Falling back to WebSocket — try Chrome/Edge 97+ or Firefox 114+ for low-latency.";st.appendChild(t)}de.forEach((t,e)=>{const n=document.createElement("tr");n.id=`url-row-${e}`,n.innerHTML=`
+    <td>${e+1}</td>
+    <td>${t}</td>
+    <td id="url-status-${e}"><span class="status-dot dot-pending"></span>pending</td>
+  `,lo.appendChild(n)});function ke(t,e,n){const s=document.getElementById(`url-status-${t}`),i={pending:"dot-pending",trying:"dot-trying",ok:"dot-ok",fail:"dot-fail"}[e];s.innerHTML=`<span class="status-dot ${i}"></span>${n}`;const r=document.getElementById(`url-row-${t}`);r.className=e==="ok"?"url-row active":"url-row"}async function So(t,e=4e3){const n=new URL(t);if(n.protocol==="http:"){const o=new URL(n);o.pathname="/certificate.sha256",o.search="";const a=new AbortController,c=setTimeout(()=>a.abort(),e);try{return(await fetch(o.toString(),{signal:a.signal})).ok}catch{return!1}finally{clearTimeout(c)}}if(typeof WebTransport>"u")return!1;let s;try{s=new WebTransport(t)}catch{return!1}let i;const r=new Promise(o=>{i=setTimeout(()=>o(!1),e)});try{return await Promise.race([s.ready.then(()=>!0,()=>!1),r])}finally{clearTimeout(i);try{s.close()}catch{}}}async function _o(){for(let t=0;t<de.length;t++){const e=de[t];if(ke(t,"trying","trying…"),le.innerHTML=G(`probing ${t+1}/${de.length}`,"yellow"),await So(e)){ke(t,"ok","✓ reachable");for(let s=t+1;s<de.length;s++)ke(s,"pending","skipped");return e}else ke(t,"fail","✗ unreachable")}return null}let Ge=ge??"",K=new Set,Mn="";function Ue(){if(zt.innerHTML="",!Ge){Fe.style.display="none";return}const t=[...K].filter(e=>e.split("/").includes(Ge));if(t.length===0){Fe.style.display="none";return}Fe.style.display="flex";for(const e of t){const n=e===Mn,s=document.createElement("button");s.type="button",s.textContent=e.split("/").pop()??e,s.title=e,s.className="quality-btn"+(n?" active":""),s.addEventListener("click",()=>{U.name=e}),zt.appendChild(s)}}let Ft=null;U.connection.established.subscribe(async t=>{if(K=new Set,Ue(),!t)return;Ft=t;const e=t.announced();try{for(;;){const n=await e.next();if(!n||Ft!==t)break;const s=n.path.toString();n.active?K=new Set([...K,s]):K=new Set([...K].filter(i=>i!==s)),Ue()}}catch{}});U.broadcast.name.subscribe(t=>{Mn=t.toString(),Ue()});function Eo(t){if(!t)return;Ge=t,U.setAttribute("name",t),_n.textContent=t,Ue();const e=new URL(window.location);e.searchParams.set("name",t),window.history.replaceState({},"",e)}fo.addEventListener("submit",t=>{t.preventDefault(),Eo(it.value.trim())});(async()=>{const t=await _o();if(!t){wo("Could not reach any relay URL. Check that the relay server is running and the host is correct."),Ct.textContent="none",le.innerHTML=G("unreachable","red");return}Ct.textContent=t,le.innerHTML=G("connecting…","yellow"),U.setAttribute("url",t),ge?(U.setAttribute("name",ge),le.innerHTML=G("connected","green")):(le.innerHTML=G("enter stream id","yellow"),it.focus())})();
