--- MoQPlugin/src/main/resources/moq-ams-player-build/assets/publish-D1fsc_dq.js
@@ -1,133 +1,0 @@
-import{E as g,S as u,e as Ze,B as qe,b as re,c as Xe,_ as pe,M as K,R as Ge,f as fe}from"./time-Cy1-YUvq.js";function m(s,e,t){function i(r,l){if(r._zod||Object.defineProperty(r,"_zod",{value:{def:l,constr:a,traits:new Set},enumerable:!1}),r._zod.traits.has(s))return;r._zod.traits.add(s),e(r,l);const c=a.prototype,d=Object.keys(c);for(let h=0;h<d.length;h++){const p=d[h];p in r||(r[p]=c[p].bind(r))}}const n=t?.Parent??Object;class o extends n{}Object.defineProperty(o,"name",{value:s});function a(r){var l;const c=t?.Parent?new o:this;i(c,r),(l=c._zod).deferred??(l.deferred=[]);for(const d of c._zod.deferred)d();return c}return Object.defineProperty(a,"init",{value:i}),Object.defineProperty(a,Symbol.hasInstance,{value:r=>t?.Parent&&r instanceof t.Parent?!0:r?._zod?.traits?.has(s)}),Object.defineProperty(a,"name",{value:s}),a}class $ extends Error{constructor(){super("Encountered Promise during synchronous parse. Use .parseAsync() instead.")}}const Ye={};function U(s){return Ye}function Je(s,e){return typeof e=="bigint"?e.toString():e}function Qe(s){return{get value(){{const e=s();return Object.defineProperty(this,"value",{value:e}),e}}}}function et(s){const e=s.startsWith("^")?1:0,t=s.endsWith("$")?s.length-1:s.length;return s.slice(e,t)}const me=Symbol("evaluating");function L(s,e,t){let i;Object.defineProperty(s,e,{get(){if(i!==me)return i===void 0&&(i=me,i=t()),i},set(n){Object.defineProperty(s,e,{value:n})},configurable:!0})}const Me="captureStackTrace"in Error?Error.captureStackTrace:(...s)=>{};function se(s){return typeof s=="object"&&s!==null&&!Array.isArray(s)}function tt(s){if(se(s)===!1)return!1;const e=s.constructor;if(e===void 0||typeof e!="function")return!0;const t=e.prototype;return!(se(t)===!1||Object.prototype.hasOwnProperty.call(t,"isPrototypeOf")===!1)}function st(s,e,t){const i=new s._zod.constr(e??s._zod.def);return(!e||t?.parent)&&(i._zod.parent=s),i}function T(s){return{}}function it(s){return Object.keys(s).filter(e=>s[e]._zod.optin==="optional"&&s[e]._zod.optout==="optional")}const nt={safeint:[Number.MIN_SAFE_INTEGER,Number.MAX_SAFE_INTEGER],int32:[-2147483648,2147483647],uint32:[0,4294967295],float32:[-34028234663852886e22,34028234663852886e22],float64:[-Number.MAX_VALUE,Number.MAX_VALUE]};function B(s,e=0){if(s.aborted===!0)return!0;for(let t=e;t<s.issues.length;t++)if(s.issues[t]?.continue!==!0)return!0;return!1}function O(s,e){return e.map(t=>{var i;return(i=t).path??(i.path=[]),t.path.unshift(s),t})}function H(s){return typeof s=="string"?s:s?.message}function W(s,e,t){const i={...s,path:s.path??[]};if(!s.message){const n=H(s.inst?._zod.def?.error?.(s))??H(e?.error?.(s))??H(t.customError?.(s))??H(t.localeError?.(s))??"Invalid input";i.message=n}return delete i.inst,delete i.continue,e?.reportInput||delete i.input,i}const ze=(s,e)=>{s.name="$ZodError",Object.defineProperty(s,"_zod",{value:s._zod,enumerable:!1}),Object.defineProperty(s,"issues",{value:e,enumerable:!1}),s.message=JSON.stringify(e,Je,2),Object.defineProperty(s,"toString",{value:()=>s.message,enumerable:!1})},ot=m("$ZodError",ze),q=m("$ZodError",ze,{Parent:Error}),at=s=>(e,t,i,n)=>{const o=i?Object.assign(i,{async:!1}):{async:!1},a=e._zod.run({value:t,issues:[]},o);if(a instanceof Promise)throw new $;if(a.issues.length){const r=new(n?.Err??s)(a.issues.map(l=>W(l,o,U())));throw Me(r,n?.callee),r}return a.value},rt=at(q),ct=s=>async(e,t,i,n)=>{const o=i?Object.assign(i,{async:!0}):{async:!0};let a=e._zod.run({value:t,issues:[]},o);if(a instanceof Promise&&(a=await a),a.issues.length){const r=new(n?.Err??s)(a.issues.map(l=>W(l,o,U())));throw Me(r,n?.callee),r}return a.value},lt=ct(q),ut=s=>(e,t,i)=>{const n=i?{...i,async:!1}:{async:!1},o=e._zod.run({value:t,issues:[]},n);if(o instanceof Promise)throw new $;return o.issues.length?{success:!1,error:new(s??ot)(o.issues.map(a=>W(a,n,U())))}:{success:!0,data:o.value}},Se=ut(q),dt=s=>async(e,t,i)=>{const n=i?Object.assign(i,{async:!0}):{async:!0};let o=e._zod.run({value:t,issues:[]},n);return o instanceof Promise&&(o=await o),o.issues.length?{success:!1,error:new s(o.issues.map(a=>W(a,n,U())))}:{success:!0,data:o.value}},xe=dt(q),ht=s=>{const e=s?`[\\s\\S]{${s?.minimum??0},${s?.maximum??""}}`:"[\\s\\S]*";return new RegExp(`^${e}$`)},pt=/^-?\d+$/,Re=/^-?\d+(?:\.\d+)?$/,ce=m("$ZodCheck",(s,e)=>{var t;s._zod??(s._zod={}),s._zod.def=e,(t=s._zod).onattach??(t.onattach=[])}),Pe={number:"number",bigint:"bigint",object:"date"},ft=m("$ZodCheckLessThan",(s,e)=>{ce.init(s,e);const t=Pe[typeof e.value];s._zod.onattach.push(i=>{const n=i._zod.bag,o=(e.inclusive?n.maximum:n.exclusiveMaximum)??Number.POSITIVE_INFINITY;e.value<o&&(e.inclusive?n.maximum=e.value:n.exclusiveMaximum=e.value)}),s._zod.check=i=>{(e.inclusive?i.value<=e.value:i.value<e.value)||i.issues.push({origin:t,code:"too_big",maximum:typeof e.value=="object"?e.value.getTime():e.value,input:i.value,inclusive:e.inclusive,inst:s,continue:!e.abort})}}),mt=m("$ZodCheckGreaterThan",(s,e)=>{ce.init(s,e);const t=Pe[typeof e.value];s._zod.onattach.push(i=>{const n=i._zod.bag,o=(e.inclusive?n.minimum:n.exclusiveMinimum)??Number.NEGATIVE_INFINITY;e.value>o&&(e.inclusive?n.minimum=e.value:n.exclusiveMinimum=e.value)}),s._zod.check=i=>{(e.inclusive?i.value>=e.value:i.value>e.value)||i.issues.push({origin:t,code:"too_small",minimum:typeof e.value=="object"?e.value.getTime():e.value,input:i.value,inclusive:e.inclusive,inst:s,continue:!e.abort})}}),gt=m("$ZodCheckNumberFormat",(s,e)=>{ce.init(s,e),e.format=e.format||"float64";const t=e.format?.includes("int"),i=t?"int":"number",[n,o]=nt[e.format];s._zod.onattach.push(a=>{const r=a._zod.bag;r.format=e.format,r.minimum=n,r.maximum=o,t&&(r.pattern=pt)}),s._zod.check=a=>{const r=a.value;if(t){if(!Number.isInteger(r)){a.issues.push({expected:i,format:e.format,code:"invalid_type",continue:!1,input:r,inst:s});return}if(!Number.isSafeInteger(r)){r>0?a.issues.push({input:r,code:"too_big",maximum:Number.MAX_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:s,origin:i,inclusive:!0,continue:!e.abort}):a.issues.push({input:r,code:"too_small",minimum:Number.MIN_SAFE_INTEGER,note:"Integers must be within the safe integer range.",inst:s,origin:i,inclusive:!0,continue:!e.abort});return}}r<n&&a.issues.push({origin:"number",input:r,code:"too_small",minimum:n,inclusive:!0,inst:s,continue:!e.abort}),r>o&&a.issues.push({origin:"number",input:r,code:"too_big",maximum:o,inclusive:!0,inst:s,continue:!e.abort})}}),vt={major:4,minor:3,patch:6},P=m("$ZodType",(s,e)=>{var t;s??(s={}),s._zod.def=e,s._zod.bag=s._zod.bag||{},s._zod.version=vt;const i=[...s._zod.def.checks??[]];s._zod.traits.has("$ZodCheck")&&i.unshift(s);for(const n of i)for(const o of n._zod.onattach)o(s);if(i.length===0)(t=s._zod).deferred??(t.deferred=[]),s._zod.deferred?.push(()=>{s._zod.run=s._zod.parse});else{const n=(a,r,l)=>{let c=B(a),d;for(const h of r){if(h._zod.def.when){if(!h._zod.def.when(a))continue}else if(c)continue;const p=a.issues.length,v=h._zod.check(a);if(v instanceof Promise&&l?.async===!1)throw new $;if(d||v instanceof Promise)d=(d??Promise.resolve()).then(async()=>{await v,a.issues.length!==p&&(c||(c=B(a,p)))});else{if(a.issues.length===p)continue;c||(c=B(a,p))}}return d?d.then(()=>a):a},o=(a,r,l)=>{if(B(a))return a.aborted=!0,a;const c=n(r,i,l);if(c instanceof Promise){if(l.async===!1)throw new $;return c.then(d=>s._zod.parse(d,l))}return s._zod.parse(c,l)};s._zod.run=(a,r)=>{if(r.skipChecks)return s._zod.parse(a,r);if(r.direction==="backward"){const c=s._zod.parse({value:a.value,issues:[]},{...r,skipChecks:!0});return c instanceof Promise?c.then(d=>o(d,a,r)):o(c,a,r)}const l=s._zod.parse(a,r);if(l instanceof Promise){if(r.async===!1)throw new $;return l.then(c=>n(c,i,r))}return n(l,i,r)}}L(s,"~standard",()=>({validate:n=>{try{const o=Se(s,n);return o.success?{value:o.data}:{issues:o.error?.issues}}catch{return xe(s,n).then(o=>o.success?{value:o.data}:{issues:o.error?.issues})}},vendor:"zod",version:1}))}),bt=m("$ZodString",(s,e)=>{P.init(s,e),s._zod.pattern=[...s?._zod.bag?.patterns??[]].pop()??ht(s._zod.bag),s._zod.parse=(t,i)=>{if(e.coerce)try{t.value=String(t.value)}catch{}return typeof t.value=="string"||t.issues.push({expected:"string",code:"invalid_type",input:t.value,inst:s}),t}}),Ie=m("$ZodNumber",(s,e)=>{P.init(s,e),s._zod.pattern=s._zod.bag.pattern??Re,s._zod.parse=(t,i)=>{if(e.coerce)try{t.value=Number(t.value)}catch{}const n=t.value;if(typeof n=="number"&&!Number.isNaN(n)&&Number.isFinite(n))return t;const o=typeof n=="number"?Number.isNaN(n)?"NaN":Number.isFinite(n)?void 0:"Infinity":void 0;return t.issues.push({expected:"number",code:"invalid_type",input:n,inst:s,...o?{received:o}:{}}),t}}),wt=m("$ZodNumberFormat",(s,e)=>{gt.init(s,e),Ie.init(s,e)});function V(s,e,t,i,n){if(s.issues.length){if(n&&!(t in i))return;e.issues.push(...O(t,s.issues))}s.value===void 0?t in i&&(e.value[t]=void 0):e.value[t]=s.value}function yt(s){const e=Object.keys(s.shape);for(const i of e)if(!s.shape?.[i]?._zod?.traits?.has("$ZodType"))throw new Error(`Invalid element at key "${i}": expected a Zod schema`);const t=it(s.shape);return{...s,keys:e,keySet:new Set(e),numKeys:e.length,optionalKeys:new Set(t)}}function _t(s,e,t,i,n,o){const a=[],r=n.keySet,l=n.catchall._zod,c=l.def.type,d=l.optout==="optional";for(const h in e){if(r.has(h))continue;if(c==="never"){a.push(h);continue}const p=l.run({value:e[h],issues:[]},i);p instanceof Promise?s.push(p.then(v=>V(v,t,h,e,d))):V(p,t,h,e,d)}return a.length&&t.issues.push({code:"unrecognized_keys",keys:a,input:e,inst:o}),s.length?Promise.all(s).then(()=>t):t}const kt=m("$ZodObject",(s,e)=>{if(P.init(s,e),!Object.getOwnPropertyDescriptor(e,"shape")?.get){const a=e.shape;Object.defineProperty(e,"shape",{get:()=>{const r={...a};return Object.defineProperty(e,"shape",{value:r}),r}})}const t=Qe(()=>yt(e));L(s._zod,"propValues",()=>{const a=e.shape,r={};for(const l in a){const c=a[l]._zod;if(c.values){r[l]??(r[l]=new Set);for(const d of c.values)r[l].add(d)}}return r});const i=se,n=e.catchall;let o;s._zod.parse=(a,r)=>{o??(o=t.value);const l=a.value;if(!i(l))return a.issues.push({expected:"object",code:"invalid_type",input:l,inst:s}),a;a.value={};const c=[],d=o.shape;for(const h of o.keys){const p=d[h],v=p._zod.optout==="optional",f=p._zod.run({value:l[h],issues:[]},r);f instanceof Promise?c.push(f.then(_=>V(_,a,h,l,v))):V(f,a,h,l,v)}return n?_t(c,l,a,r,t.value,s):c.length?Promise.all(c).then(()=>a):a}}),At=m("$ZodRecord",(s,e)=>{P.init(s,e),s._zod.parse=(t,i)=>{const n=t.value;if(!tt(n))return t.issues.push({expected:"record",code:"invalid_type",input:n,inst:s}),t;const o=[],a=e.keyType._zod.values;if(a){t.value={};const r=new Set;for(const c of a)if(typeof c=="string"||typeof c=="number"||typeof c=="symbol"){r.add(typeof c=="number"?c.toString():c);const d=e.valueType._zod.run({value:n[c],issues:[]},i);d instanceof Promise?o.push(d.then(h=>{h.issues.length&&t.issues.push(...O(c,h.issues)),t.value[c]=h.value})):(d.issues.length&&t.issues.push(...O(c,d.issues)),t.value[c]=d.value)}let l;for(const c in n)r.has(c)||(l=l??[],l.push(c));l&&l.length>0&&t.issues.push({code:"unrecognized_keys",input:n,inst:s,keys:l})}else{t.value={};for(const r of Reflect.ownKeys(n)){if(r==="__proto__")continue;let l=e.keyType._zod.run({value:r,issues:[]},i);if(l instanceof Promise)throw new Error("Async schemas not supported in object keys currently");if(typeof r=="string"&&Re.test(r)&&l.issues.length){const d=e.keyType._zod.run({value:Number(r),issues:[]},i);if(d instanceof Promise)throw new Error("Async schemas not supported in object keys currently");d.issues.length===0&&(l=d)}if(l.issues.length){e.mode==="loose"?t.value[r]=n[r]:t.issues.push({code:"invalid_key",origin:"record",issues:l.issues.map(d=>W(d,i,U())),input:r,path:[r],inst:s});continue}const c=e.valueType._zod.run({value:n[r],issues:[]},i);c instanceof Promise?o.push(c.then(d=>{d.issues.length&&t.issues.push(...O(r,d.issues)),t.value[l.value]=d.value})):(c.issues.length&&t.issues.push(...O(r,c.issues)),t.value[l.value]=c.value)}}return o.length?Promise.all(o).then(()=>t):t}});function ge(s,e){return s.issues.length&&e===void 0?{issues:[],value:void 0}:s}const Tt=m("$ZodOptional",(s,e)=>{P.init(s,e),s._zod.optin="optional",s._zod.optout="optional",L(s._zod,"values",()=>e.innerType._zod.values?new Set([...e.innerType._zod.values,void 0]):void 0),L(s._zod,"pattern",()=>{const t=e.innerType._zod.pattern;return t?new RegExp(`^(${et(t.source)})?$`):void 0}),s._zod.parse=(t,i)=>{if(e.innerType._zod.optin==="optional"){const n=e.innerType._zod.run(t,i);return n instanceof Promise?n.then(o=>ge(o,t.value)):ge(n,t.value)}return t.value===void 0?t:e.innerType._zod.run(t,i)}});var ve;class Ct{constructor(){this._map=new WeakMap,this._idmap=new Map}add(e,...t){const i=t[0];return this._map.set(e,i),i&&typeof i=="object"&&"id"in i&&this._idmap.set(i.id,e),this}clear(){return this._map=new WeakMap,this._idmap=new Map,this}remove(e){const t=this._map.get(e);return t&&typeof t=="object"&&"id"in t&&this._idmap.delete(t.id),this._map.delete(e),this}get(e){const t=e._zod.parent;if(t){const i={...this.get(t)??{}};delete i.id;const n={...i,...this._map.get(e)};return Object.keys(n).length?n:void 0}return this._map.get(e)}has(e){return this._map.has(e)}}function Et(){return new Ct}(ve=globalThis).__zod_globalRegistry??(ve.__zod_globalRegistry=Et());function Mt(s,e){return new s({type:"string",...T()})}function zt(s,e){return new s({type:"number",checks:[],...T()})}function St(s,e){return new s({type:"number",check:"number_format",abort:!1,format:"safeint",...T()})}function Oe(s,e){return new ft({check:"less_than",...T(),value:s,inclusive:!0})}function xt(s,e){return new mt({check:"greater_than",...T(),value:s,inclusive:!0})}function Ne(s){return xt(0)}const F=m("ZodMiniType",(s,e)=>{if(!s._zod)throw new Error("Uninitialized schema in ZodMiniType.");P.init(s,e),s.def=e,s.type=e.type,s.parse=(t,i)=>rt(s,t,i,{callee:s.parse}),s.safeParse=(t,i)=>Se(s,t,i),s.parseAsync=async(t,i)=>lt(s,t,i,{callee:s.parseAsync}),s.safeParseAsync=async(t,i)=>xe(s,t,i),s.check=(...t)=>s.clone({...e,checks:[...e.checks??[],...t.map(i=>typeof i=="function"?{_zod:{check:i,def:{check:"custom"},onattach:[]}}:i)]},{parent:!0}),s.with=s.check,s.clone=(t,i)=>st(s,t,i),s.brand=()=>s,s.register=((t,i)=>(t.add(s,i),s)),s.apply=t=>t(s)}),Rt=m("ZodMiniString",(s,e)=>{bt.init(s,e),F.init(s,e)});function Pt(s){return Mt(Rt)}const $e=m("ZodMiniNumber",(s,e)=>{Ie.init(s,e),F.init(s,e)});function E(s){return zt($e)}const It=m("ZodMiniNumberFormat",(s,e)=>{wt.init(s,e),$e.init(s,e)});function je(s){return St(It)}const Ot=m("ZodMiniObject",(s,e)=>{kt.init(s,e),F.init(s,e),L(s,"shape",()=>e.shape)});function Nt(s,e){const t={type:"object",shape:s??{},...T()};return new Ot(t)}const $t=m("ZodMiniRecord",(s,e)=>{At.init(s,e),F.init(s,e)});function jt(s,e,t){return new $t({type:"record",keyType:s,valueType:e,...T()})}const Lt=m("ZodMiniOptional",(s,e)=>{Tt.init(s,e),F.init(s,e)});function D(s){return new Lt({type:"optional",innerType:s})}E().check(je(),Ne(),Oe(255)).brand("u8");const Ut=E().check(je(),Ne(),Oe(Number.MAX_SAFE_INTEGER)).brand("u53");function w(s){return Ut.parse(s)}const Le=Nt({x:D(E()),y:D(E()),z:D(E()),s:D(E())}),Wt=jt(Pt(),Le),C={chat:90,audio:80,video:60,typing:40,location:20,preview:10};function be(s){return new TextEncoder().encode(JSON.stringify(s))}class X{#e;#t;constructor(e){this.#e=e}encode(e,t,i){if(i)this.#t?.close(),this.#t=this.#e.appendGroup();else if(!this.#t)throw new Error("must start with a keyframe");this.#t?.writeFrame(X.#s(e,t))}static#s(e,t){const i=Xe(t),n=e.byteLength,o=new Uint8Array(i.byteLength+n);return o.set(i,0),e instanceof Uint8Array?o.set(e,i.byteLength):e.copyTo(o.subarray(i.byteLength)),o}close(e){this.#e.close(e),this.#t?.close()}}navigator.userAgent.toLowerCase().includes("chrome");const Ft=navigator.userAgent.toLowerCase().includes("firefox");let ee;async function Bt(){return globalThis.AudioEncoder&&globalThis.AudioDecoder?!0:(ee||(console.warn("using Opus polyfill; performance may be degraded"),ee=Promise.all([pe(()=>import("./libav-opus-af-BlMWboA7-B4GfDr9_.js"),[],import.meta.url),pe(()=>import("./main-DGBFe0O7-DQ8if_La.js"),[],import.meta.url)]).then(async([s,e])=>(await e.load({LibAV:s,polyfill:!0}),!0))),await ee)}const Ht=`var __defProp = Object.defineProperty;
-var __getOwnPropNames = Object.getOwnPropertyNames;
-var __esm = (fn, res) => function __init() {
-  return fn && (res = (0, fn[__getOwnPropNames(fn)[0]])(fn = 0)), res;
-};
-var __export = (target, all) => {
-  for (var name in all)
-    __defProp(target, name, { get: all[name], enumerable: true });
-};
-
-// ../lite/src/path.ts
-var init_path = __esm({
-  "../lite/src/path.ts"() {
-    "use strict";
-  }
-});
-
-// ../lite/src/varint.ts
-var MAX_U6, MAX_U14, MAX_U30, MAX_U53, MAX_U64, MAX_U62;
-var init_varint = __esm({
-  "../lite/src/varint.ts"() {
-    "use strict";
-    MAX_U6 = 2 ** 6 - 1;
-    MAX_U14 = 2 ** 14 - 1;
-    MAX_U30 = 2 ** 30 - 1;
-    MAX_U53 = Number.MAX_SAFE_INTEGER;
-    MAX_U64 = (1n << 64n) - 1n;
-    MAX_U62 = 2n ** 62n - 1n;
-  }
-});
-
-// ../lite/src/index.ts
-init_path();
-
-// ../lite/src/time.ts
-var time_exports = {};
-__export(time_exports, {
-  Micro: () => Micro,
-  Milli: () => Milli,
-  Nano: () => Nano,
-  Second: () => Second
-});
-var Nano = {
-  zero: 0,
-  fromMicro: (us) => us * 1e3,
-  fromMilli: (ms) => ms * 1e6,
-  fromSecond: (s) => s * 1e9,
-  toMicro: (ns) => ns / 1e3,
-  toMilli: (ns) => ns / 1e6,
-  toSecond: (ns) => ns / 1e9,
-  now: () => performance.now() * 1e6,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Micro = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e3,
-  fromMilli: (ms) => ms * 1e3,
-  fromSecond: (s) => s * 1e6,
-  toNano: (us) => us * 1e3,
-  toMilli: (us) => us / 1e3,
-  toSecond: (us) => us / 1e6,
-  now: () => performance.now() * 1e3,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Milli = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e6,
-  fromMicro: (us) => us / 1e3,
-  fromSecond: (s) => s * 1e3,
-  toNano: (ms) => ms * 1e6,
-  toMicro: (ms) => ms * 1e3,
-  toSecond: (ms) => ms / 1e3,
-  now: () => performance.now(),
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-var Second = {
-  zero: 0,
-  fromNano: (ns) => ns / 1e9,
-  fromMicro: (us) => us / 1e6,
-  fromMilli: (ms) => ms / 1e3,
-  toNano: (s) => s * 1e9,
-  toMicro: (s) => s * 1e6,
-  toMilli: (s) => s * 1e3,
-  now: () => performance.now() / 1e3,
-  add: (a, b) => a + b,
-  sub: (a, b) => a - b,
-  mul: (a, b) => a * b,
-  div: (a, b) => a / b,
-  max: (a, b) => Math.max(a, b),
-  min: (a, b) => Math.min(a, b)
-};
-
-// ../lite/src/index.ts
-init_varint();
-
-// src/audio/capture-worklet.ts
-var Capture = class extends AudioWorkletProcessor {
-  #sampleCount = 0;
-  process(input) {
-    if (input.length > 1) throw new Error("only one input is supported.");
-    const channels = input[0];
-    if (channels.length === 0) return true;
-    const timestamp = time_exports.Micro.fromSecond(this.#sampleCount / sampleRate);
-    const msg = {
-      timestamp,
-      channels
-    };
-    this.port.postMessage(msg);
-    this.#sampleCount += channels[0].length;
-    return true;
-  }
-};
-registerProcessor("capture", Capture);
-`,Dt=new Blob([Ht],{type:"application/javascript"}),Kt=URL.createObjectURL(Dt),we=.001,te=.2;let ye=class Ue{static TRACK="audio/data";static PRIORITY=C.audio;enabled;muted;volume;sampleRate;groupDuration;source;#e=new u(void 0);catalog=this.#e;#t=new u(void 0);config=this.#t;#s=new u(void 0);#n=new u(void 0);root=this.#n;active=new u(!1);#i=new g;constructor(e){this.source=u.from(e?.source),this.enabled=u.from(e?.enabled??!1),this.muted=u.from(e?.muted??!1),this.volume=u.from(e?.volume??1),this.sampleRate=u.from(e?.sampleRate),this.groupDuration=e?.groupDuration??100,this.#i.run(this.#o.bind(this)),this.#i.run(this.#a.bind(this)),this.#i.run(this.#r.bind(this)),this.#i.run(this.#c.bind(this))}#o(e){const t=e.getAll([this.enabled,this.source]);if(!t)return;const[i,n]=t,o=n.getSettings(),a=e.get(this.sampleRate)??o.sampleRate,r=new AudioContext({latencyHint:"interactive",sampleRate:a});e.cleanup(()=>r.close());const l=new MediaStreamAudioSourceNode(r,{mediaStream:new MediaStream([n])});e.cleanup(()=>l.disconnect());const c=new GainNode(r,{gain:this.volume.peek()});l.connect(c),e.cleanup(()=>c.disconnect()),e.spawn(async()=>{if(await r.audioWorklet.addModule(Kt),r.state==="closed")return;const d=new AudioWorkletNode(r,"capture",{numberOfInputs:1,numberOfOutputs:0,channelCount:o.channelCount??l.channelCount});e.set(this.#s,d),c.connect(d),e.cleanup(()=>d.disconnect()),e.set(this.#n,c)})}#a(e){const t=e.getAll([this.source,this.#s]);if(!t)return;const[i,n]=t,o={codec:"opus",sampleRate:w(n.context.sampleRate),numberOfChannels:w(n.channelCount),bitrate:w(n.channelCount*32e3),container:{kind:"legacy"},jitter:w(20)};e.set(this.#t,o)}#r(e){const t=e.get(this.#n);if(!t)return;e.cleanup(()=>t.gain.cancelScheduledValues(t.context.currentTime));const i=e.get(this.muted)?0:e.get(this.volume);i<we?(t.gain.exponentialRampToValueAtTime(we,t.context.currentTime+te),t.gain.setValueAtTime(0,t.context.currentTime+te+.01)):t.gain.exponentialRampToValueAtTime(i,t.context.currentTime+te)}serve(e,t){const i=t.getAll([this.enabled,this.#s,this.#t]);if(!i)return;const[n,o,a]=i;t.set(this.active,!0,!1);const r=new X(e);t.cleanup(()=>r.close());let l;t.spawn(async()=>{await Bt();const c=new AudioEncoder({output:d=>{if(d.type!=="key")throw new Error("only key frames are supported");let h=!1;(!l||l+re.fromMilli(this.groupDuration)<=d.timestamp)&&(l=d.timestamp,h=!0),r.encode(d,d.timestamp,h)},error:d=>{console.error("encoder error",d),r.close(d),o.port.onmessage=null}});t.cleanup(()=>c.close()),console.debug("encoding audio",a),c.configure(a),o.port.onmessage=({data:d})=>{const h=d.channels.slice(0,o.channelCount),p=h.reduce((_,k)=>_+k.length,0),v=new Float32Array(p);h.reduce((_,k)=>(v.set(k,_),_+k.length),0);const f=new AudioData({format:"f32-planar",sampleRate:o.context.sampleRate,numberOfFrames:h[0].length,numberOfChannels:h.length,timestamp:d.timestamp,data:v,transfer:[v.buffer]});c.encode(f),f.close()},t.cleanup(()=>{o.port.onmessage=null})})}#c(e){const t=e.get(this.#t);if(!t)return;const i={renditions:{[Ue.TRACK]:t}};e.set(this.#e,i)}close(){this.#i.close()}};class G{static TRACK="chat/message.txt";static PRIORITY=C.chat;enabled;latest;catalog=new u(void 0);#e=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.latest=new u(""),this.#e.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{name:G.TRACK})})}serve(e,t){if(!t.get(this.enabled))return;const i=t.get(this.latest);e.writeString(i??"")}close(){this.#e.close()}}class Y{static TRACK="chat/typing.bool";static PRIORITY=C.typing;enabled;active;catalog=new u(void 0);#e=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.active=new u(!1),this.#e.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{name:Y.TRACK})})}serve(e,t){if(!t.get(this.enabled))return;const i=t.get(this.active);e.writeBool(i)}close(){this.#e.close()}}let Vt=class{message;typing;#e=new u(void 0);catalog=this.#e;#t=new g;constructor(s){this.message=new G(s?.message),this.typing=new Y(s?.typing),this.#t.run(e=>{this.#e.set({message:e.get(this.message.catalog),typing:e.get(this.typing.catalog)})})}close(){this.#t.close(),this.message.close(),this.typing.close()}};function We(s,e,t){const i=t.parse(e);s.writeJson(i)}class J{static TRACK="location/peers.json";static PRIORITY=C.location;enabled;positions=new u({});catalog=new u(void 0);signals=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.positions=u.from(e?.positions??{}),this.signals.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{name:J.TRACK})})}serve(e,t){const i=t.getAll([this.enabled,this.positions]);if(!i)return;const[n,o]=i;We(e,o,Wt)}close(){this.signals.close()}}class Q{static TRACK="location/window.json";static PRIORITY=C.location;enabled;position;handle;catalog=new u(void 0);signals=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.position=u.from(e?.position??void 0),this.handle=u.from(e?.handle??void 0),this.signals.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{initial:this.position.peek(),track:{name:Q.TRACK},handle:t.get(this.handle)})})}serve(e,t){const i=t.getAll([this.enabled,this.position]);if(!i)return;const[n,o]=i;We(e,o,Le)}close(){this.signals.close()}}let Zt=class{window;peers;catalog=new u(void 0);signals=new g;constructor(s){this.window=new Q(s?.window),this.peers=new J(s?.peers),this.signals.run(this.#e.bind(this))}#e(s){const e=s.get(this.window.catalog),t=s.get(this.peers.catalog);!e&&!t||s.set(this.catalog,{peers:t,...e})}close(){this.signals.close(),this.window.close(),this.peers.close()}};class Z{static TRACK="preview.json";static PRIORITY=C.preview;enabled;info;catalog=new u(void 0);signals=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.info=u.from(e?.info),this.signals.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{name:Z.TRACK})})}serve(e,t){const i=t.getAll([this.enabled,this.info]);if(!i)return;const[n,o]=i;e.writeJson(o)}close(){this.signals.close()}}class qt{enabled;id;name;avatar;color;catalog=new u(void 0);signals=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.id=u.from(e?.id),this.name=u.from(e?.name),this.avatar=u.from(e?.avatar),this.color=u.from(e?.color),this.signals.run(t=>{t.get(this.enabled)&&t.set(this.catalog,{id:t.get(this.id),name:t.get(this.name),avatar:t.get(this.avatar),color:t.get(this.color)})})}close(){this.signals.close()}}class _e{enabled;source;frame;#e=new u(void 0);catalog=this.#e;#t=new g;config;#s=new u(void 0);#n=new u(void 0);active=new u(!1);connection;constructor(e,t,i,n){this.frame=e,this.source=t,this.connection=i,this.enabled=u.from(n?.enabled??!1),this.config=u.from(n?.config),this.#t.run(this.#i.bind(this)),this.#t.run(this.#o.bind(this)),this.#t.run(this.#a.bind(this))}serve(e,t){if(!t.get(this.enabled))return;const i=new X(e);t.cleanup(()=>i.close());let n;t.set(this.active,!0,!1),t.spawn(async()=>{const o=new VideoEncoder({output:a=>{a.type==="key"&&(n=a.timestamp),i.encode(a,a.timestamp,a.type==="key")},error:a=>{i.close(a)}});t.cleanup(()=>o.close()),t.run(()=>{const a=t.get(this.#n);a&&o.configure(a)}),t.run(a=>{const r=a.get(this.frame);if(!r||o.state!=="configured")return;const l=this.config.peek()?.keyframeInterval??K.fromSecond(2),c=!n||n+re.fromMilli(l)<=r.timestamp;c&&(n=r.timestamp),o.encode(r,{keyFrame:c})})})}#i(e){const t=e.getAll([this.enabled,this.#n]);if(!t)return;const[i,n]=t,o={codec:n.codec,bitrate:n.bitrate?w(n.bitrate):void 0,framerate:n.framerate,codedWidth:w(n.width),codedHeight:w(n.height),optimizeForLatency:!0,container:{kind:"legacy"},jitter:n.framerate?w(Math.ceil(1e3/n.framerate)):void 0};e.set(this.#e,o)}#o(e){const t=e.getAll([this.enabled,this.source,this.#s]);if(!t)return;const[i,n,o]=t,a=n.getSettings().frameRate??30,r=e.get(this.config)??{},l=r.maxPixels??o.width*o.height,c=r.bitrateScale??.07;e.spawn(async()=>{const d=await this.#r(e);if(!d)return;const{codec:h,hardwareAcceleration:p}=d,v=30+(a-30)/2;let f=Math.round(l*c*v);if(h.startsWith("avc1"))f*=1;else if(h.startsWith("hev1"))f*=.7;else if(h.startsWith("vp09"))f*=.8;else if(h.startsWith("av01"))f*=.6;else if(h==="vp8")f*=1.1;else throw new Error(`unknown codec: ${h}`);if(f=Math.round(Math.min(f,r.maxBitrate||f)),!r.maxBitrate){const k=e.get(this.connection)?.sendBandwidth;if(k){const he=e.get(k);if(he!=null){const Ve=Math.round(he*.9);f=Math.min(f,Ve)}}}const _={codec:h,width:o.width,height:o.height,framerate:a,bitrate:f,avc:h.startsWith("avc1")?{format:"annexb"}:void 0,hevc:h.startsWith("hev1")?{format:"annexb"}:void 0,latencyMode:"realtime",hardwareAcceleration:p};e.set(this.#n,_)})}#a(e){const t=e.get(this.config),i=e.get(this.frame);if(!i)return;const n=t?.maxPixels??i.codedWidth*i.codedHeight,o=Math.min(Math.sqrt(n/(i.codedWidth*i.codedHeight)),1),a=16*Math.floor(i.codedWidth*o/16),r=16*Math.floor(i.codedHeight*o/16);e.set(this.#s,{width:a,height:r})}async#r(e){const t=e.get(this.config)?.codec??"",i=e.get(this.#s);if(!i)return;const n=["vp09.00.10.08","vp09","avc1.640028","avc1.4D401F","avc1.42E01E","avc1","av01.0.08M.08","av01","hev1.1.6.L93.B0","hev1","vp8"],o=["avc1.640028","avc1.4D401F","avc1.42E01E","avc1","vp8","vp09.00.10.08","vp09","hev1.1.6.L93.B0","hev1","av01.0.08M.08","av01"];if(!Ft)for(const a of n){if(!a.startsWith(t))continue;const r="prefer-hardware",l={codec:a,width:i.width,height:i.height,latencyMode:"realtime",hardwareAcceleration:r,avc:a.startsWith("avc1")?{format:"annexb"}:void 0,hevc:a.startsWith("hev1")?{format:"annexb"}:void 0},{supported:c}=await VideoEncoder.isConfigSupported(l);if(c)return{codec:a,hardwareAcceleration:r}}for(const a of o){if(!a.startsWith(t))continue;const r="prefer-software",l={codec:a,width:i.width,height:i.height,latencyMode:"realtime",hardwareAcceleration:r,avc:a.startsWith("avc1")?{format:"annexb"}:void 0,hevc:a.startsWith("hev1")?{format:"annexb"}:void 0},{supported:c}=await VideoEncoder.isConfigSupported(l);if(c)return{codec:a,hardwareAcceleration:r}}throw new Error("no supported codec")}close(){this.#t.close()}}function Xt(s){if(self.MediaStreamTrackProcessor){const o=performance.now()*1e3,a=new TransformStream({transform(r,l){const c=new VideoFrame(r,{timestamp:r.timestamp+o});r.close(),l.enqueue(c)}});return new self.MediaStreamTrackProcessor({track:s}).readable.pipeThrough(a)}console.warn("Using MediaStreamTrackProcessor polyfill; performance might suffer.");const e=s.getSettings();if(!e)throw new Error("track has no settings");let t,i;const n=e.frameRate??30;return new ReadableStream({async start(){t=document.createElement("video"),t.srcObject=new MediaStream([s]),await Promise.all([t.play(),new Promise(o=>{t.onloadedmetadata=o})]),i=K.now()},async pull(o){for(;;){const a=K.now();if(K.sub(a,i)<1e3/n){await new Promise(r=>requestAnimationFrame(r));continue}i=a,o.enqueue(new VideoFrame(t,{timestamp:re.fromMilli(i)}))}}})}class S{static TRACK_HD="video/hd";static TRACK_SD="video/sd";static PRIORITY=C.video;source;hd;sd;frame=new u(void 0);catalog=new u(void 0);display=new u(void 0);flip=new u(!1);signals=new g;constructor(e){this.source=u.from(e?.source);const t=e?.connection??new u(void 0);this.hd=new _e(this.frame,this.source,t,e?.hd),this.sd=new _e(this.frame,this.source,t,e?.sd),this.flip=u.from(e?.flip??!1),this.signals.run(this.#t.bind(this)),this.signals.run(this.#e.bind(this))}#e(e){const t=e.get(this.source);if(!t)return;const i=Xt(t).getReader();e.cleanup(()=>i.cancel()),e.spawn(async()=>{for(;;){const n=await Promise.race([i.read(),e.cancel]);if(!n?.value)break;this.frame.update(o=>(o?.close(),n.value)),this.display.set({width:n.value.codedWidth,height:n.value.codedHeight})}}),e.cleanup(()=>{this.frame.update(n=>{n?.close()}),this.display.set(void 0)})}#t(e){if(!e.get(this.source))return;const t=e.get(this.display);if(!t)return;const i=e.get(this.hd.catalog),n=e.get(this.sd.catalog),o={};i&&(o[S.TRACK_HD]=i),n&&(o[S.TRACK_SD]=n);const a={renditions:o,display:{width:w(t.width),height:w(t.height)},flip:e.get(this.flip)??void 0};e.set(this.catalog,a)}close(){this.signals.close(),this.hd.close(),this.sd.close(),this.frame.update(e=>{e?.close()})}}class le{static CATALOG_TRACK="catalog.json";connection;enabled;name;audio;video;location;chat;preview;user;signals=new g;constructor(e){this.connection=u.from(e?.connection),this.enabled=u.from(e?.enabled??!1),this.name=u.from(e?.name??Ze()),this.audio=new ye(e?.audio),this.video=new S({...e?.video,connection:this.connection}),this.location=new Zt(e?.location),this.chat=new Vt(e?.chat),this.preview=new Z(e?.preview),this.user=new qt(e?.user),this.signals.run(this.#e.bind(this))}#e(e){const t=e.getAll([this.enabled,this.connection]);if(!t)return;const[i,n]=t,o=e.get(this.name),a=new qe;e.cleanup(()=>a.close()),n.publish(o,a),e.spawn(this.#t.bind(this,a,e))}async#t(e,t){for(;;){const i=await e.requested();if(!i)break;t.cleanup(()=>i.track.close()),t.run(n=>{if(!n.get(i.track.state.closed))switch(i.track.name){case le.CATALOG_TRACK:this.#s(i.track,n);break;case Q.TRACK:this.location.window.serve(i.track,n);break;case J.TRACK:this.location.peers.serve(i.track,n);break;case Z.TRACK:this.preview.serve(i.track,n);break;case Y.TRACK:this.chat.typing.serve(i.track,n);break;case G.TRACK:this.chat.message.serve(i.track,n);break;case ye.TRACK:this.audio.serve(i.track,n);break;case S.TRACK_HD:this.video.hd.serve(i.track,n);break;case S.TRACK_SD:this.video.sd.serve(i.track,n);break;default:console.error("received subscription for unknown track",i.track.name),i.track.close(new Error(`Unknown track: ${i.track.name}`));break}})}}#s(e,t){if(!t.get(this.enabled)){e.writeFrame(be({}));return}const i={video:t.get(this.video.catalog),audio:t.get(this.audio.catalog),location:t.get(this.location.catalog),user:t.get(this.user.catalog),chat:t.get(this.chat.catalog),preview:t.get(this.preview.catalog)},n=be(i);e.writeFrame(n)}close(){this.signals.close(),this.audio.close(),this.video.close(),this.location.close(),this.chat.close(),this.preview.close(),this.user.close()}}class Fe{kind;#e=new u(void 0);available=this.#e;#t=new u(void 0);default=this.#t;preferred;active=new u(void 0);permission=new u(!1);#s=new u(void 0);requested=this.#s;signals=new g;constructor(e,t){this.kind=e,this.preferred=u.from(t?.preferred),this.signals.run(i=>{i.spawn(this.#n.bind(this,i)),i.event(navigator.mediaDevices,"devicechange",()=>this.permission.mutate(()=>{}))}),this.signals.run(this.#i.bind(this))}async#n(e){e.get(this.permission);let t=await Promise.race([navigator.mediaDevices.enumerateDevices().catch(()=>{}),e.cancel]);if(!t)return;if(t=t.filter(o=>o.kind===`${this.kind}input`),t.some(o=>o.deviceId==="")){console.warn(`no ${this.kind} permission`),this.#e.set(void 0),this.#t.set(void 0);return}this.permission.set(!0),t.length||console.warn(`no ${this.kind} devices found`);const i=t.find(o=>o.deviceId==="default");t=t.filter(o=>o.deviceId!=="default");let n;i&&(n=t.find(o=>o.groupId===i.groupId)),n||(this.kind==="audio"?n=t.find(o=>{const a=o.label.toLowerCase();return a.includes("default")||a.includes("communications")}):this.kind==="video"&&(n=t.find(o=>{const a=o.label.toLowerCase();return a.includes("front")||a.includes("external")||a.includes("usb")}))),n||(n=t.at(0)),this.#e.set(t),this.#t.set(n?.deviceId)}#i(e){const t=e.get(this.preferred);t&&e.get(this.#e)?.some(i=>i.deviceId===t)?this.#s.set(t):this.#s.set(e.get(this.default))}requestPermission(){this.permission.peek()||navigator.mediaDevices.getUserMedia({[this.kind]:!0}).then(e=>{this.permission.set(!0);const t=e.getTracks().at(0)?.getSettings().deviceId;t&&this.preferred.set(t),e.getTracks().forEach(i=>{i.stop()})}).catch(()=>{})}close(){this.signals.close()}}class Gt{enabled;device;constraints;source=new u(void 0);signals=new g;constructor(e){this.device=new Fe("video",e?.device),this.enabled=u.from(e?.enabled??!1),this.constraints=u.from(e?.constraints),this.signals.run(this.#e.bind(this))}#e(e){if(!e.get(this.enabled))return;const t=e.get(this.device.requested),i={...e.get(this.constraints)??{},deviceId:t?{exact:t}:void 0};e.spawn(async()=>{const n=navigator.mediaDevices.getUserMedia({video:i}).catch(()=>{});e.cleanup(()=>n.then(l=>l?.getTracks().forEach(c=>{c.stop()})));const o=await Promise.race([n,e.cancel]);if(!o)return;this.device.permission.set(!0);const a=o.getVideoTracks()[0];if(!a)return;const r=a.getSettings();e.set(this.device.active,r.deviceId),e.set(this.source,a)})}close(){this.signals.close(),this.device.close()}}let Yt=class{file=new u(void 0);signals=new g;source=new u({});enabled;constructor(s){this.enabled=u.from(s.enabled??!1),this.file=u.from(s.file),this.signals.run(e=>{const t=e.getAll([this.file,this.enabled]);if(!t)return;const[i]=t;this.#e(i,e).catch(n=>{console.error("Failed to decode file:",n)})})}async#e(s,e){const t=s.type;if(t.startsWith("image/"))await this.#t(s,e);else if(t.startsWith("video/")||t.startsWith("audio/"))await this.#s(s,e);else throw new Error(`Unsupported file type: ${t}`)}async#t(s,e){const t=new Image,i=URL.createObjectURL(s);t.src=i,await t.decode(),e.cleanup(()=>URL.revokeObjectURL(i));const n=document.createElement("canvas");n.width=t.width,n.height=t.height;const o=n.getContext("2d");if(!o)throw new Error("Failed to create 2D canvas context");const a=setInterval(()=>{o.drawImage(t,0,0)},1e3/30);e.cleanup(()=>clearInterval(a));const r=n.captureStream(30).getVideoTracks()[0];if(!r)throw new Error("Failed to capture video track from canvas stream");e.set(this.source,{video:r},{})}async#s(s,e){const t=document.createElement("video"),i=URL.createObjectURL(s);t.src=i,t.loop=!0,t.muted=!0,await new Promise((r,l)=>{t.onloadedmetadata=()=>r(),t.onerror=()=>l(new Error("Failed to load video"))}),await t.play(),e.cleanup(()=>{t.pause(),URL.revokeObjectURL(i)});const n=t.captureStream(),o=n.getVideoTracks()[0],a=n.getAudioTracks()[0];if(!o&&!a)throw new Error("Failed to capture any tracks from video element");e.set(this.source,{video:o,audio:a},{})}close(){this.signals.close()}};class Jt{enabled;device;constraints;source=new u(void 0);signals=new g;constructor(e){this.device=new Fe("audio",e?.device),this.enabled=u.from(e?.enabled??!1),this.constraints=u.from(e?.constraints),this.signals.run(this.#e.bind(this))}#e(e){if(!e.get(this.enabled))return;const t=e.get(this.device.requested),i={...e.get(this.constraints)??{},deviceId:t!==void 0?{exact:t}:void 0};e.spawn(async()=>{const n=navigator.mediaDevices.getUserMedia({audio:i}).catch(()=>{});e.cleanup(()=>n.then(l=>l?.getTracks().forEach(c=>{c.stop()})));const o=await Promise.race([n,e.cancel]);if(!o)return;this.device.permission.set(!0);const a=o.getAudioTracks()[0];if(!a)return;const r=a.getSettings();t===void 0&&this.device.preferred.set(r.deviceId),e.set(this.device.active,r.deviceId),e.set(this.source,a)})}close(){this.signals.close(),this.device.close()}}class Qt{enabled;video;audio;source=new u(void 0);signals=new g;constructor(e){this.enabled=u.from(e?.enabled??!1),this.video=u.from(e?.video),this.audio=u.from(e?.audio),this.signals.run(this.#e.bind(this))}#e(e){if(!e.get(this.enabled))return;const t=e.get(this.video),i=e.get(this.audio);let n;typeof self.CaptureController<"u"&&(n=new CaptureController,n.setFocusBehavior("no-focus-change")),e.spawn(async()=>{const o=await Promise.race([navigator.mediaDevices.getDisplayMedia({video:t,audio:i,controller:n,preferCurrentTab:!1,selfBrowserSurface:"exclude",surfaceSwitching:"include"}).catch(()=>{}),e.cancel]);if(!o)return;const a=o.getVideoTracks().at(0),r=o.getAudioTracks().at(0);e.cleanup(()=>a?.stop()),e.cleanup(()=>r?.stop()),e.set(this.source,{video:a,audio:r})})}close(){this.signals.close()}}const es=["url","name","muted","invisible","source"],ts=new FinalizationRegistry(s=>s.close());class ss extends HTMLElement{static observedAttributes=es;state={source:new u(void 0),muted:new u(!1),invisible:new u(!1)};connection;broadcast;#e=new u(void 0);video=new u(void 0);audio=new u(void 0);file=new u(void 0);#t;#s;#n;#i=new u(!1);signals=new g;constructor(){super(),ts.register(this,this.signals),this.connection=new Ge({enabled:this.#i}),this.signals.cleanup(()=>this.connection.close()),this.#t=new u(!1),this.#s=new u(!1),this.#n=new u(!1),this.signals.run(i=>{const n=i.get(this.state.muted),o=i.get(this.state.invisible);this.#t.set(!o),this.#s.set(!n),this.#n.set(!n||!o)}),this.broadcast=new le({connection:this.connection.established,enabled:this.#i,audio:{enabled:this.#s},video:{hd:{enabled:this.#t}}}),this.signals.cleanup(()=>this.broadcast.close());const e=()=>{this.#e.set(this.querySelector("video"))},t=new MutationObserver(e);t.observe(this,{childList:!0,subtree:!0}),this.signals.cleanup(()=>t.disconnect()),e(),this.signals.run(i=>{const n=i.get(this.#e);if(!n)return;const o=i.get(this.broadcast.video.source);if(!o){n.style.display="none";return}n.srcObject=new MediaStream([o]),n.style.display="block",i.cleanup(()=>{n.srcObject=null})}),this.signals.run(this.#o.bind(this))}connectedCallback(){this.#i.set(!0)}disconnectedCallback(){this.#i.set(!1)}attributeChangedCallback(e,t,i){if(t!==i)if(e==="url")this.connection.url.set(i?new URL(i):void 0);else if(e==="name")this.broadcast.name.set(fe(i??""));else if(e==="source")if(i==="camera"||i==="screen"||i==="file"||i===null)this.state.source.set(i);else throw new Error(`Invalid source: ${i}`);else if(e==="muted")this.state.muted.set(i!==null);else if(e==="invisible")this.state.invisible.set(i!==null);else{const n=e;throw new Error(`Invalid attribute: ${n}`)}}#o(e){const t=e.get(this.state.source);if(!t)return;if(t==="camera"){const n=new Gt({enabled:this.#t});this.signals.run(a=>{const r=a.get(n.source);this.broadcast.video.source.set(r)});const o=new Jt({enabled:this.#s});this.signals.run(a=>{const r=a.get(o.source);this.broadcast.audio.source.set(r)}),e.set(this.video,n),e.set(this.audio,o),e.cleanup(()=>{n.close(),o.close()});return}if(t==="screen"){const n=new Qt({enabled:this.#n});this.signals.run(o=>{const a=o.get(n.source);a&&(o.set(this.broadcast.video.source,a.video),o.set(this.broadcast.audio.source,a.audio))}),e.set(this.video,n),e.set(this.audio,n),e.cleanup(()=>{n.close()});return}if(t==="file"||t instanceof File){const n=new Yt({file:t instanceof File?t:void 0,enabled:this.#n});this.signals.run(o=>{const a=o.get(n.source);this.broadcast.video.source.set(a.video),this.broadcast.audio.source.set(a.audio)}),e.cleanup(()=>{n.close()});return}const i=t;throw new Error(`Invalid source: ${i}`)}get url(){return this.connection.url.peek()}set url(e){this.connection.url.set(e?new URL(e):void 0)}get name(){return this.broadcast.name.peek()}set name(e){this.broadcast.name.set(fe(e))}get source(){return this.state.source.peek()}set source(e){this.state.source.set(e)}get muted(){return this.state.muted.peek()}set muted(e){this.state.muted.set(e)}get invisible(){return this.state.invisible.peek()}set invisible(e){this.state.invisible.set(e)}}customElements.define("moq-publish",ss);const Be=navigator.userAgent.toLowerCase().includes("firefox"),ie={aac:"mp4a.40.2",opus:"opus",av1:"av01.0.08M.08",h264:"avc1.640028",h265:"hev1.1.6.L93.B0",vp9:"vp09.00.10.08",vp8:"vp8"};async function ke(s){return globalThis.AudioEncoder?(await AudioEncoder.isConfigSupported({codec:ie[s],numberOfChannels:2,sampleRate:48e3})).supported===!0:!1}async function I(s){const e=await VideoEncoder.isConfigSupported({codec:ie[s],width:1280,height:720,hardwareAcceleration:"prefer-software"}),t=await VideoEncoder.isConfigSupported({codec:ie[s],width:1280,height:720,hardwareAcceleration:"prefer-hardware"});return{hardware:Be||t.config?.hardwareAcceleration!=="prefer-hardware"?void 0:t.supported===!0,software:e.supported===!0}}async function is(){return{webtransport:typeof WebTransport<"u"?Be?"partial":"full":"partial",audio:{capture:typeof AudioWorkletNode<"u",encoding:{aac:await ke("aac"),opus:await ke("opus")?"full":"partial"}},video:{capture:typeof MediaStreamTrackProcessor<"u"?"full":typeof OffscreenCanvas<"u"?"partial":"none",encoding:typeof VideoEncoder<"u"?{h264:await I("h264"),h265:await I("h265"),vp8:await I("vp8"),vp9:await I("vp9"),av1:await I("av1")}:void 0}}}function b(s,e,...t){const i=document.createElement(s);if(!e)return i;const{style:n,classList:o,dataset:a,attributes:r,...l}=e;return n&&Object.assign(i.style,n),o&&i.classList.add(...o),a&&Object.entries(a).forEach(([c,d])=>{i.dataset[c]=d}),r&&Object.entries(r).forEach(([c,d])=>{i.setAttribute(c,d)}),t&&t.forEach(c=>{typeof c=="string"?i.appendChild(document.createTextNode(c)):i.appendChild(c)}),Object.assign(i,l),i}const Ae=navigator.userAgent.toLowerCase().includes("firefox"),ns=["show","details"];class os extends HTMLElement{#e=new u("warning");#t=new u(!1);#s=new u(void 0);#n=new u(!1);#i;static observedAttributes=ns;constructor(){super(),is().then(e=>this.#s.set(e)).catch(e=>console.error("Failed to detect publish support:",e))}connectedCallback(){this.#i=new g,this.#i.run(this.#a.bind(this))}disconnectedCallback(){this.#i?.close(),this.#i=void 0}attributeChangedCallback(e,t,i){if(e==="show"){const n=i??"warning";if(n==="always"||n==="warning"||n==="error"||n==="never")this.show=n;else throw new Error(`Invalid show: ${n}`)}else if(e==="details")this.details=i!==null;else{const n=e;throw new Error(`Invalid attribute: ${n}`)}}get show(){return this.#e.peek()}set show(e){this.#e.set(e)}get details(){return this.#t.peek()}set details(e){this.#t.set(e)}#o(e){return e.webtransport==="none"||!e.audio.encoding||!e.video.encoding||!e.audio.capture||!Object.values(e.audio.encoding).some(t=>t===!0||t==="full"||t==="partial")||!Object.values(e.video.encoding).some(t=>t.software||t.hardware)?"none":e.video.capture==="partial"||!Object.values(e.video.encoding).some(t=>t.hardware)?"partial":"full"}#a(e){const t=e.get(this.#s);if(!t||e.get(this.#n))return;const i=e.get(this.#e);if(i==="never")return;const n=this.#o(t);if(i==="warning"&&n==="full"||i==="error"&&n!=="none")return;const o=b("div",{style:{margin:"0 auto",maxWidth:"28rem",padding:"1rem"}});this.appendChild(o),e.cleanup(()=>this.removeChild(o)),this.#r(o,n,e),e.get(this.#t)&&this.#c(o,t,e)}#r(e,t,i){const n=b("div",{style:{display:"flex",flexDirection:"row",gap:"1rem",flexWrap:"wrap",justifyContent:"space-between",alignItems:"center"}}),o=b("div",{style:{fontWeight:"bold"}});t==="full"?o.textContent="🟢 Full Browser Support":t==="partial"?o.textContent="🟡 Partial Browser Support":t==="none"&&(o.textContent="🔴 No Browser Support");const a=b("button",{type:"button",style:{fontSize:"14px"}});i.event(a,"click",()=>{this.#t.update(l=>!l)}),i.run(l=>{a.textContent=l.get(this.#t)?"Details ➖":"Details ➕"});const r=b("button",{type:"button",style:{fontSize:"14px"}},"Close ❌");i.event(r,"click",()=>{this.#n.set(!0)}),n.appendChild(o),n.appendChild(a),n.appendChild(r),e.appendChild(n),i.cleanup(()=>e.removeChild(n))}#c(e,t,i){const n=b("div",{style:{display:"grid",gridTemplateColumns:"1fr 1fr 1fr",columnGap:"0.5rem",rowGap:"0.2rem",backgroundColor:"rgba(0, 0, 0, 0.6)",borderRadius:"0.5rem",padding:"1rem",fontSize:"0.875rem"}}),o=c=>c?"🟢 Yes":"🔴 No",a=c=>c?.hardware?"🟢 Hardware":c?.software?`🟡 Software${Ae?"*":""}`:"🔴 No",r=c=>c==="full"?"🟢 Full":c==="partial"?"🟡 Polyfill":"🔴 None",l=(c,d,h)=>{const p=b("div",{style:{gridColumnStart:"1",fontWeight:"bold",textAlign:"right"}},c),v=b("div",{style:{gridColumnStart:"2",textAlign:"center"}},d),f=b("div",{style:{gridColumnStart:"3"}},h);n.appendChild(p),n.appendChild(v),n.appendChild(f)};if(l("WebTransport","",r(t.webtransport)),l("Capture","Audio",o(t.audio.capture)),l("","Video",r(t.video.capture)),l("Encoding","Opus",r(t.audio.encoding.opus)),l("","AAC",o(t.audio.encoding.aac)),l("","AV1",a(t.video.encoding?.av1)),l("","H.265",a(t.video.encoding?.h265)),l("","H.264",a(t.video.encoding?.h264)),l("","VP9",a(t.video.encoding?.vp9)),l("","VP8",a(t.video.encoding?.vp8)),Ae){const c=b("div",{style:{gridColumnStart:"1",gridColumnEnd:"4",textAlign:"center",fontSize:"0.875rem",fontStyle:"italic"}},"Hardware acceleration is ",b("a",{href:"https://github.com/w3c/webcodecs/issues/896"},"undetectable")," on Firefox.");n.appendChild(c)}e.appendChild(n),i.cleanup(()=>e.removeChild(n))}}customElements.define("moq-publish-support",os);const y=document.getElementById("publisher"),ue=document.getElementById("errors"),as=document.getElementById("url-rows"),He=document.getElementById("info-name"),ne=document.getElementById("info-url"),x=document.getElementById("info-status"),A=document.getElementById("stream-input"),rs=document.querySelectorAll("input[name=source]"),Te=document.getElementById("chk-muted"),Ce=document.getElementById("chk-invisible"),M=document.getElementById("btn-publish"),oe=document.getElementById("btn-stop"),cs=document.getElementById("broadcast-preview");function R(s,e){return`<span class="badge badge-${e}">${s}</span>`}function ls(s){const e=document.createElement("div");e.className="banner banner-error",e.innerHTML=`⛔ ${s}`,ue.appendChild(e)}const j=new URLSearchParams(window.location.search),de=window.location.pathname.split("/").filter(Boolean)[0]??"live";function us(){return Math.random().toString(36).slice(2,9)}const De=j.get("name")??j.get("broadcast")??j.get("id")??us(),ds=j.get("url")??j.get("host")??null;A.value=De;He.textContent=`${de}/${De}/publish`;function Ke(){const s=A.value.trim();cs.textContent=s?`→ ${de}/${s}/publish`:""}A.addEventListener("input",Ke);Ke();const hs=typeof VideoEncoder<"u";if(!hs){const s=document.createElement("div");s.className="banner banner-warn",s.innerHTML="⚠️ <strong>WebCodecs is not supported in this browser.</strong> Publishing requires WebCodecs for video encoding. Chrome 94+, Edge 94+, or Safari 17.4+ required.",ue.appendChild(s)}rs.forEach(s=>{s.addEventListener("change",()=>{y.setAttribute("source",s.value)})});Te.addEventListener("change",()=>{Te.checked?y.setAttribute("muted",""):y.removeAttribute("muted")});Ce.addEventListener("change",()=>{Ce.checked?y.setAttribute("invisible",""):y.removeAttribute("invisible")});function ps(s){if(s&&s.includes("://"))try{return new URL(s).hostname}catch{}return s?s.split(":")[0]:window.location.hostname||"localhost"}const Ee=ps(ds),ae=window.location.protocol,z=[`${ae}//${Ee}/moq`,`${ae}//${Ee}:4443/moq`];if(ae==="http:"){const s=window.location.href.replace(/^http:/,"https:"),e=document.createElement("div");e.className="banner banner-warn",e.innerHTML=`⚠️ <strong>Page loaded over HTTP.</strong> WebTransport requires HTTPS — publishing may fail. <a href="${s}" style="color:inherit;font-weight:bold;">Switch to HTTPS →</a>`,ue.appendChild(e)}z.forEach((s,e)=>{const t=document.createElement("tr");t.id=`url-row-${e}`,t.innerHTML=`
-    <td>${e+1}</td>
-    <td>${s}</td>
-    <td id="url-status-${e}"><span class="status-dot dot-pending"></span>pending</td>
-  `,as.appendChild(t)});function N(s,e,t){const i=document.getElementById(`url-status-${s}`),n={pending:"dot-pending",trying:"dot-trying",ok:"dot-ok",fail:"dot-fail"}[e];i.innerHTML=`<span class="status-dot ${n}"></span>${t}`;const o=document.getElementById(`url-row-${s}`);o.className=e==="ok"?"url-row active":"url-row"}async function fs(s,e=4e3){const t=new AbortController,i=setTimeout(()=>t.abort(),e);try{return await fetch(s,{method:"HEAD",mode:"no-cors",signal:t.signal}),!0}catch(n){return n.name==="AbortError"?!1:!(n instanceof TypeError)}finally{clearTimeout(i)}}async function ms(){for(let s=0;s<z.length;s++){const e=z[s];if(N(s,"trying","trying…"),x.innerHTML=R(`probing ${s+1}/${z.length}`,"yellow"),await fs(e)){N(s,"ok","✓ reachable");for(let i=s+1;i<z.length;i++)N(i,"pending","skipped");return e}else N(s,"fail","✗ unreachable")}return null}y.setAttribute("source","camera");x.innerHTML=R("preview","blue");M.addEventListener("click",async()=>{const s=A.value.trim();if(!s){A.focus();return}M.disabled=!0,x.innerHTML=R("probing relay…","yellow");const e=await ms();if(!e){ls("Could not reach any relay URL. Check that the relay server is running and the host is correct."),ne.textContent="none",x.innerHTML=R("unreachable","red"),M.disabled=!1;return}const t=`${de}/${s}/publish`;ne.textContent=e,He.textContent=t,y.broadcast.video.hd.config.set({codec:"avc1"}),y.setAttribute("url",e),y.setAttribute("name",t),x.innerHTML=R("publishing","green"),M.style.display="none",oe.style.display="inline-block",A.disabled=!0});oe.addEventListener("click",()=>{y.removeAttribute("url"),ne.textContent="—",x.innerHTML=R("preview","blue"),oe.style.display="none",M.style.display="inline-block",M.disabled=!1,A.disabled=!1,z.forEach((s,e)=>N(e,"pending","pending"))});
