+++ MoQPlugin/src/main/resources/moq-ams-player-build/play.html
@@ -249,7 +249,7 @@
     /* Compat section */
     moq-watch-support { display: block; margin-bottom: 14px; }
   </style>
-  <script type="module" crossorigin src="./assets/play-C3Y9epmR.js"></script>
+  <script type="module" crossorigin src="./assets/play-DA0pAR74.js"></script>
   <link rel="modulepreload" crossorigin href="./assets/time-Cy1-YUvq.js">
 </head>
 <body>
