+++ MoQPlugin/src/main/resources/moq-ams-player-build/publish.html
@@ -175,7 +175,7 @@
     /* Compat section */
     moq-publish-support { display: block; margin-bottom: 14px; }
   </style>
-  <script type="module" crossorigin src="./assets/publish-D1fsc_dq.js"></script>
+  <script type="module" crossorigin src="./assets/publish-uZUVINoe.js"></script>
   <link rel="modulepreload" crossorigin href="./assets/time-Cy1-YUvq.js">
 </head>
 <body>
